% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-FnAGBayesianMDL.R
\name{actualiza_carpeta}
\alias{actualiza_carpeta}
\title{Actualizador de archivos de una carpeta a otra}
\usage{
actualiza_carpeta(comienzos_de_archivo, source_raiz, carpeta_destino)
}
\arguments{
\item{comienzos_de_archivo}{este es el vector de los comienzos de los nombres
de los archivos; eg si comienzos_de_archivo = c("MDL V","Fn AG Bayesian
MDL.R"), entonces se buscarán los archivos que comienzen con "MDL V" y se
copiará el archivo con la terminación mas grande, ie si existen los
archivos "Fn AG Bayesian MDL.R", "MDL V01", "MDL V02", "MDL V03" se copiará
el archivo "Fn AG Bayesian MDL.R" (por ser único) y el archivo "MDL V03"
(por ser el que tiene el nombre mas grande alfanumericamente)}

\item{source_raiz}{dirección de donde se toman los archivos a copiar}

\item{carpeta_destino}{dirección a donde se quieren copiar los archivos}
}
\description{
Actualizador de archivos de una carpeta a otra
}
