% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-FnAGBayesianMDL.R
\name{AG_BMDL_r_paso}
\alias{AG_BMDL_r_paso}
\title{Algoritmo genético con  Bayesian-MDL}
\usage{
AG_BMDL_r_paso(param)
}
\arguments{
\item{param}{}
}
\description{
Algoritmo genético con  Bayesian-MDL
}
