% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpt_lm.R
\name{as.ts.cpt_lm}
\alias{as.ts.cpt_lm}
\alias{changepoints.cpt_lm}
\title{Methods for \code{cpt_lm}}
\usage{
\method{as.ts}{cpt_lm}(x, ...)

\method{changepoints}{cpt_lm}(x, ...)
}
\arguments{
\item{x}{A \code{cpt_lm} object}

\item{...}{arguments passed to methods}
}
\description{
Methods for \code{cpt_lm}
}
\examples{
cpts <- segment(DataCPSim)
as.ts(cpts)
changepoints(cpts)
}
