% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{as.ts.lm}
\alias{as.ts.lm}
\alias{changepoints.lm}
\title{Compatibility layer for \code{lm}}
\usage{
\method{as.ts}{lm}(x, ...)

\method{changepoints}{lm}(x, ...)
}
\arguments{
\item{x}{An \code{lm} object}

\item{...}{arguments passed to methods}
}
\description{
Compatibility layer for \code{lm}
}
\examples{
cpts <- segment(DataCPSim)
as.ts(cpts)
changepoints(cpts)
}
