% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mod_cpt.R, R/class-seg_basket.R,
%   R/class-seg_cpt.R, R/class-tidycpt.R, R/pkg-GA.R, R/pkg-changepoint.R,
%   R/pkg-wbs.R
\name{as.ts.mod_cpt}
\alias{as.ts.mod_cpt}
\alias{as.ts.seg_basket}
\alias{as.ts.seg_cpt}
\alias{as.ts.tidycpt}
\alias{as.ts.ga}
\alias{as.ts.cpt}
\alias{as.ts.wbs}
\title{Retrieve a time series from a more complex object}
\usage{
\method{as.ts}{mod_cpt}(x, ...)

\method{as.ts}{seg_basket}(x, ...)

\method{as.ts}{seg_cpt}(x, ...)

\method{as.ts}{tidycpt}(x, ...)

\method{as.ts}{ga}(x, ...)

\method{as.ts}{cpt}(x, ...)

\method{as.ts}{wbs}(x, ...)
}
\arguments{
\item{x}{An object coercible into a time series. See \code{\link[stats:ts]{stats::as.ts()}}.}

\item{...}{arguments passed to methods}
}
\value{
A \code{\link[stats:ts]{stats::ts()}} object.
}
\description{
Retrieve a time series from a more complex object
}
\examples{
cpts <- segment(DataCPSim, method = "ga", maxiter = 5)
as.ts(cpts)

cpts <- segment(DataCPSim, method = "wbs")
as.ts(cpts)

}
\seealso{
\code{\link[stats:ts]{stats::ts()}}
}
