% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoint2broom.R, R/cpt_gbmdl2broom.R
\name{augment.cpt}
\alias{augment.cpt}
\alias{tidy.cpt}
\alias{glance.cpt}
\alias{tidy.cpt_gbmdl}
\alias{glance.cpt_gbmdl}
\title{Broom compatibility layer for changepoint}
\usage{
\method{augment}{cpt}(x, ...)

\method{tidy}{cpt}(x, ...)

\method{glance}{cpt}(x, ...)

\method{tidy}{cpt_gbmdl}(x, ...)

\method{glance}{cpt_gbmdl}(x, ...)
}
\description{
Broom compatibility layer for changepoint
}
\examples{
if (require(changepoint)) {
  cpts <- cpt.meanvar(DataCPSim)
  y <- augment(cpts)
  class(y)
  y
  tidy(cpts)
  glance(cpts)
}
}
