% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom_cpt_gbmdl.R
\name{augment.cpt_gbmdl}
\alias{augment.cpt_gbmdl}
\alias{tidy.cpt_gbmdl}
\alias{glance.cpt_gbmdl}
\alias{as.ts.cpt_gbmdl}
\alias{changepoints.cpt_gbmdl}
\title{Broom compatibility layer for changepoint}
\usage{
\method{augment}{cpt_gbmdl}(x, ...)

\method{tidy}{cpt_gbmdl}(x, ...)

\method{glance}{cpt_gbmdl}(x, ...)

\method{as.ts}{cpt_gbmdl}(x, ...)

\method{changepoints}{cpt_gbmdl}(x, ...)
}
\description{
Broom compatibility layer for changepoint
}
\examples{
cpts <- lista_AG
class(cpts) <- c("cpt_gbmdl", class(cpts))
y <- augment(cpts)
class(y)
y
tidy(cpts)
glance(cpts)
cpts <- lista_AG
class(cpts) <- c("cpt_bmdl", class(cpts))
as.ts(cpts)
}
