% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-FnAGBayesianMDL.R
\name{Bayesaian_MDL_1_cp}
\alias{Bayesaian_MDL_1_cp}
\title{Bayesian MDL para un vector de puntos de cambio}
\usage{
Bayesaian_MDL_1_cp(
  cp,
  x,
  rf_type,
  initial_val_optim,
  mat_low_upp,
  vec_dist_a_priori,
  mat_phi,
  ajuste_bloque
)
}
\arguments{
\item{cp}{}

\item{x}{}

\item{rf_type}{}

\item{initial_val_optim}{}

\item{mat_low_upp}{}

\item{vec_dist_a_priori}{}

\item{mat_phi}{}

\item{ajuste_bloque}{description}
}
\description{
Bayesian MDL para un vector de puntos de cambio
}
