% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-FnAGBayesianMDL.R
\name{Bloq_LogPrio_NHPP}
\alias{Bloq_LogPrio_NHPP}
\title{Bloque de log-a priori NHPP}
\usage{
Bloq_LogPrio_NHPP(vec_dist_a_priori, theta, mat_phi)
}
\arguments{
\item{vec_dist_a_priori}{}

\item{theta}{}

\item{mat_phi}{}
}
\description{
Bloque de log-a priori NHPP
}
