% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhpp.R
\name{Bloq_LogVero_NHPP}
\alias{Bloq_LogVero_NHPP}
\title{Bloque de log-verosimilitud NHPP}
\usage{
Bloq_LogVero_NHPP(vec_d_i, tau1, tau2, theta, nhpp_dist = "W")
}
\arguments{
\item{vec_d_i}{vector de días en los que hubo rebases entre el los puntos de
cambio tau1 y tau2}

\item{tau1}{valor del primer punto de cambio}

\item{tau2}{valor del segundo punto de cambio}

\item{theta}{vector de parámetros de verosimilitud del NHPP}
}
\description{
Bloque de log-verosimilitud NHPP
}
