% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bogota_pm}
\alias{bogota_pm}
\title{Paritculate matter in Bogotá, Colombia}
\format{
An object of class \code{tbl_ts} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1096 rows and 2 columns.
}
\usage{
bogota_pm
}
\value{
A \code{\link[tsibble:tsibble]{tsibble::tsibble()}}
}
\description{
Particulate matter of less than 2.5 microns of diameter in Bogotá, Colombia.
Daily readings from 2018-2020 are included.
}
\examples{
str(bogota_pm)


}
\keyword{datasets}
