% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_genetic.R
\name{cpt_bmdl_genetic}
\alias{cpt_bmdl_genetic}
\title{Algoritmo genético de Bayesian MDL a un paso}
\usage{
cpt_bmdl_genetic(x, param, destdir = tempdir())
}
\arguments{
\item{x}{a vector that can be coerced into a \code{ts}}

\item{param}{description. See \link{param}.}
}
\value{
A \code{cpt_list} object
}
\description{
Algoritmo genético de Bayesian MDL a un paso
}
\examples{
\dontrun{
lista_AG <- cpt_bmdl_genetic(DataCPSim, param)
lista_AG <- cpt_bmdl_genetic(rlnorm_ts_1, param)
}

}
