% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoint.R
\name{cpt_candidates}
\alias{cpt_candidates}
\alias{cpt_best}
\alias{cpt_meanvar}
\title{Compatibility with changepoint package}
\usage{
cpt_candidates(lista_AG)

cpt_best(x)

cpt_meanvar(lista_AG)
}
\arguments{
\item{lista_AG}{list of candidate changepoints returned by \code{\link[=AG_BMDL_r_paso]{AG_BMDL_r_paso()}}}
}
\value{
\itemize{
\item For \code{cpt_candidates()}, a \code{\link[tibble:tibble]{tibble::tibble()}} where each row represents a
possible set of changepoints and a BMDL score.
\item For \code{cpt_best()}, a single vector of changepoints with the lowest BDML
score.
\item For \code{cpt_meanvar()}, a \code{ch}
}
}
\description{
Compatibility with changepoint package
}
\examples{
cpts_list <- cpt_candidates(lista_AG)
cpts <- cpt_meanvar(lista_AG)
class(cpts)

}
