% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpt_gbmdl.R
\name{cpt_gbmdl}
\alias{cpt_gbmdl}
\alias{validate_cpt_gbmdl}
\alias{new_cpt_gbmdl}
\title{Class for candidate changepoints using Genetic BMDL heuristic}
\usage{
cpt_gbmdl(x, ...)

validate_cpt_gbmdl(x)

new_cpt_gbmdl(
  x = numeric(),
  nhpp_dist = c("W", "EW", "GGO", "MO", "GO")[1],
  vec_dist_a_priori = c("Gamma", "Gamma"),
  mat_phi = matrix(c(1, 3, 2, 1.2), ncol = 2),
  num_generations = 50,
  generation_size = 50,
  max_num_cp = 20
)
}
\arguments{
\item{x}{a numeric vector}

\item{nhpp_dist}{toma valores en c("W","EW","GGO","MO","GO") y es el nombre de
la función de tasa del NHPP}

\item{vec_dist_a_priori}{vector de los nobmres de las distribuciones a priori
que se utilizan; eg c("Gamma","Gamma") y c("Gamma","Gamma","Gamma")}

\item{mat_phi}{matriz cuyos renglones tiene los parámetros de las
distribuciones a priori; cada renglón tiene todos los parametros de una
distribución}

\item{num_generations}{número de generaciones}

\item{generation_size}{tamaño de las generaciones}
}
\description{
Class for candidate changepoints using Genetic BMDL heuristic
}
\examples{
cpts <- cpt_gbmdl(DataCPSim)
str(cpts)
}
