% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-cptmod.R
\name{cptmod-generics}
\alias{cptmod-generics}
\alias{as.ts.cptmod}
\alias{nobs.cptmod}
\alias{logLik.cptmod}
\alias{fitted.cptmod}
\alias{residuals.cptmod}
\alias{model_variance}
\alias{glance.cptmod}
\title{Methods for cptmod objects}
\usage{
\method{as.ts}{cptmod}(x, ...)

\method{nobs}{cptmod}(object, ...)

\method{logLik}{cptmod}(object, ...)

\method{fitted}{cptmod}(object, ...)

\method{residuals}{cptmod}(object, ...)

model_variance(object, ...)

\method{glance}{cptmod}(x, ...)
}
\arguments{
\item{x}{A \code{cptmod} object, typically the output from one of the \verb{fit_*()}
functions}

\item{...}{some methods for this generic function require additional
   arguments.}

\item{object}{any object from which a log-likelihood value, or a
   contribution to a log-likelihood value, can be extracted.}
}
\description{
Methods for cptmod objects
}
\examples{
cpts <- fit_meanshift(DataCPSim, tau = 365)
as.ts(cpts)
nobs(cpts)
logLik(cpts)
fitted(cpts)
residuals(cpts)
changepoints(cpts)
glance(cpts)
}
