% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-FnAGBayesianMDL.R
\name{D_Bloq_LogPost_NHPP}
\alias{D_Bloq_LogPost_NHPP}
\title{Derivada bloque de log posterior NHPP}
\usage{
D_Bloq_LogPost_NHPP(
  vec_d_i,
  tau1,
  tau2,
  rf_type,
  theta,
  vec_dist_a_priori,
  mat_phi
)
}
\arguments{
\item{vec_d_i}{}

\item{tau1}{}

\item{tau2}{}

\item{rf_type}{}

\item{theta}{}

\item{vec_dist_a_priori}{}

\item{mat_phi}{}
}
\description{
Derivada bloque de log posterior NHPP
}
