% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-FnAGBayesianMDL.R
\name{D_Bloq_LogPrio_NHPP}
\alias{D_Bloq_LogPrio_NHPP}
\title{Derivada de bloque de log-a priori NHPP}
\usage{
D_Bloq_LogPrio_NHPP(vec_dist_a_priori, theta, mat_phi)
}
\arguments{
\item{vec_dist_a_priori}{vector que determina cual es la distribución a
priori de los parametros; por ahora se tiene programado
vec_dist_a_priori=c("Gamma","Gamma") y
vec_dist_a_priori=c("Gamma","Gamma","Gamma")}

\item{theta}{vector de parámetros de verosimilitud}

\item{mat_phi}{matriz cuyos renglones tiene los parámetros de las
distribuciones a priori; cada renglón tiene todos los parametros de una
distribución}
}
\description{
Derivada de bloque de log-a priori NHPP
}
