% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-FnAGBayesianMDL.R
\name{D_Bloq_LogVero_NHPP}
\alias{D_Bloq_LogVero_NHPP}
\title{Derivadas de bloque de log-verosimilitud NHPP}
\usage{
D_Bloq_LogVero_NHPP(vec_d_i, tau1, tau2, rf_type, theta)
}
\arguments{
\item{vec_d_i}{vector de días en los que hubo revases entre el los puntos de
cambio tau1 y tau2}

\item{tau1}{valor del primer punto de cambio}

\item{tau2}{valor del segundo punto de cambio}

\item{rf_type}{nombre de tasa de NHPP}

\item{theta}{vector de parámetros de verosimilitud del NHPP}
}
\description{
Derivadas de bloque de log-verosimilitud NHPP
}
