% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DataCPSimRebases}
\alias{DataCPSimRebases}
\title{DataCPSimRebases}
\format{
An object of class \code{integer} of length 360.
}
\usage{
DataCPSimRebases
}
\description{
DataCPSimRebases
}
\keyword{datasets}
