% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deg_free}
\alias{deg_free}
\title{Retrieve the degrees of freedom from a \code{logLik} object}
\usage{
deg_free(x)
}
\arguments{
\item{x}{An object that implements a method for \code{\link[stats:logLik]{stats::logLik()}}.}
}
\value{
The \code{df} attribute of the \code{\link[stats:logLik]{stats::logLik()}} of the given object.
}
\description{
Retrieve the degrees of freedom from a \code{logLik} object
}
\examples{
# Retrieve the degrees of freedom model a changepoint model
DataCPSim |>
  segment() |>
  as.model() |>
  deg_free()
  
}
