% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-FnAGBayesianMDL.R
\name{Ejecuta_multiples_pruebas_AG}
\alias{Ejecuta_multiples_pruebas_AG}
\title{Title}
\usage{
Ejecuta_multiples_pruebas_AG(archivo_excel, hoja_excel)

Ejecuta_multiples_pruebas_AG(archivo_excel, hoja_excel)
}
\arguments{
\item{archivo_excel}{nombre del archivo excel; eg "Parametros pruebas V01.xlsx"}

\item{hoja_excel}{nombre de la hoja que se ejecuta; eg "V01"}
}
\description{
Title

Ejecución de multiples pruebas de AG extrayendo de un archivo excel
}
