% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_temperature}
\alias{england_temperature}
\title{Mean annual temperature in Central England}
\format{
An object of class \code{tbl_ts} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 362 rows and 2 columns.
}
\usage{
england_temperature
}
\description{
Mean annual temperature in Central England
}
\keyword{datasets}
