% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exceedances}
\alias{exceedances}
\alias{split_by_tau}
\alias{deg_free}
\title{Calculate overages}
\usage{
exceedances(x)

split_by_tau(x, tau)

deg_free(x, ...)
}
\arguments{
\item{x}{a time series object}
}
\description{
Calculate overages
}
\examples{
exceedances(DataCPSim)
split_by_tau(as.ts(lista_AG), changepoints(lista_AG))
deg_free(segment(DataCPSim))
}
