% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{exceedances.ts}
\alias{exceedances.ts}
\alias{exceedances.double}
\alias{MDL.logLik}
\alias{MBIC.logLik}
\title{Generic methods for classes from \link{stats}}
\usage{
\method{exceedances}{ts}(x, ...)

\method{exceedances}{double}(x, ...)

\method{MDL}{logLik}(object, ...)

\method{MBIC}{logLik}(object, ...)
}
\arguments{
\item{x}{A numeric vector}

\item{...}{arguments passed to methods}

\item{object}{A \code{tidycpt} object}
}
\description{
Generic methods for classes from \link{stats}
}
\examples{
MDL(fit_meanshift_ar1(CET, tau = c(42, 330)))
}
\references{
Zhang and Seigmmund (2007) for MBIC: \doi{10.1111/j.1541-0420.2006.00662.x}
}
