% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_bmdl.R
\name{extrae_mat_MAP}
\alias{extrae_mat_MAP}
\title{Extrae matriz con estimadores MAP}
\usage{
extrae_mat_MAP(cp, x)
}
\arguments{
\item{cp}{vector de tamaño max_num_cp con entradas m, tau_0=1 , ...,
tau_{m+1}, 0, ..., 0}

\item{x}{description}

\item{rf_type}{nombre de tasa de NHPP}

\item{vec_dist_a_priori}{nombres de distribuciones a priori}

\item{mat_phi}{matriz cuyos renglones tiene los parámetros de las
distribuciones a priori}
}
\value{
regresa una matriz cuya primera columna es la log-posterior evaluada
en los estimadores MAP; sus siguientes columnas tiene los parametros de
cada regimen.
}
\description{
Extrae matriz con estimadores MAP
}
\examples{
chromo <- chromosome_best(lista_AG$segmenter)
extrae_mat_MAP(chromo, as.ts(lista_AG))
}
