% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_lm.R
\name{fit_lm}
\alias{fit_lm}
\title{Regression-based model fitting}
\usage{
fit_lm(x, tau, trends = FALSE, ...)
}
\description{
Regression-based model fitting
}
\examples{
tau <- c(365, 826)
mod <- fit_lm(DataCPSim, tau)
logLik(mod)
deg_free(mod)

cpts = c(1700, 1739, 1988)
ids <- match(cpts, lubridate::year(index(CET)))
mod <- fit_lm(CET, tau = ids)

mod <- fit_lm(CET, tau = ids, trends = TRUE)
logLik(mod)
deg_free(mod)
}
