% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhpp.R
\name{fit_nhpp_region}
\alias{fit_nhpp_region}
\title{Estimadores MAP}
\usage{
fit_nhpp_region(
  t,
  tau_left,
  tau_right,
  initial_val_optim = c(0.1, 0.5),
  mat_low_upp = matrix(c(c(1e-04, 1e-08), c(10, 1e+05)), nrow = 2),
  vec_dist_a_priori = c("Gamma", "Gamma"),
  mat_phi = matrix(c(1, 3, 2, 1.2), ncol = 2),
  ...
)
}
\arguments{
\item{mat_low_upp}{matriz con lugares donde buscar; cada renglon es para un
parámetro del NHPP}

\item{vec_dist_a_priori}{nombres de distribuciones a priori}

\item{mat_phi}{matriz cuyos renglones tiene los parámetros de las
distribuciones a priori}

\item{vec_d_i}{vector de días de un régimen}

\item{tau1}{valor del primer punto de cambio}

\item{tau2}{valor del segundo punto de cambio}
}
\value{
regresa un el resultado de optim
}
\description{
Estimadores MAP
}
\examples{
fit_nhpp_region(exceedances(lista_AG), 0, 575)
fit_nhpp_region(exceedances(lista_AG), 0, 575, initial_val_optim = c(1, 10))


}
