% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-FnAGBayesianMDL.R
\name{fun_1_genera_texto_dist}
\alias{fun_1_genera_texto_dist}
\title{Genera los textos para los pdf y RData de una distribución}
\usage{
fun_1_genera_texto_dist(dist, parametros_dist)
}
\arguments{
\item{dist}{}

\item{parametros_dist}{}
}
\description{
Genera los textos para los pdf y RData de una distribución
}
