% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-FnAGBayesianMDL.R
\name{fun_n_genera_texto_dist}
\alias{fun_n_genera_texto_dist}
\title{Genera los textos para los pdf y RData de las n distribuciones}
\usage{
fun_n_genera_texto_dist(param)
}
\arguments{
\item{param}{}
}
\description{
Genera los textos para los pdf y RData de las n distribuciones
}
