% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcion_media_acumulada.R
\name{funcion_media_acumulada}
\alias{funcion_media_acumulada}
\title{funcion media}
\usage{
funcion_media_acumulada(i = 1, d, alpha, sigma, tau)
}
\description{
funcion media
}
\examples{
funcion_media_acumulada(
  i = 1, d = sample(1:1096), alpha = runif(6), 
  sigma = runif(6), tau = sample(1:1096, 5))


}
