% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/pkg-GA.R
\name{as.mod_default.default}
\alias{as.mod_default.default}
\alias{ga-generics}
\alias{glance.ga}
\alias{params.ga}
\alias{as.ts.ga}
\alias{nobs.ga}
\title{Compatibility layer for GA}
\usage{
\method{as.mod_default}{default}(object, ...)

\method{glance}{ga}(x, ...)

\method{params}{ga}(x, ...)

\method{as.ts}{ga}(x, ...)

\method{nobs}{ga}(object, ...)
}
\arguments{
\item{object}{A \code{ga} object.}

\item{...}{arguments passed to methods}

\item{x}{A \code{GA} object returned by \code{\link[GA:ga]{GA::ga()}}}
}
\description{
Compatibility layer for GA
}
\examples{
cpt <- segment(DataCPSim, method = "ga", maxiter = 5)
as.mod_default(cpt$segmenter)
cpts <- segment(DataCPSim, method = "ga", maxiter = 10)
class(cpts)
y <- augment(cpts)
class(y)
y
tidy(cpts)
glance(cpts)
cpts <- segment(DataCPSim, method = "ga", maxiter = 5)
as.ts(cpts)

cpts <- segment(DataCPSim, method = "ga", maxiter = 5)
nobs(cpts)

}
