% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-FnAGBayesianMDL.R
\name{gen_texto_m}
\alias{gen_texto_m}
\title{gen_texto_m
Genera el texto correspondiente a la variable m para el archivo .txt para JAGS}
\usage{
gen_texto_m(n_puntos_cambio, mas_derecha = "", destdir = tempdir())
}
\arguments{
\item{n_puntos_cambio}{numero de puntos de cambio}

\item{mas_derecha}{es una variable que fue utilizada para que
pudiera ejecutarse en R y en JAGS; toma los valores "" (vacío) y "+".}
}
\value{
Regresa el texto correspondiente a la variable m
}
\description{
gen_texto_m
Genera el texto correspondiente a la variable m para el archivo .txt para JAGS
}
\examples{
\dontrun{
mejor_cp <- historia_mejores[which(
  Bayesaian_MDL_k_cp(
    historia_mejores, x, param$rf_type, param$initial_val_optim,
    param$mat_low_upp, param$vec_dist_a_priori, param$mat_phi
  ) == min(Bayesaian_MDL_k_cp(
    historia_mejores, x, param$rf_type, param$initial_val_optim,
    param$mat_low_upp, param$vec_dist_a_priori, param$mat_phi
  ))
), ]
}
}
