% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generative.R
\name{genera_insumos_bloque_sin_theta}
\alias{genera_insumos_bloque_sin_theta}
\alias{split_by_tau}
\title{Agrupamos los insumos por bloques}
\usage{
genera_insumos_bloque_sin_theta(cp, x)

split_by_tau(x, tau)
}
\arguments{
\item{cp}{vector de tamaño max_num_cp con entradas m, tau_0=1 , ...,
tau_{m+1}, 0, ..., 0}

\item{x}{datos de rebases}
}
\value{
regresa una lista con una matriz de 5 columnas la cual tiene
los el tau_{i-1}, tau_i, theta_1, theta_2 y theta_3, y también
una lista con los días de cada régimen
}
\description{
Agrupamos los insumos por bloques
}
\examples{
genera_insumos_bloque_sin_theta(chromosome_best(lista_AG), lista_AG$data)

split_by_tau(as.ts(lista_AG), changepoints(lista_AG))
}
