% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-FnAGBayesianMDL.R
\name{genera_texto_eval_param}
\alias{genera_texto_eval_param}
\title{Generador de texto para evaluar una lista param}
\usage{
genera_texto_eval_param(renglon_parametros)

genera_texto_eval_param(renglon_parametros)
}
\arguments{
\item{renglon_parametros}{renglón extraido del excel}
}
\description{
Generador de texto para evaluar una lista param

Generador de texto para evaluar una lista param
}
\examples{
library(readxl)
library(xlsx)
archivo_excel <- "Parametros pruebas V01.xlsx"
hoja_excel <- "V01"
matriz_parametros <- read.xlsx(xlsxFile = archivo_excel, sheet = hoja_excel, skipEmptyRows = FALSE,colNames = T)
renglon_parametros <- matriz_parametros[2,]
texto_list_param <- genera_texto_eval_param(renglon_parametros)
eval(parse(text=texto_list_param))
archivo_excel <- "Parametros pruebas V01.xlsx"
hoja_excel <- "V01"
matriz_parametros <- read.xlsx(xlsxFile = archivo_excel, sheet = hoja_excel, skipEmptyRows = FALSE,colNames = T)
renglon_parametros <- matriz_parametros[2,]
texto_list_param <- genera_texto_eval_param(renglon_parametros)
eval(parse(text=texto_list_param))
}
