% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generative.R
\name{generate_mean_cumulative}
\alias{generate_mean_cumulative}
\title{generate_mean_cumulative
Genera el texto correspondiente a la variable m para el archivo .txt para JAGS}
\usage{
generate_mean_cumulative(
  n_puntos_cambio,
  mas_derecha = "",
  destdir = tempdir()
)
}
\arguments{
\item{n_puntos_cambio}{numero de puntos de cambio}

\item{mas_derecha}{es una variable que fue utilizada para que
pudiera ejecutarse en R y en JAGS; toma los valores "" (vacío) y "+".}
}
\value{
Regresa el texto correspondiente a la variable m
}
\description{
generate_mean_cumulative
Genera el texto correspondiente a la variable m para el archivo .txt para JAGS
}
\examples{
\dontrun{
mejor_cp <- historia_mejores[which(
  Bayesaian_MDL_k_cp(
    historia_mejores, x, param$rf_type, param$initial_val_optim,
    param$mat_low_upp, param$vec_dist_a_priori, param$mat_phi
  ) == min(Bayesaian_MDL_k_cp(
    historia_mejores, x, param$rf_type, param$initial_val_optim,
    param$mat_low_upp, param$vec_dist_a_priori, param$mat_phi
  ))
), ]
}
d <- exceedances(lista_AG$data)
tau <- cpt_best(lista_AG)
theta <- cpt_best_params(lista_AG)
alpha <- theta$alpha
sigma <- theta$beta
f <- generate_mean_cumulative(length(tau))
f(1:length(d), d, tau, alpha, sigma)
}
