% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoint.R, R/wbs.R
\name{glance.cpt}
\alias{glance.cpt}
\alias{params.cpt}
\alias{as.ts.cpt}
\alias{nobs.cpt}
\alias{logLik.cpt}
\alias{changepoints.cpt}
\alias{params.wbs}
\alias{as.ts.wbs}
\alias{nobs.wbs}
\alias{logLik.wbs}
\alias{changepoints.wbs}
\title{Compatibility layer for changepoint}
\usage{
\method{glance}{cpt}(x, ...)

\method{params}{cpt}(x, ...)

\method{as.ts}{cpt}(x, ...)

\method{nobs}{cpt}(object, ...)

\method{logLik}{cpt}(object, ...)

\method{changepoints}{cpt}(x, ...)

\method{params}{wbs}(x, ...)

\method{as.ts}{wbs}(x, ...)

\method{nobs}{wbs}(object, ...)

\method{logLik}{wbs}(object, ...)

\method{changepoints}{wbs}(x, ...)
}
\arguments{
\item{x}{A \code{cpt} object returned by \code{\link[changepoint:cpt.meanvar]{changepoint::cpt.meanvar()}}}

\item{...}{arguments passed to methods}

\item{object}{A \code{cpt} object.}
}
\description{
Compatibility layer for changepoint
}
\examples{
cpts <- segment(DataCPSim, method = "pelt")
class(cpts)
y <- augment(cpts)
class(y)
y
tidy(cpts)
glance(cpts)
cpts <- segment(DataCPSim, method = "pelt")
as.ts(cpts)

cpts <- segment(DataCPSim, method = "pelt")
nobs(cpts)

cpts <- segment(DataCPSim, method = "pelt", penalty = "BIC")
logLik(cpts)
cpts <- segment(DataCPSim, method = "pelt", penalty = "AIC")
logLik(cpts)

cpts <- segment(DataCPSim, method = "pelt")
changepoints(cpts)

cpts <- segment(DataCPSim, method = "wbs")
as.ts(cpts)

cpts <- segment(DataCPSim, method = "wbs")
nobs(cpts)

cpts <- segment(DataCPSim, method = "wbs")
logLik(cpts)

cpts <- segment(DataCPSim, method = "wbs")
changepoints(cpts$segmenter)

}
