% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpt_gbmdl.R
\name{glance.cpt_gbmdl}
\alias{glance.cpt_gbmdl}
\alias{as.ts.cpt_gbmdl}
\alias{length.cpt_gbmdl}
\alias{changepoints.cpt_gbmdl}
\alias{diagnose.cpt_gbmdl}
\title{Broom compatibility layer for changepoint}
\usage{
\method{glance}{cpt_gbmdl}(x, ...)

\method{as.ts}{cpt_gbmdl}(x, ...)

\method{length}{cpt_gbmdl}(x, ...)

\method{changepoints}{cpt_gbmdl}(x, ...)

\method{diagnose}{cpt_gbmdl}(x, ...)
}
\description{
Broom compatibility layer for changepoint
}
\examples{
cpts <- lista_AG
y <- augment(lista_AG)
class(y)
y
tidy(cpts)
glance(cpts)
as.ts(lista_AG)
length(lista_AG)

diagnose(lista_AG)

}
