% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA.R
\name{glance.ga}
\alias{glance.ga}
\alias{params.ga}
\alias{as.ts.ga}
\alias{nobs.ga}
\alias{logLik.ga}
\alias{logLik.ts}
\alias{changepoints.ga}
\title{Compatibility layer for GA}
\usage{
\method{glance}{ga}(x, ...)

\method{params}{ga}(x, ...)

\method{as.ts}{ga}(x, ...)

\method{nobs}{ga}(object, ...)

\method{logLik}{ga}(object, ...)

\method{logLik}{ts}(object, ...)

\method{changepoints}{ga}(x, ...)
}
\arguments{
\item{x}{A \code{GA} object returned by \code{\link[GA:ga]{GA::ga()}}}

\item{...}{arguments passed to methods}

\item{object}{A \code{ga} object.}
}
\description{
Compatibility layer for GA
}
\details{
loc.ind is the binary input of length N
Note \code{X[1]} cannot be a changepoint,
so \code{loc.ind[1]} is set to be zero
loc.ind = 0/1. 1 corresponds a CPT location, 0 is not.
}
\examples{
cpts <- segment(DataCPSim, method = "ga", maxiter = 10)
class(cpts)
y <- augment(cpts)
class(y)
y
tidy(cpts)
glance(cpts)
cpts <- segment(DataCPSim, method = "ga", maxiter = 5)
as.ts(cpts)

cpts <- segment(DataCPSim, method = "ga", maxiter = 5)
nobs(cpts)

cpts <- segment(DataCPSim, method = "ga", maxiter = 5)
logLik(cpts)

x <- england_temperature$annual_mean_temp
logLik(as.ts(x), loc.ind = round(runif(length(x))))
cpts <- segment(DataCPSim, method = "ga", maxiter = 5)
changepoints(cpts$segmenter)

}
\author{
Xueheng Shi
}
