% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-FnAGBayesianMDL.R
\name{graf_puntos_cambio_repetidos}
\alias{graf_puntos_cambio_repetidos}
\title{Grafica los n puntos de cambio más repetidos sobre el plot.stepfun de los datos}
\usage{
graf_puntos_cambio_repetidos(lista_AG)
}
\arguments{
\item{lista_AG}{}
}
\value{
guarda un pdf con la gráfica
}
\description{
Grafica los n puntos de cambio más repetidos sobre el plot.stepfun de los datos
}
