% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{grafica_ajuste_NHPP}
\alias{grafica_ajuste_NHPP}
\title{Graficando un ajuste de NHPP}
\usage{
grafica_ajuste_NHPP(
  d_i,
  tau1,
  tau2,
  initial_val_optim,
  mat_low_upp,
  rf_type,
  vec_dist_a_priori,
  mat_phi
)
}
\arguments{
\item{d_i}{vector de días de revases}

\item{tau1}{description}

\item{tau2}{description}

\item{mat_low_upp}{description}

\item{mat_phi}{description}
}
\description{
Esta función solo grafica un ajuste de NHPP
}
\examples{
\dontrun{
grafica_ajuste_NHPP(exceedances(lista_AG$data), 
tau1 = cpt_best(lista_AG), tau2 = cpt_best(lista_AG),
lista_AG$param$initial_val_optim, 
lista_AG$param$mat_low_upp,
lista_AG$param$rf_type,
lista_AG$vec_dist_a_priori,
lista_AG$param$mat_phi
)
}

}
