% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-FnAGBayesianMDL.R
\name{grafica_datos_e_intervalos}
\alias{grafica_datos_e_intervalos}
\title{Graficas de los intervalos y los datos}
\usage{
grafica_datos_e_intervalos(lista_AG, param)
}
\arguments{
\item{lista_AG}{}

\item{param}{}
}
\description{
Graficas de los intervalos y los datos
}
