% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{grafica_escalonada}
\alias{grafica_escalonada}
\title{Grafica escalonada para un proceso poisson}
\usage{
grafica_escalonada(res, col_segmets, vec_xlim = NULL, vec_ylim = NULL)
}
\arguments{
\item{res}{vector que contiene los lugares donde se quiere un escal칩n de tama침o 1}

\item{col_segmets}{color de las lineas}

\item{vec_xlim}{limites en el eje x de la gr치fica}

\item{vec_ylim}{limites en el eje y de la gr치fica}
}
\description{
Grafica escalonada para un proceso poisson
}
\examples{
res <- c(10, 20, 33, 36, 43, 51, 56, 69, 82, 89)
grafica_escalonada(res, col_segmets = "blue")

}
