% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-FnAGBayesianMDL.R
\name{graficas_BMDL}
\alias{graficas_BMDL}
\title{Graficas del AG con BMDL}
\usage{
graficas_BMDL(lista_AG, param)
}
\arguments{
\item{lista_AG}{}

\item{param}{}
}
\value{
una gráfica con :
1 Datos reales
2 Evolución del algoritmo genético
3 Puntos de cambio que más se repitieron
4 Number of change points in the best chromosomes
}
\description{
Graficas del AG con BMDL
}
