% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_musa-okumoto.R
\name{imusaokomoto}
\alias{imusaokomoto}
\alias{mmusaokomoto}
\alias{log_likelihood_region_musaokomoto}
\alias{log_prior_region_musaokomoto}
\alias{D_log_prior_region_musaokomoto}
\alias{D_log_likelihood_region_musaokomoto}
\title{Evaluanción de rate function (la derivada de la mean)}
\usage{
imusaokomoto(t, theta)

mmusaokomoto(t, theta)

log_likelihood_region_musaokomoto(t, tau_left, tau_right, theta)

log_prior_region_musaokomoto(theta, hyperparameters)

D_log_prior_region_musaokomoto(theta, hyperparameters)

D_log_likelihood_region_musaokomoto(
  vec_d_i,
  tau1,
  tau2,
  theta,
  nhpp_dist = "W"
)
}
\arguments{
\item{t}{valor a evaluar (real)}

\item{theta}{vector de parámetros de verosimilitud del NHPP}

\item{vec_d_i}{vector de días en los que hubo revases entre el los puntos de
cambio tau1 y tau2}

\item{tau1}{valor del primer punto de cambio}

\item{tau2}{valor del segundo punto de cambio}

\item{nhpp_dist}{nombre de tasa de NHPP}

\item{vec_dist_a_priori}{vector que determina cual es la distribución a
priori de los parametros; por ahora se tiene programado
vec_dist_a_priori=c("Gamma","Gamma") y
vec_dist_a_priori=c("Gamma","Gamma","Gamma")}

\item{mat_phi}{matriz cuyos renglones tiene los parámetros de las
distribuciones a priori; cada renglón tiene todos los parametros de una
distribución}
}
\value{
regresa la evaluación

regresa la evaluación
}
\description{
Evaluanción de rate function (la derivada de la mean)
}
\examples{
t <- c(1.4, 2.8)
theta <- c(1.2, 2.1, 3.2)

imusaokomoto(t, nhpp_dist, theta)

t <- c(1.4, 2.8)
theta <- c(1.2, 2.1, 3.2)
mmusaokomoto(t, nhpp_dist, theta)

log_likelihood_region(DataCPSim, 0, 575, theta = c(0.5, 0.5))

hyper <- data.frame(
  hyperprior_shape = c(1, 3),
  hyperprior_scale = c(2, 1.2)
)
log_prior_region(theta = c(0.5, 0.5), hyper)
}
