% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{italy_grads}
\alias{italy_grads}
\title{Italian graduates}
\format{
An object of class \code{tbl_ts} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 88 rows and 17 columns.
}
\source{
\url{https://seriestoriche.istat.it/}
}
\usage{
italy_grads
}
\description{
Italian graduates
}
\keyword{datasets}
