% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_bmdl.R
\name{junta_1_puntos_cambio}
\alias{junta_1_puntos_cambio}
\title{Hace un hijo de dos padres}
\usage{
junta_1_puntos_cambio(padres, mat_cp)
}
\arguments{
\item{padres}{vector de longitud dos con índice de papa e índice de mama}

\item{mat_cp}{matriz con tiempos de cambio de dimension k por max_num_cp}
}
\description{
Hace un hijo de dos padres
}
