% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-FnAGBayesianMDL.R
\name{junta_k_puntos_cambio}
\alias{junta_k_puntos_cambio}
\title{Hace k hijos de k parejas de padres}
\usage{
junta_k_puntos_cambio(mat_padres, mat_cp)
}
\arguments{
\item{mat_cp}{matriz con cromosomas cambio de tamaño max_num_cp con entradas
m,tau_0=1,...,tau_{m+1}=N,0,...,0}

\item{parejas_padres}{matriz de kx2 la cual contiene en sus renglones las
parejas de padres}
}
\value{
regresa una matriz de las mismas dimensiones que mat_cp, pero con los
nuevos cromosomas
}
\description{
Hace k hijos de k parejas de padres
}
