% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{label_priors}
\alias{label_priors}
\title{Genera los textos para los pdf y RData de las n distribuciones}
\usage{
label_priors(x)
}
\description{
Genera los textos para los pdf y RData de las n distribuciones
}
