% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lista_AG}
\alias{lista_AG}
\title{Output from running genetic algorithm on DataCPSim}
\format{
An object of class \code{tidycpt} of length 1096.
}
\usage{
lista_AG
}
\description{
Output from running genetic algorithm on DataCPSim
}
\keyword{datasets}
