% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lmshift.R
\name{lmshift-generics}
\alias{lmshift-generics}
\alias{as.ts.lmshift}
\alias{logLik.lmshift}
\alias{glance.lmshift}
\title{Methods for lmshift objects}
\usage{
\method{as.ts}{lmshift}(x, ...)

\method{logLik}{lmshift}(object, ...)

\method{glance}{lmshift}(x, ...)
}
\arguments{
\item{x}{An \code{lmshift} object, typically the output from \code{\link[=fit_lmshift]{fit_lmshift()}}`}

\item{...}{some methods for this generic function require additional
   arguments.}

\item{object}{any object from which a log-likelihood value, or a
   contribution to a log-likelihood value, can be extracted.}
}
\description{
Methods for lmshift objects
}
\examples{
cpts <- fit_lmshift(DataCPSim, tau = 365)
as.ts(cpts)
}
