% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{log_likelihood_norm}
\alias{log_likelihood_norm}
\alias{logLik_cp}
\alias{best_cp}
\title{Likelihood}
\usage{
log_likelihood_norm(mod)

logLik_cp(x, i = 1)

best_cp(x)
}
\description{
Likelihood
}
\examples{
ds <- as.data.frame(rlnorm_ts_1)
ds["t"] <- 1:length(rlnorm_ts_1)
mod <- lm(x ~ 1, data = ds)
mu <- coef(mod)
sigma <- sqrt(mean(mod$residuals^2))
logLik(mod)
log_likelihood_norm(mod)
log_likelihood_norm_param(rlnorm_ts_1, mean = coef(mod), sd = sqrt(mean(mod$residuals^2)))
likelihood_dnorm(rlnorm_ts_1, mean = mu, sd = sigma)

mod_cp <- lm(x ~ t > 826, data = ds)
logLik(mod_cp)
logLik_cp(rlnorm_ts_1, 634)
logLik_cp(rlnorm_ts_1, 826)
cp <- best_cp(rlnorm_ts_1)
plot(1:length(rlnorm_ts_1), cp)
plot(1:length(rlnorm_ts_2), best_cp(rlnorm_ts_2))
plot(1:length(rlnorm_ts_3), best_cp(rlnorm_ts_3))
}
