% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_musa-okumoto.R
\name{log_likelihood_region_musaokomoto}
\alias{log_likelihood_region_musaokomoto}
\title{Log-likelihood}
\usage{
log_likelihood_region_musaokomoto(t, tau_left, tau_right, theta)
}
\description{
Log-likelihood
}
\examples{
log_likelihood_region(DataCPSim, 0, 575, theta = c(0.5, 0.5))

}
