% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_weibull.R
\name{log_likelihood_region_weibull}
\alias{log_likelihood_region_weibull}
\title{Log-likelihood}
\usage{
log_likelihood_region_weibull(t, tau_left, tau_right, theta)
}
\description{
Log-likelihood
}
\examples{
log_likelihood_region_weibull(DataCPSim, 0, 575, theta = c(0.5, 2))

}
