% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{logLik.ts}
\alias{logLik.ts}
\alias{exceedances.ts}
\alias{exceedances.double}
\alias{MDL.logLik}
\alias{MBIC.logLik}
\title{Generic methods for classes from \link{stats}}
\usage{
\method{logLik}{ts}(object, ...)

\method{exceedances}{ts}(x, ...)

\method{exceedances}{double}(x, ...)

\method{MDL}{logLik}(object, ...)

\method{MBIC}{logLik}(object, ...)
}
\arguments{
\item{object}{any object from which a log-likelihood value, or a
   contribution to a log-likelihood value, can be extracted.}

\item{...}{some methods for this generic function require additional
   arguments.}

\item{x}{A numeric vector}
}
\description{
Generic methods for classes from \link{stats}
}
\details{
loc.ind is the binary input of length N
Note \code{X[1]} cannot be a changepoint,
so \code{loc.ind[1]} is set to be zero
loc.ind = 0/1. 1 corresponds a CPT location, 0 is not.
}
\examples{
logLik(as.ts(CET), loc.ind = round(runif(length(CET))))
logLik(as.ts(CET), loc.ind = tau2binary(c(42, 81, 330), n = length(CET)))
logLik(fit_meanshift(CET, tau = c(42, 81, 330)))
MDL(logLik(as.ts(CET), loc.ind = tau2binary(123, n = length(CET))))
}
\references{
Zhang and Seigmmund (2007) for MBIC: \doi{10.1111/j.1541-0420.2006.00662.x}
}
\author{
Xueheng Shi
}
