% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-fun_cpt.R, R/tidychangepoint-package.R
\name{ls_models}
\alias{ls_models}
\alias{ls_pkgs}
\alias{ls_methods}
\alias{ls_penalties}
\alias{ls_cpt_penalties}
\alias{ls_coverage}
\title{Algorithmic coverage through tidychangepoint}
\usage{
ls_models()

ls_pkgs()

ls_methods()

ls_penalties()

ls_cpt_penalties()

ls_coverage()
}
\value{
A \link[tibble:tibble]{tibble::tibble} or \code{character}
}
\description{
Algorithmic coverage through tidychangepoint
}
\examples{
# List all model-fitting functions
ls_models()

# List packages supported by tidychangepoint
ls_pkgs()

# List methods supported by segment()
ls_methods()

# List penalty functions provided by tidychangepoint
ls_penalties()

# List penalty functions supported by changepoint
ls_cpt_penalties()

# List combinations of method, model, and penalty supported by tidychangepoint
ls_coverage()

}
\seealso{
\code{\link[=segment]{segment()}}
}
