% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-MaMBOltzmann.R, R/3-GABoltzmannRanks.R
\name{MaMBOltzmann}
\alias{MaMBOltzmann}
\alias{GABoltzmannMutation2}
\title{Boltzman}
\usage{
MaMBOltzmann(mat_cp, Fitness, Mutation, Temperature)

GABoltzmannMutation2(x, param, Mutation = 0.03, Temperature = 27)
}
\arguments{
\item{mat_cp}{description}

\item{Fitness}{description}

\item{Mutation}{description}

\item{Temperature}{description}

\item{x}{An \code{integer} vector. See \link{DataCPSimRebases}.}

\item{param}{A \code{list} or parameters. See \link{param}.}
}
\value{
A \code{list} of length 2:
\itemize{
\item A \code{numeric} vector of \code{FitnessGeneraciones}
\item A \code{matrix} with \code{k} rows and \code{N} columns
}
}
\description{
What does this function do?
}
\examples{
\dontrun{
  GABoltzmannMutation2(sims$ts1, param)
  
}
}
