% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-FnAGBayesianMDL.R
\name{mata_k_tau_volado}
\alias{mata_k_tau_volado}
\title{Elimina algunos de las tiempos de cambio de los k cromosomas}
\usage{
mata_k_tau_volado(mat_cp, prob_volado)
}
\arguments{
\item{mat_cp}{matriz cuyos renglones son vectores de cromosomas de tamaño
max_num_cp con entradas m,tau_0,...,tau_{m+1},0,...,0}

\item{prob_volado}{probabilidad de quitar un tiempo de cambio existente}
}
\value{
regresa una matriz a la cual se le quitaron a sus cromosomas algunos
puntos de cambio
}
\description{
Elimina algunos de las tiempos de cambio de los k cromosomas
}
