% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhpp.R
\name{mcdf}
\alias{mcdf}
\title{Mean cumulative exceedances function}
\usage{
mcdf(x, dist = "weibull")
}
\arguments{
\item{x}{An \code{nhpp} object}

\item{dist}{Name of the distribution}
}
\value{
a numeric vector of length equal to the \link{exceedances} of \code{x}
}
\description{
Mean cumulative exceedances function
}
\examples{
nhpp <- fit_nhpp(lista_AG$segmenter, tau = changepoints(lista_AG))
mcdf(nhpp)

}
