% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mde_rain}
\alias{mde_rain}
\alias{mde_rain_monthly}
\title{Rainfall in Medellín, Colombia}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 185705 rows and 8 columns.

An object of class \code{xts} (inherits from \code{zoo}) with 444 rows and 1 columns.
}
\usage{
mde_rain

mde_rain_monthly
}
\description{
Rainfall in Medellín, Colombia
}
\details{
Daily rainfall measurements for 13 different weather stations positioned
around Medellín, Colombia.
Variables:
\itemize{
\item \code{station_id}:
\item \code{lat}, \code{long}: latitude and longitude for the weather station
\item \code{date}, \code{year}, \code{month}, \code{day}: date variables
\item \code{rainfall}: daily rainfall (in cubic centimeters) as measured by the weather station
}

\itemize{
\item \code{mean_rainfall}: average rainfall across all weather stations
}
}
\references{
\href{https://www.openstreetmap.org/?mlat=6.244747&mlon=-75.574828&zoom=12}{OpenStreetMap}
}
\keyword{datasets}
