% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{mean_fn}
\alias{mean_fn}
\title{Evaluanción de mean function (la integral de la rate)}
\usage{
mean_fn(t, rf_type, theta)
}
\arguments{
\item{t}{valor a evaluar (real)}

\item{rf_type}{tipo de mean function (ver libro de Eliane pdf 40 eq (3.12));
posibles exponentiated-Weibull (EW),the Musa–Okumoto (MO),
the Goel–Okumoto (GO), and a generalized Goel–Okumoto (GGO)}

\item{theta}{parámetros de mean function.Para
\itemize{
\item W: \eqn{\alpha = \theta[1], \sigma = \theta[2]}
\item EW: \eqn{\alpha= \theta[1], \beta = \theta[2]} y \eqn{\sigma = \theta[3]}
\item GGO: \eqn{\alpha = \theta[1], \beta = \theta[2]} y \eqn{\gamma = \theta[3]}
\item MO: \eqn{\alpha = \theta[1], \beta = \theta[2]}
\item GO: \eqn{\alpha = \theta[1], \beta = \theta[2]}
}}
}
\value{
regresa la evaluación
}
\description{
Evaluanción de mean function (la integral de la rate)
}
\examples{
t <- c(1.4, 2.8)
theta <- c(1.2, 2.1, 3.2)
rf_type <- "EW"
mean_fn(t, rf_type, theta)
rf_type <- "GGO"
mean_fn(t, rf_type, theta)
rf_type <- "MO"
mean_fn(t, rf_type, theta)
rf_type <- "GO"
mean_fn(t, rf_type, theta)

}
