% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_meanshift.R
\name{meanshift-generics}
\alias{meanshift-generics}
\alias{as.ts.meanshift}
\alias{logLik.meanshift}
\alias{nobs.meanshift}
\alias{glance.meanshift}
\title{Methods for meanshift objects}
\usage{
\method{as.ts}{meanshift}(x, ...)

\method{logLik}{meanshift}(object, ...)

\method{nobs}{meanshift}(object, ...)

\method{glance}{meanshift}(x, ...)
}
\arguments{
\item{x}{A \code{meanshift} object, typically the output from \code{\link[=fit_meanshift]{fit_meanshift()}}`}

\item{...}{some methods for this generic function require additional
   arguments.}

\item{object}{any object from which a log-likelihood value, or a
   contribution to a log-likelihood value, can be extracted.}
}
\description{
Methods for meanshift objects
}
\examples{
cpts <- fit_meanshift(DataCPSim, tau = 365)
as.ts(cpts)
}
