% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{medellin_rainfall}
\alias{medellin_rainfall}
\title{Rainfall in Medellín, Colombia}
\format{
An object of class \code{xts} (inherits from \code{zoo}) with 444 rows and 1 columns.
}
\usage{
medellin_rainfall
}
\description{
Rainfall in Medellín, Colombia
}
\keyword{datasets}
