% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mlb_diffs}
\alias{mlb_diffs}
\title{Differences between leagues in Major League Baseball}
\format{
An object of class \code{tbl_ts} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 99 rows and 6 columns.
}
\source{
The \code{Lahman} package
}
\usage{
mlb_diffs
}
\description{
The difference in various statistics between the
American League and the National League from 1925 to 2023.
Statistics include:
\itemize{
\item \code{PA}: The total number of plate appearances
\item \code{hr_rate_diff}: The difference in home runs per plate appearance
\item \code{bavg_dff}: The difference in batting average
\item \code{obp_diff}: The difference in on-base percentage
\item \code{slg_diff}: The difference in slugging percentage
}
}
\keyword{datasets}
