% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mlb_hrs}
\alias{mlb_hrs}
\title{Differences between leagues in Major League Baseball}
\format{
An object of class \code{tbl_ts} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 98 rows and 2 columns.
}
\usage{
mlb_hrs
}
\description{
Differences between leagues in Major League Baseball
}
\keyword{datasets}
