% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mod_cpt.R
\name{mod_cpt-generics}
\alias{mod_cpt-generics}
\alias{as.ts.mod_cpt}
\alias{nobs.mod_cpt}
\alias{logLik.mod_cpt}
\alias{fitted.mod_cpt}
\alias{residuals.mod_cpt}
\alias{model_variance}
\alias{coef.mod_cpt}
\alias{augment.mod_cpt}
\alias{tidy.mod_cpt}
\alias{glance.mod_cpt}
\alias{plot.mod_cpt}
\alias{print.mod_cpt}
\title{Methods for mod_cpt objects}
\usage{
\method{as.ts}{mod_cpt}(x, ...)

\method{nobs}{mod_cpt}(object, ...)

\method{logLik}{mod_cpt}(object, ...)

\method{fitted}{mod_cpt}(object, ...)

\method{residuals}{mod_cpt}(object, ...)

model_variance(object, ...)

\method{coef}{mod_cpt}(object, ...)

\method{augment}{mod_cpt}(x, ...)

\method{tidy}{mod_cpt}(x, ...)

\method{glance}{mod_cpt}(x, ...)

\method{plot}{mod_cpt}(x, ...)

\method{print}{mod_cpt}(x, ...)
}
\arguments{
\item{x}{A \code{mod_cpt} object, typically the output from one of the \verb{fit_*()}
functions or \code{\link[=as.model]{as.model()}}}

\item{...}{some methods for this generic function require additional
   arguments.}

\item{object}{any object from which a log-likelihood value, or a
   contribution to a log-likelihood value, can be extracted.}
}
\description{
Methods for generic functions applied to \code{\link[=mod_cpt]{mod_cpt()}} objects
}
\details{
All changepoint detection models inherit from \linkS4class{mod_cpt}: the
base class for changepoint detection models.
These models are created by one of the \verb{fit_*()} functions, or by
\code{\link[=as.model]{as.model()}}.
The methods documented here work on these objects.
}
\examples{
cpts <- fit_meanshift_norm(DataCPSim, tau = 365)
as.ts(cpts)
nobs(cpts)
logLik(cpts)
fitted(cpts)
residuals(cpts)
changepoints(cpts)
augment(cpts)
tidy(cpts)
glance(cpts)
plot(fit_meanshift_norm(CET, tau = 330))
plot(fit_trendshift(CET, tau = 330))
plot(fit_lmshift(CET, tau = 330, deg_poly = 2))
plot(fit_lmshift(CET, tau = 330, deg_poly = 10))
}
\seealso{
\code{\link[=mod_cpt]{mod_cpt()}}

\code{\link[broom:reexports]{broom::augment()}}

\code{\link[broom:reexports]{broom::tidy()}}
}
