% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mod_default.R
\name{mod_default-generics}
\alias{mod_default-generics}
\alias{as.ts.mod_default}
\alias{nobs.mod_default}
\alias{logLik.mod_default}
\alias{fitted.mod_default}
\alias{residuals.mod_default}
\alias{model_variance}
\alias{augment.mod_default}
\alias{tidy.mod_default}
\alias{glance.mod_default}
\alias{plot.mod_default}
\title{Methods for mod_default objects}
\usage{
\method{as.ts}{mod_default}(x, ...)

\method{nobs}{mod_default}(object, ...)

\method{logLik}{mod_default}(object, ...)

\method{fitted}{mod_default}(object, ...)

\method{residuals}{mod_default}(object, ...)

model_variance(object, ...)

\method{augment}{mod_default}(x, ...)

\method{tidy}{mod_default}(x, ...)

\method{glance}{mod_default}(x, ...)

\method{plot}{mod_default}(x, ...)
}
\arguments{
\item{x}{A \code{mod_default} object, typically the output from one of the \verb{fit_*()}
functions}

\item{...}{some methods for this generic function require additional
   arguments.}

\item{object}{any object from which a log-likelihood value, or a
   contribution to a log-likelihood value, can be extracted.}

\item{use_time_index}{Should the x-axis labels be the time indices? Or the
time labels?}
}
\description{
Methods for mod_default objects
}
\examples{
cpts <- fit_meanshift(DataCPSim, tau = 365)
as.ts(cpts)
nobs(cpts)
logLik(cpts)
fitted(cpts)
residuals(cpts)
changepoints(cpts)
augment(cpts)
tidy(cpts)
glance(cpts)
plot(fit_meanshift(CET, tau = 300))
}
\seealso{
\code{\link[broom:reexports]{broom::augment()}}

\code{\link[broom:reexports]{broom::tidy()}}
}
