% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-seg_default.R, R/class-tidycpt.R,
%   R/generics.R, R/pkg-GA.R, R/pkg-changepoint.R, R/pkg-wbs.R
\name{model_fit.seg_default}
\alias{model_fit.seg_default}
\alias{model_fit.tidycpt}
\alias{model_fit}
\alias{model_fit.ga}
\alias{model_fit.cpt}
\alias{model_fit.wbs}
\title{Retrieve the model that a segmenter used}
\usage{
\method{model_fit}{seg_default}(object, ...)

\method{model_fit}{tidycpt}(object, ...)

model_fit(x, ...)

\method{model_fit}{ga}(object, ...)

\method{model_fit}{cpt}(object, ...)

\method{model_fit}{wbs}(object, ...)
}
\arguments{
\item{object}{An object, typically returned by \code{\link[=segment]{segment()}}}

\item{...}{currently ignored}
}
\description{
Retrieve the model that a segmenter used
}
