% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-FnAGBayesianMDL.R
\name{muta_k_cp_BMDL}
\alias{muta_k_cp_BMDL}
\title{Mutaciones un cp en el caso BMDL}
\usage{
muta_k_cp_BMDL(mat_cp, x, param)
}
\arguments{
\item{mat_cp}{}

\item{x}{}

\item{param}{}
}
\value{
regreas una mat_cp mutada
}
\description{
Mutaciones un cp en el caso BMDL
}
