% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_weibull.R
\name{mweibull}
\alias{mweibull}
\title{Evaluanción de mean function (la integral de la rate)}
\usage{
mweibull(x, shape, scale = 1)
}
\arguments{
\item{t}{valor a evaluar (real)}

\item{nhpp_dist}{tipo de mean function (ver libro de Eliane pdf 40 eq (3.12));
posibles exponentiated-Weibull (EW),the Musa–Okumoto (MO),
the Goel–Okumoto (GO), and a generalized Goel–Okumoto (GGO)}

\item{theta}{parámetros de mean function.Para
\itemize{
\item W: \eqn{\alpha = \theta[1], \sigma = \theta[2]}
\item EW: \eqn{\alpha= \theta[1], \beta = \theta[2]} y \eqn{\sigma = \theta[3]}
\item GGO: \eqn{\alpha = \theta[1], \beta = \theta[2]} y \eqn{\gamma = \theta[3]}
\item MO: \eqn{\alpha = \theta[1], \beta = \theta[2]}
\item GO: \eqn{\alpha = \theta[1], \beta = \theta[2]}
}}
}
\value{
regresa la evaluación
}
\description{
Evaluanción de mean function (la integral de la rate)
}
\examples{
mweibull(1, shape = 1, scale = 1)
plot(x = 1:10, y = mweibull(1:10, shape = 1, scale = 1))
plot(x = 1:10, y = mweibull(1:10, shape = 1, scale = 2))
plot(x = 1:10, y = mweibull(1:10, shape = 0.5, scale = 2))
plot(x = 1:10, y = mweibull(1:10, shape = 0.5, scale = 100))
plot(x = 1:10, y = mweibull(1:10, shape = 2, scale = 2))
plot(x = 1:10, y = mweibull(1:10, shape = 2, scale = 100))

}
