% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpt_gbmdl.R
\name{new_cpt_gbmdl}
\alias{new_cpt_gbmdl}
\alias{validate_cpt_gbmdl}
\alias{cpt_gbmdl}
\alias{plot.cpt_gbmdl}
\alias{print.cpt_gbmdl}
\alias{summary.cpt_gbmdl}
\alias{cpt_best_bmdl}
\alias{cpt_best_bmdl_string}
\alias{cpt_best_params}
\alias{chromosome_best}
\alias{cpt_best}
\alias{chromo2tau}
\alias{mat_cp_2_tbl}
\alias{write_cpt_gbmdl}
\title{Class for candidate changepoints using Genetic BMDL heuristic}
\usage{
new_cpt_gbmdl(
  x = numeric(),
  param = list(),
  num_generations = 50,
  generation_size = 50,
  max_num_cp = 20
)

validate_cpt_gbmdl(x)

cpt_gbmdl(x, ...)

\method{plot}{cpt_gbmdl}(x, ...)

\method{print}{cpt_gbmdl}(x, ...)

\method{summary}{cpt_gbmdl}(object, ...)

cpt_best_bmdl(x)

cpt_best_bmdl_string(x)

cpt_best_params(x)

chromosome_best(x)

cpt_best(x)

chromo2tau(chromo)

mat_cp_2_tbl(mat_cp)

write_cpt_gbmdl(x, destdir = tempdir(), data_name_slug = "data")
}
\arguments{
\item{x}{a numeric vector}

\item{num_generations}{número de generaciones}

\item{generation_size}{tamaño de las generaciones}

\item{data_name_slug}{character string that will identify the data set used
in the file name}
}
\description{
Class for candidate changepoints using Genetic BMDL heuristic
}
\examples{
cpts <- cpt_gbmdl(DataCPSim, param)
str(cpts)
cpt_best_params(lista_AG$segmenter)
chromo <- c(4, 1, 557, 877 , 905, 986, 1096, 0, 0, 0)
chromo2tau(chromo)
write_cpt_gbmdl(lista_AG$segmenter)
}
