% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mod_cpt.R
\name{new_mod_cpt}
\alias{new_mod_cpt}
\alias{validate_mod_cpt}
\alias{mod_cpt}
\title{Base class for changepoint models}
\usage{
new_mod_cpt(
  x = numeric(),
  tau = integer(),
  region_params = tibble::tibble(),
  model_params = double(),
  fitted_values = double(),
  model_name = character(),
  ...
)

validate_mod_cpt(x)

mod_cpt(x, ...)
}
\arguments{
\item{x}{a numeric vector coercible into a \code{ts} object}

\item{tau}{indices of the changepoint set}

\item{region_params}{A \code{\link[tibble:tibble]{tibble::tibble()}} with one row for each region
defined by the changepoint set \code{tau}. Each variable represents a parameter
estimated in that region.}

\item{model_params}{A numeric vector of parameters estimated by the model
across the entire data set (not just in each region).}

\item{fitted_values}{Fitted values returned by the model on the original
data set.}

\item{model_name}{A \code{character} vector giving the model's name.}

\item{...}{currently ignored}
}
\description{
Base class for changepoint models
}
\examples{
cpt <- mod_cpt(CET)
str(cpt)
as.ts(cpt)
changepoints(cpt)
}
