% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/seg_default.R
\name{evaluate_cpts}
\alias{evaluate_cpts}
\alias{new_seg_default}
\alias{validate_seg_default}
\alias{seg_default}
\alias{num_candidates}
\alias{evaluate_cpts.seg_default}
\alias{evaluate_cpts.list}
\alias{evaluate_cpts.tbl_df}
\title{Default class for candidate changepoint sets}
\usage{
evaluate_cpts(x, ...)

new_seg_default(
  x = numeric(),
  algorithm = NA,
  cpt_list = list(),
  params = list(),
  ...
)

validate_seg_default(x)

seg_default(x, ...)

num_candidates(x)

\method{evaluate_cpts}{seg_default}(x, ...)

\method{evaluate_cpts}{list}(x, ...)

\method{evaluate_cpts}{tbl_df}(x, ...)
}
\arguments{
\item{x}{a numeric vector coercible into a \code{ts} object}

\item{...}{currently ignored}

\item{algorithm}{Algorithm used to find the changepoints}

\item{cpt_list}{a possibly empty \code{list()} of candidate changepoints}

\item{params}{a possibly empty \code{list()} of parameters}
}
\description{
Default class for candidate changepoint sets
}
\examples{
seg <- seg_default(DataCPSim)
str(seg)
as.ts(seg)
changepoints(seg)
}
