% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_cpt.R
\name{new_tbl_cpt}
\alias{new_tbl_cpt}
\alias{validate_tbl_cpt}
\alias{tbl_cpt}
\alias{plot.tbl_cpt}
\alias{summarize2}
\alias{summarize2.tbl_cpt}
\alias{print.tbl_cpt}
\alias{summary.tbl_cpt}
\title{Class for segmented time series}
\usage{
new_tbl_cpt(x = numeric(), segmenter = NULL, ...)

validate_tbl_cpt(x)

tbl_cpt(x, ...)

\method{plot}{tbl_cpt}(x, ...)

summarize2(.data, ...)

\method{summarize2}{tbl_cpt}(.data, ...)

\method{print}{tbl_cpt}(x, ...)

\method{summary}{tbl_cpt}(object, ...)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
Class for segmented time series
}
\examples{
seg <- tbl_cpt(DataCPSim)
str(seg)
summarize2(seg)
plot(seg)
}
