% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pad_tau}
\alias{pad_tau}
\alias{unpad_tau}
\alias{is_valid_tau}
\alias{validate_tau}
\title{Pad and unpad changepoint sets}
\usage{
pad_tau(tau, n)

unpad_tau(padded_tau)

is_valid_tau(tau, n)

validate_tau(tau, n)
}
\arguments{
\item{tau}{a numeric vector of changepoint indices}

\item{n}{the length of the original time series}

\item{padded_tau}{Output from \code{\link[=pad_tau]{pad_tau()}}}
}
\value{
\itemize{
\item \code{\link[=pad_tau]{pad_tau()}}: an \code{integer} vector that starts with 0 and ends in \eqn{n}.
}

\itemize{
\item \code{\link[=unpad_tau]{unpad_tau()}}: an \code{integer} vector stripped of its first and last entries.
}

\itemize{
\item \code{\link[=is_valid_tau]{is_valid_tau()}}: a \code{logical} if all of the entries are between 2 and
\eqn{n-1}.
}

\itemize{
\item \code{\link[=validate_tau]{validate_tau()}}: an \code{integer} vector with only the \code{\link[base:unique]{base::unique()}}
entries between 2 and \eqn{n-1}, inclusive.
}
}
\description{
Pad and unpad changepoint sets
}
\examples{
is_valid_tau(0, length(DataCPSim))
is_valid_tau(1, length(DataCPSim))
is_valid_tau(c(42, 42), length(DataCPSim))
is_valid_tau(826, length(DataCPSim))
is_valid_tau(1096, length(DataCPSim))
is_valid_tau(1097, length(DataCPSim))

validate_tau(0, length(DataCPSim))
validate_tau(1, length(DataCPSim))
validate_tau(826, length(DataCPSim))
validate_tau(c(826, 826), length(DataCPSim))
validate_tau(1096, length(DataCPSim))
validate_tau(1097, length(DataCPSim))
validate_tau(c(-4, 0, 1, 4, 5, 5, 824, 1096, 1097, 182384), length(DataCPSim))

}
