% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalty.R
\name{penalization_MDL}
\alias{penalization_MDL}
\title{Penalización MDL}
\usage{
penalization_MDL(cp, nhpp_dist = "W", N)
}
\arguments{
\item{cp}{vector de extendido de puntos de cambio}
}
\description{
Penalización MDL
}
\examples{
mat_cp <- sim_k_cp_BMDL(DataCPSim)
penalization_MDL(mat_cp, N = max(exceedances(DataCPSim)))


}
