% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalty.R
\name{penalty_mdl}
\alias{penalty_mdl}
\title{Penalizaci贸n MDL}
\usage{
penalty_mdl(padded_tau, N = NULL)
}
\arguments{
\item{padded_tau}{Output from \code{\link[=pad_tau]{pad_tau()}}}

\item{N}{The length of the time series?}
}
\value{
regresa la evaluaci贸n de la penalizaci贸n
\deqn{
   P_{\theta,\tau} = \sum_{i=1}^{m+1}\dfrac{\ln(\tau_i-\tau_{i-1})}{2}+\ln(m)+\sum_{i=2}^m\ln(\tau_i)
 }
}
\description{
Penalizaci贸n MDL
}
\examples{
x <- segment(DataCPSim, method = "pelt")
tau <- changepoints(x)
penalty_mdl(pad_tau(tau, n = length(x)))

}
