% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-GA.R
\name{plot.tidyga}
\alias{plot.tidyga}
\title{Plot GA information}
\usage{
\method{plot}{tidyga}(x, ...)
}
\arguments{
\item{x}{A \code{tidyga} object}

\item{...}{currently ignored}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Plot GA information
}
\examples{
\donttest{
x <- segment(DataCPSim, method = "ga-coen", maxiter = 5)
plot(x$segmenter)
}
}
