% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_confint}
\alias{plot_confint}
\title{Graficas de los intervalos y los datos}
\usage{
plot_confint(cpt_list)
}
\arguments{
\item{lista_AG}{description}
}
\description{
Graficas de los intervalos y los datos
}
\examples{
plot_confint(lista_AG$segmenter)

}
