% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpt_gbmdl_graphics.R
\name{plot_cpt_repetidos}
\alias{plot_cpt_repetidos}
\title{Grafica los n puntos de cambio más repetidos sobre el plot.stepfun de los datos}
\usage{
plot_cpt_repetidos(
  x,
  destdir = tempdir(),
  data_name_slug = "data",
  pdf = FALSE,
  limit = 100
)
}
\arguments{
\item{x}{description}

\item{limit}{al final se generan unas graficas de los
mejores puntos de cambio, este parámetro dicta cuantos cromosomas se
graficarán}
}
\value{
guarda un pdf con la gráfica
}
\description{
Grafica los n puntos de cambio más repetidos sobre el plot.stepfun de los datos
}
\examples{
plot_cpt_repetidos(lista_AG$segmenter)

}
