% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpt_gbmdl_graphics.R
\name{plot_exceedances}
\alias{plot_exceedances}
\title{Grafica escalonada para un proceso poisson}
\usage{
plot_exceedances(x, ...)
}
\arguments{
\item{x}{a time series vector}
}
\description{
Grafica escalonada para un proceso poisson
}
\examples{
plot_exceedances(DataCPSim)
plot_exceedances(rlnorm_ts_1)
plot_exceedances(rlnorm_ts_2)
plot_exceedances(rlnorm_ts_3)

}
