% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-seg_basket.R
\name{plot_history}
\alias{plot_history}
\alias{plot_best_chromosome}
\alias{plot_cpt_repeated}
\title{Plot seg_basket information}
\usage{
plot_history(x, ...)

plot_best_chromosome(x)

plot_cpt_repeated(x, i = nrow(x$basket))
}
\arguments{
\item{x}{A \code{seg_basket} object}

\item{...}{currently ignored}

\item{i}{index of basket to show}
}
\description{
Plot seg_basket information
}
\examples{
x <- segment(DataCPSim, method = "random", popSize = 10)
plot_history(x$segmenter)
x <- segment(DataCPSim, method = "random", num_generations = 10)
plot_best_chromosome(x$segmenter)
x <- segment(DataCPSim, method = "random", k = 10)
plot_cpt_repeated(x$segmenter)
plot_cpt_repeated(x$segmenter, 5)
}
