% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pm_25}
\alias{pm_25}
\title{Mediciones}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1096 rows and 1 columns.
}
\usage{
pm_25
}
\description{
A short description...
}
\keyword{datasets}
