% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pow}
\alias{pow}
\title{pow}
\usage{
pow(x, y)
}
\arguments{
\item{n_puntos_cambio}{numero de puntos de cambio}

\item{mas_derecha}{es una variable que fue utilizada para que
pudiera ejecutarse en R y en JAGS; toma los valores "" (vacío) y "+".}
}
\value{
Regresa el texto correspondiente a la variable m
}
\description{
pow
}
