% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_pow_d}
\alias{print_pow_d}
\title{print_pow_d}
\usage{
print_pow_d(i_d, i_sig_alp)
}
\arguments{
\item{n_puntos_cambio}{numero de puntos de cambio}

\item{mas_derecha}{es una variable que fue utilizada para que
pudiera ejecutarse en R y en JAGS; toma los valores "" (vacío) y "+".}
}
\value{
Regresa el texto correspondiente a la variable m
}
\description{
print_pow_d
}
