% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_bmdl.R
\name{probs_vec_MDL}
\alias{probs_vec_MDL}
\title{Probabilidades a partir de mat_MDL}
\usage{
probs_vec_MDL(vec_MDL, probs_rank0_MDL1)
}
\arguments{
\item{vec_MDL}{vector con valores MDL

OBSERVACIÓN: Esto regresa numeros negativos, los cuales mientras más negativo mejor, ya que
dará que es un mejor vector de tiempos de cambio. Es decir, un MDL de -6000 es
mejor que -4000}

\item{probs_rank0_MDL1}{description}
}
\value{
regresa un vector de probabilidades
}
\description{
Probabilidades a partir de mat_MDL
}
