% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-RandomKeys.R
\name{RandomKeys}
\alias{RandomKeys}
\title{Generate random keys}
\usage{
RandomKeys(N)
}
\arguments{
\item{N}{number of keys}
}
\description{
Generate random keys
}
\examples{
RandomKeys(5)
RandomKeys(24)

}
