% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{rate_fn}
\alias{rate_fn}
\title{Evaluanción de rate function (la derivada de la mean)}
\usage{
rate_fn(t, rf_type, theta)
}
\arguments{
\item{t}{valor a evaluar (real)}

\item{rf_type}{tipo de rate function (ver articulo A Gibbs Sampling Algorithm
to Estimate the Occurrence of Ozone Exceedances in Mexico City  para el
caso de Weibull (W) y ver libro de Eliane pdf 40 eq (3.12)); posibles
exponentiated-Weibull (EW),the Musa–Okumoto (MO), the Goel–Okumoto (GO),
and a generalized Goel–Okumoto (GGO). Las cuales tienen 3,3,2 y 2
parámetros}

\item{theta}{parámetros de mean function. Para
\itemize{
\item W: \eqn{\alpha = \theta[1]}, \eqn{\sigma = \theta[2]}
\item EW: \eqn{\alpha = \theta[1], \beta = \theta[2]} y \eqn{\sigma = \theta[3]}
\item GGO: \eqn{\alpha = \theta[1], \beta = \theta[2]} y \eqn{\gamma = \theta[3]}
\item MO: \eqn{\alpha = \theta[1], \beta = \theta[2]}
\item GO: \eqn{\alpha = \theta[1], \beta = \theta[2]}
}}
}
\value{
regresa la evaluación
}
\description{
Evaluanción de rate function (la derivada de la mean)
}
\examples{
t <- c(1.4, 2.8)
theta <- c(1.2, 2.1, 3.2)

rf_type <- "EW"
rate_fn(t, rf_type, theta)
rf_type <- "GGO"
rate_fn(t, rf_type, theta)
rf_type <- "MO"
rate_fn(t, rf_type, theta)
rf_type <- "GO"
rate_fn(t, rf_type, theta)

}
