% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mod_cpt.R, R/class-seg_basket.R,
%   R/class-seg_cpt.R, R/class-tidycpt.R, R/mod_meanshift.R, R/mod_nhpp.R,
%   R/pkg-GA.R, R/pkg-changepoint.R, R/pkg-changepointGA.R, R/pkg-segmented.R,
%   R/pkg-strucchange.R, R/pkg-wbs.R, R/reexports.R
\docType{import}
\name{as.ts.mod_cpt}
\alias{as.ts.mod_cpt}
\alias{nobs.mod_cpt}
\alias{logLik.mod_cpt}
\alias{fitted.mod_cpt}
\alias{residuals.mod_cpt}
\alias{coef.mod_cpt}
\alias{augment.mod_cpt}
\alias{tidy.mod_cpt}
\alias{glance.mod_cpt}
\alias{plot.mod_cpt}
\alias{print.mod_cpt}
\alias{summary.mod_cpt}
\alias{as.ts.seg_basket}
\alias{plot.seg_basket}
\alias{as.ts.seg_cpt}
\alias{glance.seg_cpt}
\alias{nobs.seg_cpt}
\alias{print.seg_cpt}
\alias{summary.seg_cpt}
\alias{as.ts.tidycpt}
\alias{augment.tidycpt}
\alias{tidy.tidycpt}
\alias{glance.tidycpt}
\alias{plot.tidycpt}
\alias{print.tidycpt}
\alias{summary.tidycpt}
\alias{logLik.meanshift_lnorm}
\alias{logLik.nhpp}
\alias{glance.nhpp}
\alias{as.ts.ga}
\alias{nobs.ga}
\alias{as.ts.cpt}
\alias{logLik.cpt}
\alias{nobs.cpt}
\alias{as.ts.cptga}
\alias{nobs.cptga}
\alias{as.ts.segmented}
\alias{as.ts.stepmented}
\alias{as.ts.breakpointsfull}
\alias{nobs.breakpointsfull}
\alias{as.ts.wbs}
\alias{nobs.wbs}
\alias{reexports}
\alias{augment}
\alias{tidy}
\alias{glance}
\title{Objects exported from other packages}
\usage{
\method{as.ts}{mod_cpt}(x, ...)

\method{nobs}{mod_cpt}(object, ...)

\method{logLik}{mod_cpt}(object, ...)

\method{fitted}{mod_cpt}(object, ...)

\method{residuals}{mod_cpt}(object, ...)

\method{coef}{mod_cpt}(object, ...)

\method{augment}{mod_cpt}(x, ...)

\method{tidy}{mod_cpt}(x, ...)

\method{glance}{mod_cpt}(x, ...)

\method{plot}{mod_cpt}(x, ...)

\method{print}{mod_cpt}(x, ...)

\method{summary}{mod_cpt}(object, ...)

\method{as.ts}{seg_basket}(x, ...)

\method{plot}{seg_basket}(x, ...)

\method{as.ts}{seg_cpt}(x, ...)

\method{glance}{seg_cpt}(x, ...)

\method{nobs}{seg_cpt}(object, ...)

\method{print}{seg_cpt}(x, ...)

\method{summary}{seg_cpt}(object, ...)

\method{as.ts}{tidycpt}(x, ...)

\method{augment}{tidycpt}(x, ...)

\method{tidy}{tidycpt}(x, ...)

\method{glance}{tidycpt}(x, ...)

\method{plot}{tidycpt}(x, use_time_index = FALSE, ylab = NULL, ...)

\method{print}{tidycpt}(x, ...)

\method{summary}{tidycpt}(object, ...)

\method{logLik}{meanshift_lnorm}(object, ...)

\method{logLik}{nhpp}(object, ...)

\method{glance}{nhpp}(x, ...)

\method{as.ts}{ga}(x, ...)

\method{nobs}{ga}(object, ...)

\method{as.ts}{cpt}(x, ...)

\method{logLik}{cpt}(object, ...)

\method{nobs}{cpt}(object, ...)

\method{as.ts}{cptga}(x, ...)

\method{nobs}{cptga}(object, ...)

\method{as.ts}{segmented}(x, ...)

\method{as.ts}{stepmented}(x, ...)

\method{as.ts}{breakpointsfull}(x, ...)

\method{nobs}{breakpointsfull}(object, ...)

\method{as.ts}{wbs}(x, ...)

\method{nobs}{wbs}(object, ...)
}
\arguments{
\item{...}{some methods for this generic function require additional
   arguments.}

\item{object}{any object from which a log-likelihood value, or a
   contribution to a log-likelihood value, can be extracted.}

\item{use_time_index}{Should the x-axis labels be the time indices? Or the
time labels?}
}
\examples{
# Plot a meanshift model fit
plot(fit_meanshift_norm(CET, tau = 330))
plot(fit_meanshift_norm(CET, tau = 330), plot.title.position = "plot")

#' # Plot a trendshift model fit
plot(fit_trendshift(CET, tau = 330))

#' # Plot a quadratic polynomial model fit
plot(fit_lmshift(CET, tau = 330, deg_poly = 2))

#' # Plot a 4th degree polynomial model fit
plot(fit_lmshift(CET, tau = 330, deg_poly = 10))

# Plot a segmented time series
plot(segment(CET, method = "pelt"))

# Plot a segmented time series and show the time labels on the x-axis
plot(segment(CET, method = "pelt"), use_time_index = TRUE)

# Label the y-axis correctly
segment(CET, method = "pelt") |>
  plot(use_time_index = TRUE, ylab = "Degrees Celsius")
# Summarize a tidycpt object
summary(segment(CET, method = "pelt"))
summary(segment(DataCPSim, method = "pelt"))
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{augment}}, \code{\link[generics]{glance}}, \code{\link[generics]{tidy}}}
}}

