% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_bmdl.R
\name{revisor_param}
\alias{revisor_param}
\title{Validador de la lista param para ejecutar el AG-BMDL}
\usage{
revisor_param(
  param,
  r = 50,
  k = 50,
  max_num_cp = 40,
  rf_type = c("W", "EW", "GGO", "MO", "GO")[1],
  vec_dist_a_priori = c("Gamma", "Gamma"),
  mat_phi = matrix(c(1, 3, 2, 1.2), ncol = 2)
)
}
\arguments{
\item{param}{es la lista original de parámetros la cual contiene todos los
siguientes}

\item{r}{número de generaciones}

\item{k}{tamaño de las generaciones}

\item{max_num_cp}{el máximo número de rebases. Este parámetro se ocupa en
particular para que todos los cromosomas quepan en una matriz.}

\item{rf_type}{toma valores en c("W","EW","GGO","MO","GO") y es el nombre de
la función de tasa del NHPP}

\item{vec_dist_a_priori}{vector de los nobmres de las distribuciones a priori
que se utilizan; eg c("Gamma","Gamma") y c("Gamma","Gamma","Gamma")}

\item{mat_phi}{matriz cuyos renglones tiene los parámetros de las
distribuciones a priori; cada renglón tiene todos los parametros de una
distribución}
}
\value{
regresa un mensaje en caso de que no sea de la longitud correcta el
vector
}
\description{
Validador de la lista param para ejecutar el AG-BMDL
}
\examples{
revisor_param(param)
}
