% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rlnorm_ts_1}
\alias{rlnorm_ts_1}
\alias{rlnorm_ts_2}
\alias{rlnorm_ts_3}
\title{Simulated time series data}
\format{
An object of class \code{ts} of length 1096.

An object of class \code{ts} of length 1096.

An object of class \code{ts} of length 1096.
}
\usage{
rlnorm_ts_1

rlnorm_ts_2

rlnorm_ts_3
}
\description{
Randomly-generated time series data, using the \code{\link[stats:Lognormal]{stats::rlnorm()}} function.
\itemize{
\item For \code{ts_1}, there is one changepoint located at 826.
\item For \code{ts_2}, there are two changepoints, located at 366 and 731.
\item For \code{ts_3}, there are three changepoints, located at 548, 823, and 973.
}
}
\examples{
plot(rlnorm_ts_1)
plot(rlnorm_ts_2)
plot(rlnorm_ts_3)

}
\seealso{
\code{\link[stats:ts]{stats::ts()}}
}
\keyword{datasets}
