% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-seg_cpt.R
\name{as.seg_cpt.seg_cpt}
\alias{as.seg_cpt.seg_cpt}
\alias{seg-cpt-generics}
\alias{as.ts.seg_cpt}
\alias{glance.seg_cpt}
\alias{nobs.seg_cpt}
\alias{print.seg_cpt}
\alias{seg_params.seg_cpt}
\title{Methods for seg_cpt objects}
\usage{
\method{as.seg_cpt}{seg_cpt}(object, ...)

\method{as.ts}{seg_cpt}(x, ...)

\method{glance}{seg_cpt}(x, ...)

\method{nobs}{seg_cpt}(object, ...)

\method{print}{seg_cpt}(x, ...)

\method{seg_params}{seg_cpt}(x, ...)
}
\arguments{
\item{object}{A \code{seg_cpt} object}

\item{...}{arguments passed to methods}

\item{x}{An \code{seg_cpt} object}
}
\description{
Methods for seg_cpt objects
}
