% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_boltzmann.R
\name{segment_boltzmann}
\alias{segment_boltzmann}
\alias{MaMBOltzmann}
\title{Segment a times series using Boltzmann criteria for a genetic algorithm}
\usage{
segment_boltzmann(x, r = 10, k = 50, Mutation = 0.03, Temperature = 27)

MaMBOltzmann(mat_cp, Fitness, Mutation, Temperature)
}
\arguments{
\item{x}{An \code{integer} vector. See \code{\link[=exceedances]{exceedances()}}.}

\item{r}{number of generations}

\item{k}{size of generation}

\item{Mutation}{description}

\item{Temperature}{description}

\item{mat_cp}{description}

\item{Fitness}{description}
}
\value{
A \code{list} of length 2:
\itemize{
\item A \code{numeric} vector of \code{FitnessGeneraciones}
\item A \code{matrix} with \code{k} rows and \code{N} columns
}
}
\description{
What does this function do?
}
\examples{
\dontrun{
x <- segment_boltzmann(DataCPSim, r = 5)
y <- segment_boltzmann(rlnorm_ts_1)
  
}
}
