% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_ga_shi.R
\name{segment_ga_shi}
\alias{segment_ga_shi}
\title{Segment a time series using a genetic algorithm}
\usage{
segment_ga_shi(x, ...)
}
\arguments{
\item{x}{A time series}

\item{...}{arguments passed to \code{\link[GA:ga]{GA::ga()}}}
}
\description{
Segment a time series using a genetic algorithm
}
\examples{
\dontrun{
x <- england_temperature$annual_mean_temp
res <- segment_ga_shi(x, maxiter = 50)
summary(res)
str(res)
plot(res)
}
}
\author{
Xueheng Shi
}
