% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_manual.R
\name{segment_manual}
\alias{segment_manual}
\title{Manually segment a times series}
\usage{
segment_manual(x, tau, ...)
}
\arguments{
\item{x}{a numeric vector coercible into a \link[stats:ts]{stats::ts} object}

\item{...}{arguments passed to methods}
}
\description{
Manually segment a times series
}
\examples{
segment_manual(CET, tau = c(84, 330))
segment_manual(CET, tau = NULL)

}
