% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_random.R
\name{segment_random}
\alias{segment_random}
\alias{random_cpts}
\title{Segment a times series using randomly selected candidate changepoints}
\usage{
segment_random(x, num_generations = 10, ...)

random_cpts(x, n = 10, k = round(log(length(as.ts(x)))), ...)
}
\arguments{
\item{x}{A times series}

\item{num_generations}{Number of generations}

\item{...}{currently ignored}

\item{n}{Number of random changepoints sets to generate}

\item{k}{Rate function for Poisson distribution of mean changepoint set length.}
}
\description{
Segment a times series using randomly selected candidate changepoints
}
\examples{
segment_random(DataCPSim, 10)

random_cpts(DataCPSim, n = 5)

}
