% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_bmdl.R
\name{selec_k_pares_de_padres}
\alias{selec_k_pares_de_padres}
\title{Seleciona k pares de padres}
\usage{
selec_k_pares_de_padres(vec_probs)
}
\arguments{
\item{vec_probs}{vector de probabilidades de selección de cada uno de los
cromosomas}
}
\description{
Seleciona k pares de padres
}
