% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-FnAGBayesianMDL.R
\name{sim_k_cp_BMDL}
\alias{sim_k_cp_BMDL}
\title{Simula k vectores change point para Bayesian MDL}
\usage{
sim_k_cp_BMDL(x, param)
}
\arguments{
\item{x}{vector de excedentes}

\item{param}{lista de parámetros globales}
}
\value{
regresa una matriz de k por max_num_cp+3, la cual en cada renglón tiene
una simulación de un vector de tiempos de cambio
}
\description{
Simula k vectores change point para Bayesian MDL
}
