% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sims}
\alias{sims}
\title{Simulated time series data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1096 rows and 3 columns.
}
\usage{
sims
}
\description{
Simulated time series data
}
\keyword{datasets}
