% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom.R
\name{tidy}
\alias{tidy}
\alias{augment}
\alias{glance}
\title{Broom compatibility layer}
\usage{
tidy(x, ...)

augment(x, ...)

glance(x, ...)
}
\description{
Broom compatibility layer
}
\seealso{
\code{\link[broom:reexports]{broom::tidy()}}

\code{\link[broom:reexports]{broom::augment()}}

\code{\link[broom:reexports]{broom::glance()}}
}
