% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-tidycpt.R
\name{tidycpt-generics}
\alias{tidycpt-generics}
\alias{as.ts.tidycpt}
\alias{nobs.tidycpt}
\alias{augment.tidycpt}
\alias{tidy.tidycpt}
\alias{glance.tidycpt}
\alias{plot.tidycpt}
\title{Generic functions for tidycpt objects}
\usage{
\method{as.ts}{tidycpt}(x, ...)

\method{nobs}{tidycpt}(object, ...)

\method{augment}{tidycpt}(x, ...)

\method{tidy}{tidycpt}(x, ...)

\method{glance}{tidycpt}(x, ...)

\method{plot}{tidycpt}(x, use_time_index = FALSE, ...)
}
\arguments{
\item{x}{An \code{tidycpt} object, typically the output from \code{\link[=segment]{segment()}}`}

\item{...}{arguments passed to methods}

\item{object}{A fitted model object.}

\item{use_time_index}{Should the x-axis labels be the time indices? Or the
time labels?}
}
\description{
Generic functions for tidycpt objects
}
\examples{
plot(segment(CET, method = "pelt"))
plot(segment(CET, method = "pelt"), use_time_index = TRUE)
}
\seealso{
\code{\link[stats:nobs]{stats::nobs()}}

\code{\link[broom:reexports]{broom::augment()}}

\code{\link[broom:reexports]{broom::tidy()}}

\code{\link[broom:reexports]{broom::glance()}}
}
