% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidychangepoint-package.R
\name{tidycpt_coverage}
\alias{tidycpt_coverage}
\title{Algorithmic coverage through tidychangepoint}
\usage{
tidycpt_coverage()
}
\value{
A \link[tibble:tibble]{tibble::tibble} showing the coverage of \code{tidychangepoint}
}
\description{
Algorithmic coverage through tidychangepoint
}
\examples{
tidycpt_coverage()

}
