% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vec_ptype2.logLik.logLik}
\alias{vec_ptype2.logLik.logLik}
\title{Vectors implementation for logLik}
\usage{
\method{vec_ptype2}{logLik.logLik}(x, y, ...)
}
\arguments{
\item{x, y}{Vector types.}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
Vectors implementation for logLik
}
\examples{
a <- logLik(lm(mpg ~ disp, data = mtcars))
b <- logLik(lm(mpg ~ am, data = mtcars))
vec_ptype2(a, b)
c(a, b)
dplyr::bind_rows(tibble::tibble(ll = a), tibble::tibble(ll = b))
vec_cast(a, b)
}
\seealso{
\code{\link[stats:logLik]{stats::logLik()}}
}
