% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{which_over_mean}
\alias{which_over_mean}
\title{Calculate overages}
\usage{
which_over_mean(x)
}
\arguments{
\item{x}{a time series object}
}
\description{
Calculate overages
}
\examples{
which_over_mean(DataCPSim)
}
