% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-FnAGBayesianMDL.R
\name{AG_BMDL_1_paso}
\alias{AG_BMDL_1_paso}
\title{Algoritmo genético de Bayesian MDL a un paso}
\usage{
AG_BMDL_1_paso(x, mat_cp, param)
}
\arguments{
\item{x}{}

\item{mat_cp}{}

\item{param}{}
}
\description{
Algoritmo genético de Bayesian MDL a un paso
}
