% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as.mod_default}
\alias{as.mod_default}
\title{Convert a segmenter to a \code{mod_default} object}
\usage{
as.mod_default(x, ...)
}
\arguments{
\item{...}{currently ignored}
}
\description{
Convert a segmenter to a \code{mod_default} object
}
