% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as.model}
\alias{as.model}
\title{Convert a segmenter to a model object}
\usage{
as.model(object, ...)
}
\arguments{
\item{object}{An object, typically returned by \code{\link[=segment]{segment()}}}

\item{...}{currently ignored}
}
\description{
Convert a segmenter to a model object
}
