% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/utils.R
\name{as.model}
\alias{as.model}
\alias{as.model.default}
\alias{is_model}
\alias{is_segmenter}
\title{Convert a segmenter to a model object}
\usage{
as.model(object, ...)

\method{as.model}{default}(object, ...)

is_model(x, ...)

is_segmenter(object, ...)
}
\arguments{
\item{object}{A segmenter object, typically returned by \code{\link[=segment]{segment()}}}

\item{...}{currently ignored}

\item{x}{An object, typically returned by \verb{fit_*()}}
}
\description{
Convert a segmenter to a model object
}
\examples{
cpt <- segment(DataCPSim, method = "ga", maxiter = 5)
as.model(cpt$segmenter)
cpt <- segment(DataCPSim, method = "ga-taimal", model_fn_args = list(threshold = 80), maxiter = 5)
as.model(cpt$segmenter)$model_params
x <- fit_nhpp(CET, tau = 330)
is_model(x)
x <- segment(CET, method = "pelt")
is_segmenter(x$segmenter)
}
