% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{as.segmenter}
\alias{as.segmenter}
\title{Convert a wild-caught segmenter object into a \code{segmenter} object}
\usage{
as.segmenter(object, ...)
}
\arguments{
\item{object}{An object resulting from a changepoint detection algorithm.}

\item{...}{currently ignored}
}
\description{
Convert a wild-caught segmenter object into a \code{segmenter} object
}
\seealso{
Other tidychangepoint-generics: 
\code{\link{changepoints.mod_cpt}()},
\code{\link{diagnose.seg_basket}()},
\code{\link{exceedances}()},
\code{\link{fitness.seg_basket}()}
}
\concept{tidychangepoint-generics}
