% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_year}
\alias{as_year}
\title{Convert a date into a year}
\usage{
as_year(x)
}
\arguments{
\item{x}{an object coercible into a \link[base:Dates]{base::Date}. See \code{\link[base:as.Date]{base::as.Date()}}.}
}
\value{
A \code{character} vector representing the years of the input
}
\description{
Convert a date into a year
}
\examples{
# Retrieve only the year
as_year("1988-01-01")

}
