% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom_cpt_gbmdl.R, R/changepoint.R
\name{tidy.cpt_gbmdl}
\alias{tidy.cpt_gbmdl}
\alias{glance.cpt_gbmdl}
\alias{as.ts.cpt_gbmdl}
\alias{augment.cpt}
\alias{tidy.cpt}
\alias{glance.cpt}
\alias{as.ts.cpt}
\alias{changepoints.cpt}
\title{Compatibility layer for changepoint}
\usage{
\method{tidy}{cpt_gbmdl}(x, ...)

\method{glance}{cpt_gbmdl}(x, ...)

\method{as.ts}{cpt_gbmdl}(x, ...)

\method{augment}{cpt}(x, ...)

\method{tidy}{cpt}(x, ...)

\method{glance}{cpt}(x, ...)

\method{as.ts}{cpt}(x, ...)

\method{changepoints}{cpt}(x, ...)
}
\description{
Compatibility layer for changepoint
}
\examples{
cpts <- lista_AG
class(cpts) <- c("cpt_bmdl", class(cpts))
as.ts(cpts)
cpts <- segment(DataCPSim, method = "cpt-pelt")
class(cpts)
y <- augment(cpts)
class(y)
y
tidy(cpts)
glance(cpts)
cpts <- segment(DataCPSim, method = "cpt-pelt")
as.ts(cpts)

cpts <- segment(DataCPSim, method = "cpt-pelt")
changepoints(cpts)

}
