% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpt_gbmdl2broom.R
\name{augment.cpt_gbmdl}
\alias{augment.cpt_gbmdl}
\title{Broom compatibility layer for changepoint}
\usage{
\method{augment}{cpt_gbmdl}(x, ...)
}
\description{
Broom compatibility layer for changepoint
}
\examples{
cpts <- lista_AG
class(cpts) <- c("cpt_gbmdl", class(cpts))
y <- augment(cpts)
class(y)
y
tidy(cpts)
glance(cpts)
}
