% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_bmdl.R
\name{Bayesaian_MDL_1_cp}
\alias{Bayesaian_MDL_1_cp}
\alias{Bayesaian_MDL_k_cp}
\title{Validador de la lista param para ejecutar el AG-BMDL}
\usage{
Bayesaian_MDL_1_cp(cp, x)

Bayesaian_MDL_k_cp(mat_cp, x)
}
\arguments{
\item{cp}{description}

\item{x}{description}

\item{param}{es la lista original de parámetros la cual contiene todos los
siguientes}

\item{rf_type}{description}

\item{vec_dist_a_priori}{description}

\item{mat_phi}{description}
}
\value{
regresa un vector de tamaño \code{k} (el numero de cromosomas por
generación) con los valores del bayesian MDL
}
\description{
Validador de la lista param para ejecutar el AG-BMDL

Bayesian MDL para un vector de puntos de cambio
}
