% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpt_gbmdl_algorithm.R
\name{Bayesaian_MDL_1_cp}
\alias{Bayesaian_MDL_1_cp}
\alias{Bayesaian_MDL_k_cp}
\title{Bayesian MDL para un vector de puntos de cambio}
\usage{
Bayesaian_MDL_1_cp(cp, x)

Bayesaian_MDL_k_cp(mat_cp, x)
}
\arguments{
\item{cp}{description}

\item{x}{description}
}
\value{
regresa un vector de tamaño \code{k} (el numero de cromosomas por
generación) con los valores del bayesian MDL
}
\description{
Bayesian MDL para un vector de puntos de cambio

Bayesian MDL para un vector de puntos de cambio
}
