% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_nhpp.R, R/seg_default.R, R/tidycpt.R
\name{BMDL.nhpp}
\alias{BMDL.nhpp}
\alias{BMDL.seg_default}
\alias{BMDL}
\alias{BMDL.tidycpt}
\title{Bayesian Maximum Descriptive Length}
\usage{
\method{BMDL}{nhpp}(object, ...)

\method{BMDL}{seg_default}(object, ...)

BMDL(object, ...)

\method{BMDL}{tidycpt}(object, ...)
}
\arguments{
\item{object}{any object from which a log-likelihood value, or a
   contribution to a log-likelihood value, can be extracted.}

\item{...}{some methods for this generic function require additional
   arguments.}
}
\description{
Bayesian Maximum Descriptive Length
}
\examples{
x <- segment(DataCPSim, method = "pelt")
BMDL(x)
y <- segment(DataCPSim, method = "manual", cpts = 826)
BMDL(y)
z <- segment(DataCPSim, method = "single-best")
BMDL(z)
}
\seealso{
\code{\link[=MDL]{MDL()}}
}
