% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/mod_nhpp.R
\name{BMDL}
\alias{BMDL}
\alias{BMDL.default}
\alias{BMDL.nhpp}
\title{Bayesian Maximum Descriptive Length}
\usage{
BMDL(object, ...)

\method{BMDL}{default}(object, ...)

\method{BMDL}{nhpp}(object, ...)
}
\arguments{
\item{object}{any object from which a log-likelihood value, or a
   contribution to a log-likelihood value, can be extracted.}

\item{...}{some methods for this generic function require additional
   arguments.}
}
\description{
Bayesian Maximum Descriptive Length
}
\examples{
x <- segment(DataCPSim, method = "pelt")
BMDL(x$nhpp)
y <- segment(DataCPSim, method = "manual", cpts = 826)
BMDL(y$nhpp)
z <- segment(DataCPSim, method = "single-best")
BMDL(z$nhpp)
}
\seealso{
\code{\link[=MDL]{MDL()}}
}
