% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalty.R
\name{bmdl}
\alias{bmdl}
\title{Bayesian MDL para un vector de puntos de cambio}
\usage{
bmdl(x, tau)
}
\arguments{
\item{x}{description}

\item{cp}{description}
}
\description{
Bayesian MDL para un vector de puntos de cambio
}
\examples{
bmdl(DataCPSim, tau = 826)
bmdl(DataCPSim, tau = changepoints(segment(DataCPSim, method = "cpt-pelt")))
bmdl(DataCPSim, tau = changepoints(segment(DataCPSim, method = "single-best")))

}
