% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/mod_nhpp.R
\name{BMDL}
\alias{BMDL}
\alias{BMDL.default}
\alias{BMDL.nhpp}
\title{Bayesian Maximum Descriptive Length}
\usage{
BMDL(object, ...)

\method{BMDL}{default}(object, ...)

\method{BMDL}{nhpp}(object, ...)
}
\arguments{
\item{object}{any object from which a log-likelihood value, or a
   contribution to a log-likelihood value, can be extracted.}

\item{...}{some methods for this generic function require additional
   arguments.}
}
\description{
Bayesian Maximum Descriptive Length
}
\examples{
BMDL(fit_nhpp(DataCPSim, tau = NULL))
BMDL(fit_nhpp(DataCPSim, tau = c(365, 830)))
}
\seealso{
\code{\link[=MDL]{MDL()}}
}
