% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bogota_pm}
\alias{bogota_pm}
\title{Particulate matter in Bogotá, Colombia}
\format{
An object of class \code{xts} (inherits from \code{zoo}) with 1096 rows and 1 columns.
}
\usage{
bogota_pm
}
\description{
Particulate matter of less than 2.5 microns of diameter in Bogotá, Colombia.
}
\details{
Daily readings from 2018-2020 are included.
}
\examples{
class(bogota_pm)

}
\keyword{datasets}
