% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-GA.R
\name{build_gabin_population}
\alias{build_gabin_population}
\title{Build an initial population set for GA algorithms}
\usage{
build_gabin_population(x, ...)
}
\arguments{
\item{x}{a numeric vector coercible into a \link[stats:ts]{stats::ts} object}

\item{...}{arguments passed to methods}
}
\description{
Build an initial population set for GA algorithms
}
\examples{
f <- build_gabin_population(CET)
segment(CET, method = "ga", population = f)
}
\seealso{
\code{\link[GA:ga_Population]{GA::gabin_Population()}}
}
