% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhpp.R
\name{cdf_exceedances_est}
\alias{cdf_exceedances_est}
\title{Mean cumulative exceedances function}
\usage{
cdf_exceedances_est(t, tau, theta, n, dist = "weibull")
}
\arguments{
\item{t}{The exceedances of a time series}

\item{tau}{A set of changepoints}

\item{theta}{A data.frame of best-fit parameters}

\item{n}{the legnth of the original time series}

\item{dist}{Name of the distribution}
}
\value{
a numeric vector of length equal to the \link{exceedances} of \code{x}
}
\description{
Mean cumulative exceedances function
}
\examples{
tau <- changepoints(lista_AG)
theta <- cpt_best_params(lista_AG$segmenter)
cdf_exceedances_est(exceedances(as.ts(lista_AG)), tau, theta, length(lista_AG))

}
