% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mod_cpt.R, R/class-seg_basket.R,
%   R/class-seg_cpt.R, R/class-tidycpt.R, R/generics.R, R/pkg-GA.R,
%   R/pkg-changepoint.R, R/pkg-wbs.R
\name{changepoints.mod_cpt}
\alias{changepoints.mod_cpt}
\alias{changepoints.seg_basket}
\alias{changepoints.seg_cpt}
\alias{changepoints.tidycpt}
\alias{changepoints_labels}
\alias{changepoints}
\alias{changepoints.default}
\alias{seg_params}
\alias{changepoints.ga}
\alias{changepoints.cpt}
\alias{changepoints.wbs}
\title{Extract changepoints}
\usage{
\method{changepoints}{mod_cpt}(x, ...)

\method{changepoints}{seg_basket}(x, ...)

\method{changepoints}{seg_cpt}(x, ...)

\method{changepoints}{tidycpt}(x, ...)

changepoints_labels(x, ...)

changepoints(x, ...)

\method{changepoints}{default}(x, ...)

seg_params(x, ...)

\method{changepoints}{ga}(x, ...)

\method{changepoints}{cpt}(x, ...)

\method{changepoints}{wbs}(x, ...)
}
\arguments{
\item{x}{A \code{tidycpt} object}

\item{...}{arguments passed to methods}
}
\description{
Not to be confused with \code{\link[wbs:changepoints]{wbs::changepoints()}}, which returns different
information.
}
\examples{
cpts <- segment(DataCPSim, method = "ga", maxiter = 5)
changepoints(cpts$segmenter)

cpts <- segment(DataCPSim, method = "wbs")
changepoints(cpts$segmenter)

}
\seealso{
\code{\link[wbs:changepoints]{wbs::changepoints()}}

Other tidychangepoint-generics: 
\code{\link{as.segmenter}()},
\code{\link{diagnose.seg_basket}()},
\code{\link{exceedances}()},
\code{\link{fitness.seg_basket}()}

Other tidychangepoint-generics: 
\code{\link{as.segmenter}()},
\code{\link{diagnose.seg_basket}()},
\code{\link{exceedances}()},
\code{\link{fitness.seg_basket}()}
}
\concept{tidychangepoint-generics}
