% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA.R, R/changepoint.R, R/mod_lmshift.R,
%   R/mod_meanshift.R, R/mod_nhpp.R, R/seg_default.R, R/tidycpt.R, R/wbs.R
\name{changepoints.ga}
\alias{changepoints.ga}
\alias{changepoints.cpt}
\alias{changepoints.lmshift}
\alias{changepoints.meanshift}
\alias{changepoints.nhpp}
\alias{changepoints.seg_default}
\alias{changepoints}
\alias{changepoints.tidycpt}
\alias{changepoints_labels}
\alias{params}
\alias{changepoints.wbs}
\title{Extract changepoints}
\usage{
\method{changepoints}{ga}(x, ...)

\method{changepoints}{cpt}(x, ...)

\method{changepoints}{lmshift}(x, ...)

\method{changepoints}{meanshift}(x, ...)

\method{changepoints}{nhpp}(x, ...)

\method{changepoints}{seg_default}(x, ...)

changepoints(x, ...)

\method{changepoints}{tidycpt}(x, ...)

changepoints_labels(x, ...)

params(x, ...)

\method{changepoints}{wbs}(x, ...)
}
\arguments{
\item{x}{A \code{tidycpt} object}

\item{...}{arguments passed to methods}
}
\description{
Extract changepoints
}
\examples{
cpts <- segment(DataCPSim, method = "ga", maxiter = 5)
changepoints(cpts$segmenter)

cpts <- segment(DataCPSim, method = "pelt")
changepoints(cpts)

cpts <- segment(DataCPSim, method = "wbs")
changepoints(cpts$segmenter)

}
