% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycpt.R
\name{changepoints}
\alias{changepoints}
\alias{changepoints.tidycpt}
\alias{as.ts.tidycpt}
\alias{length.tidycpt}
\alias{logLik.tidycpt}
\alias{MBIC}
\alias{MBIC.tidycpt}
\alias{nobs.tidycpt}
\alias{augment.tidycpt}
\alias{tidy.tidycpt}
\alias{glance.tidycpt}
\alias{regions}
\alias{regions.tidycpt}
\alias{plot.tidycpt}
\alias{plot_mcdf}
\alias{diagnose}
\alias{diagnose.tidycpt}
\title{Generic functions for tidycpt objects}
\usage{
changepoints(x, ...)

\method{changepoints}{tidycpt}(x, ...)

\method{as.ts}{tidycpt}(x, ...)

\method{length}{tidycpt}(x, ...)

\method{logLik}{tidycpt}(object, ...)

MBIC(object, ...)

\method{MBIC}{tidycpt}(object, ...)

\method{nobs}{tidycpt}(object, ...)

\method{augment}{tidycpt}(x, ...)

\method{tidy}{tidycpt}(x, ...)

\method{glance}{tidycpt}(x, ...)

regions(x, ...)

\method{regions}{tidycpt}(x, ...)

\method{plot}{tidycpt}(x, ...)

plot_mcdf(x, ...)

diagnose(x, ...)

\method{diagnose}{tidycpt}(x, ...)
}
\arguments{
\item{x}{A \code{tidycpt} object}

\item{...}{arguments passed to methods}

\item{object}{A \code{tidycpt} object}
}
\description{
Generic functions for tidycpt objects
}
\examples{
plot_mcdf(segment(DataCPSim))
diagnose(segment(DataCPSim))
diagnose(segment(DataCPSim, method = "single-best"))
diagnose(segment(DataCPSim, method = "cpt-pelt"))
diagnose(segment(test_set()))
diagnose(segment(test_set(n = 2, sd = 4), method = "cpt-pelt"))

}
\references{
Zhang and Seigmmund (2007) for MBIC: \doi{10.1111/j.1541-0420.2006.00662.x}
}
\seealso{
\code{\link[stats:logLik]{stats::logLik()}}

\code{\link[stats:nobs]{stats::nobs()}}

\code{\link[broom:reexports]{broom::augment()}}

\code{\link[broom:reexports]{broom::tidy()}}

\code{\link[broom:reexports]{broom::glance()}}
}
