% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_default.R, R/tidycpt.R
\name{plot_history}
\alias{plot_history}
\alias{changepoints}
\alias{changepoints.tidycpt}
\alias{changepoints_labels}
\alias{as.ts.tidycpt}
\alias{exceedances}
\alias{exceedances.tidycpt}
\alias{length.tidycpt}
\alias{logLik.tidycpt}
\alias{MBIC}
\alias{MBIC.tidycpt}
\alias{MDL}
\alias{BMDL}
\alias{BMDL.tidycpt}
\alias{nobs.tidycpt}
\alias{augment.tidycpt}
\alias{tidy.tidycpt}
\alias{glance.tidycpt}
\alias{compare_models}
\alias{compare_algorithms}
\alias{params}
\alias{plot.tidycpt}
\alias{diagnose}
\alias{diagnose.tidycpt}
\alias{file_name}
\title{Generic functions for tidycpt objects}
\usage{
plot_history(x, ...)

changepoints(x, ...)

\method{changepoints}{tidycpt}(x, ...)

changepoints_labels(x, ...)

\method{as.ts}{tidycpt}(x, ...)

exceedances(x, ...)

\method{exceedances}{tidycpt}(x, ...)

\method{length}{tidycpt}(x, ...)

\method{logLik}{tidycpt}(object, ...)

MBIC(object, ...)

\method{MBIC}{tidycpt}(object, ...)

MDL(object, ...)

BMDL(object, ...)

\method{BMDL}{tidycpt}(object, ...)

\method{nobs}{tidycpt}(object, ...)

\method{augment}{tidycpt}(x, ...)

\method{tidy}{tidycpt}(x, ...)

\method{glance}{tidycpt}(x, ...)

compare_models(x, ...)

compare_algorithms(x, ...)

params(x, ...)

\method{plot}{tidycpt}(x, ...)

diagnose(x, ...)

\method{diagnose}{tidycpt}(x, ...)

file_name(x, data_name_slug = "data")
}
\arguments{
\item{x}{A \code{tidycpt} object}

\item{...}{arguments passed to methods}

\item{object}{A \code{tidycpt} object}

\item{data_name_slug}{character string that will identify the data set used
in the file name}
}
\description{
Generic functions for tidycpt objects
}
\examples{
x <- segment(DataCPSim, method = "random", k = 10)
plot_history(x$segmenter)
x <- segment(DataCPSim, method = "pelt")
BMDL(x)
y <- segment(DataCPSim, method = "manual", cpts = 826)
BMDL(y)
z <- segment(DataCPSim, method = "single-best")
BMDL(z)
diagnose(segment(DataCPSim))
diagnose(segment(DataCPSim, method = "single-best"))
diagnose(segment(DataCPSim, method = "pelt"))
diagnose(segment(test_set()))
diagnose(segment(test_set(n = 2, sd = 4), method = "pelt"))

file_name(segment(DataCPSim, method = "pelt"))
}
\references{
Zhang and Seigmmund (2007) for MBIC: \doi{10.1111/j.1541-0420.2006.00662.x}
}
\seealso{
\code{\link[stats:logLik]{stats::logLik()}}

\code{\link[stats:nobs]{stats::nobs()}}

\code{\link[broom:reexports]{broom::augment()}}

\code{\link[broom:reexports]{broom::tidy()}}

\code{\link[broom:reexports]{broom::glance()}}
}
