% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mod_default.R, R/class-seg_default.R,
%   R/class-tidycpt.R, R/generics.R, R/pkg-GA.R, R/pkg-changepoint.R,
%   R/pkg-wbs.R
\name{changepoints.mod_default}
\alias{changepoints.mod_default}
\alias{changepoints.seg_default}
\alias{changepoints.tidycpt}
\alias{changepoints_labels}
\alias{changepoints}
\alias{params}
\alias{changepoints.ga}
\alias{changepoints.cpt}
\alias{changepoints.wbs}
\title{Extract changepoints}
\usage{
\method{changepoints}{mod_default}(x, ...)

\method{changepoints}{seg_default}(x, ...)

\method{changepoints}{tidycpt}(x, ...)

changepoints_labels(x, ...)

changepoints(x, ...)

params(x, ...)

\method{changepoints}{ga}(x, ...)

\method{changepoints}{cpt}(x, ...)

\method{changepoints}{wbs}(x, ...)
}
\arguments{
\item{x}{A \code{tidycpt} object}

\item{...}{arguments passed to methods}
}
\description{
Not to be confused with \code{\link[wbs:changepoints]{wbs::changepoints()}}, which returns different
information.
}
\examples{
cpts <- segment(DataCPSim, method = "ga", maxiter = 5)
changepoints(cpts$segmenter)

cpts <- segment(DataCPSim, method = "pelt")
changepoints(cpts)

cpts <- segment(DataCPSim, method = "wbs")
changepoints(cpts$segmenter)

}
\seealso{
\code{\link[wbs:changepoints]{wbs::changepoints()}}

Other tidychangepoint-generics: 
\code{\link{diagnose.seg_default}()},
\code{\link{exceedances}()},
\code{\link{fitness.seg_default}()}

Other tidychangepoint-generics: 
\code{\link{diagnose.seg_default}()},
\code{\link{exceedances}()},
\code{\link{fitness.seg_default}()}
}
\concept{tidychangepoint-generics}
