% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-tidycpt.R
\name{compare_models}
\alias{compare_models}
\alias{compare_algorithms}
\title{Compare various models for a give changepoint set}
\usage{
compare_models(x, ...)

compare_algorithms(x, ...)
}
\arguments{
\item{x}{A \code{tidycpt} object}

\item{...}{currently ignored}
}
\description{
Compare various models for a give changepoint set
}
