% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-changepoint.R
\name{cpt-generics}
\alias{cpt-generics}
\alias{as.segmenter.cpt}
\alias{as.ts.cpt}
\alias{logLik.cpt}
\alias{nobs.cpt}
\alias{seg_params.cpt}
\title{Compatibility layer for changepoint}
\usage{
\method{as.segmenter}{cpt}(object, ...)

\method{as.ts}{cpt}(x, ...)

\method{logLik}{cpt}(object, ...)

\method{nobs}{cpt}(object, ...)

\method{seg_params}{cpt}(x, ...)
}
\arguments{
\item{object}{A \code{cpt} object.}

\item{...}{arguments passed to methods}

\item{x}{A \code{cpt} object returned by \code{\link[changepoint:cpt.meanvar]{changepoint::cpt.meanvar()}}}
}
\description{
Methods for generic functions applied to \code{cpt} objects
}
\examples{
cpts <- segment(DataCPSim, method = "pelt")
x <- cpts$segmenter
class(x)
as.segmenter(x)
as.ts(x) 
changepoints(x)
fitness(x)
logLik(x)
model_name(x)
model_args(x)
nobs(x)
seg_params(x)

}
\seealso{
\link[changepoint:changepoint-package]{changepoint::changepoint-package}
}
