% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoint.R, R/seg_default.R
\name{cpt-generics}
\alias{cpt-generics}
\alias{glance.cpt}
\alias{params.cpt}
\alias{as.ts.cpt}
\alias{nobs.cpt}
\alias{logLik.cpt}
\alias{nobs.seg_default}
\alias{length.seg_default}
\title{Compatibility layer for changepoint}
\usage{
\method{glance}{cpt}(x, ...)

\method{params}{cpt}(x, ...)

\method{as.ts}{cpt}(x, ...)

\method{nobs}{cpt}(object, ...)

\method{logLik}{cpt}(object, ...)

\method{nobs}{seg_default}(object, ...)

\method{length}{seg_default}(x, ...)
}
\arguments{
\item{x}{A \code{cpt} object returned by \code{\link[changepoint:cpt.meanvar]{changepoint::cpt.meanvar()}}}

\item{...}{arguments passed to methods}

\item{object}{A \code{seg_default} object}
}
\description{
Compatibility layer for changepoint
}
\examples{
cpts <- segment(DataCPSim, method = "pelt")
class(cpts)
y <- augment(cpts)
class(y)
y
tidy(cpts)
glance(cpts)
cpts <- segment(DataCPSim, method = "pelt")
as.ts(cpts)

cpts <- segment(DataCPSim, method = "pelt")
nobs(cpts)

cpts <- segment(DataCPSim, method = "pelt", penalty = "BIC")
logLik(cpts)
cpts <- segment(DataCPSim, method = "pelt", penalty = "AIC")
logLik(cpts)

}
