% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-seg_default.R, R/pkg-changepoint.R
\name{best_cpt}
\alias{best_cpt}
\alias{cpt-generics}
\alias{glance.cpt}
\alias{params.cpt}
\alias{as.ts.cpt}
\alias{nobs.cpt}
\alias{logLik.cpt}
\title{Compatibility layer for changepoint}
\usage{
best_cpt(x, ...)

\method{glance}{cpt}(x, ...)

\method{params}{cpt}(x, ...)

\method{as.ts}{cpt}(x, ...)

\method{nobs}{cpt}(object, ...)

\method{logLik}{cpt}(object, ...)
}
\arguments{
\item{x}{A \code{cpt} object returned by \code{\link[changepoint:cpt.meanvar]{changepoint::cpt.meanvar()}}}

\item{...}{arguments passed to methods}

\item{object}{A \code{cpt} object.}
}
\description{
Compatibility layer for changepoint
}
\examples{
cpts <- segment(DataCPSim, method = "pelt")
class(cpts)
y <- augment(cpts)
class(y)
y
tidy(cpts)
glance(cpts)
cpts <- segment(DataCPSim, method = "pelt")
as.ts(cpts)

cpts <- segment(DataCPSim, method = "pelt")
nobs(cpts)

cpts <- segment(DataCPSim, method = "pelt", penalty = "BIC")
logLik(cpts$segmenter)
cpts <- segment(DataCPSim, method = "pelt", penalty = "AIC")
logLik(cpts$segmenter)

}
