% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-seg_default.R, R/class-tidycpt.R,
%   R/generics.R
\name{diagnose.seg_default}
\alias{diagnose.seg_default}
\alias{diagnose.tidycpt}
\alias{diagnose}
\title{Diagnose the fit of a segmented time series}
\usage{
\method{diagnose}{seg_default}(x, ...)

\method{diagnose}{tidycpt}(x, ...)

diagnose(x, ...)
}
\arguments{
\item{x}{A \code{tidycpt} object}

\item{...}{currently ignored}
}
\description{
Diagnose the fit of a segmented time series
}
\examples{
x <- segment(DataCPSim, method = "random", num_generations = 5)
diagnose(x)
diagnose(x$segmenter)
diagnose(segment(DataCPSim))
diagnose(segment(DataCPSim, method = "single-best"))
diagnose(segment(DataCPSim, method = "pelt"))
diagnose(segment(test_set()))
diagnose(segment(test_set(n = 2, sd = 4), method = "pelt"))

}
