% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mod_cpt.R, R/class-seg_basket.R,
%   R/class-tidycpt.R, R/generics.R
\name{diagnose.mod_cpt}
\alias{diagnose.mod_cpt}
\alias{diagnose.seg_basket}
\alias{diagnose.tidycpt}
\alias{diagnose}
\title{Diagnose the fit of a segmented time series}
\usage{
\method{diagnose}{mod_cpt}(x, ...)

\method{diagnose}{seg_basket}(x, ...)

\method{diagnose}{tidycpt}(x, ...)

diagnose(x, ...)
}
\arguments{
\item{x}{A \code{tidycpt} object}

\item{...}{currently ignored}
}
\description{
Diagnose the fit of a segmented time series
}
\examples{
diagnose(fit_meanshift_norm(CET, tau = 330))
\dontrun{
x <- segment(DataCPSim, method = "coen", num_generations = 3)
plot(x$segmenter)
diagnose(x$segmenter)
}
diagnose(segment(DataCPSim))
diagnose(segment(DataCPSim, method = "single-best"))
diagnose(segment(DataCPSim, method = "pelt"))
diagnose(segment(test_set()))
diagnose(segment(test_set(n = 2, sd = 4), method = "pelt"))
}
\seealso{
Other tidychangepoint-generics: 
\code{\link{changepoints.mod_cpt}()},
\code{\link{exceedances}()},
\code{\link{fitness.seg_basket}()}
}
\concept{tidychangepoint-generics}
