% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_nhpp.R, R/seg_default.R, R/stats.R,
%   R/tidycpt.R
\name{exceedances.nhpp}
\alias{exceedances.nhpp}
\alias{exceedances.seg_default}
\alias{exceedances.ts}
\alias{exceedances.double}
\alias{exceedances}
\alias{exceedances.tidycpt}
\title{Compute exceedances of a threshold for a time series}
\usage{
\method{exceedances}{nhpp}(x, ...)

\method{exceedances}{seg_default}(x, ...)

\method{exceedances}{ts}(x, ...)

\method{exceedances}{double}(x, threshold = mean(x, na.rm = TRUE), ...)

exceedances(x, ...)

\method{exceedances}{tidycpt}(x, ...)
}
\arguments{
\item{x}{a numeric vector coercible into a \link[stats:ts]{stats::ts} object}

\item{...}{arguments passed to methods}

\item{threshold}{A value above which to exceed. Default is the \code{\link[=mean]{mean()}}}
}
\description{
Compute exceedances of a threshold for a time series
}
