% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-tidycpt.R
\name{file_name}
\alias{file_name}
\title{Obtain a descriptive filename for a tidycpt object}
\usage{
file_name(x, data_name_slug = "data")
}
\arguments{
\item{x}{A \code{tidycpt} object}

\item{data_name_slug}{character string that will identify the data set used
in the file name}
}
\value{
A \code{character} string giving a unique file name.
}
\description{
Obtain a descriptive filename for a tidycpt object
}
\examples{
DataCPSim |>
  segment(method = "pelt") |>
  file_name()
}
