% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_lm.R
\name{fit_lm}
\alias{fit_lm}
\title{Regression-based model fitting}
\usage{
fit_lm(x, tau)
}
\description{
Regression-based model fitting
}
\examples{
tau <- c(365, 826)
mod <- fit_lm(DataCPSim, tau)
logLik(mod)
deg_free(mod)

}
