% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lmshift.R
\name{fit_lmshift}
\alias{fit_lmshift}
\alias{fit_lmshift_ar1}
\title{Regression-based model fitting}
\usage{
fit_lmshift(x, tau, trends = FALSE, ...)

fit_lmshift_ar1(x, tau, ...)
}
\arguments{
\item{x}{A time series}

\item{tau}{a set of indices representing a changepoint set}

\item{trends}{logical indicating whether you want trends within regions}

\item{...}{arguments passed to \code{\link[stats:lm]{stats::lm()}}}

\item{ar1}{logicial indicating whether you want auto-regressive lag 1 errors}
}
\description{
Regression-based model fitting
}
\examples{
tau <- c(365, 826)
mod <- fit_lmshift(DataCPSim, tau)
logLik(mod)
deg_free(mod)

cpts <- c(1700, 1739, 1988)
ids <- time2tau(cpts, as_year(time(CET)))
mod <- fit_lmshift(CET, tau = ids)
glance(mod)
glance(fit_lmshift(CET, tau = ids, trends = TRUE))
glance(fit_lmshift_ar1(CET, tau = ids))
glance(fit_lmshift_ar1(CET, tau = ids, trends = TRUE))
}
