% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lmshift.R
\name{fit_lmshift}
\alias{fit_lmshift}
\title{Regression-based model fitting}
\usage{
fit_lmshift(x, tau, trends = FALSE, ar1 = FALSE, ...)
}
\arguments{
\item{x}{A time series}

\item{tau}{a set of indices representing a changepoint set}

\item{trends}{logical indicating whether you want trends within regions}

\item{ar1}{logicial indicating whether you want auto-regressive lag 1 errors}

\item{...}{arguments passed to \code{\link[stats:lm]{stats::lm()}}}
}
\description{
Regression-based model fitting
}
\examples{
tau <- c(365, 826)
mod <- fit_lmshift(DataCPSim, tau)
logLik(mod)
deg_free(mod)

cpts <- c(1700, 1739, 1988)
ids <- time2tau(cpts, substr(time(CET), 1, 4))
mod <- fit_lmshift(CET, tau = ids)
glance(mod)
glance(fit_lmshift(CET, tau = ids, trends = TRUE))
glance(fit_lmshift(CET, tau = ids, ar1 = TRUE))
glance(fit_lmshift(CET, tau = ids, trends = TRUE, ar1 = TRUE))
}
