% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lmshift.R
\name{fit_lmshift}
\alias{fit_lmshift}
\alias{fit_lmshift_ar1}
\title{Regression-based model fitting}
\usage{
fit_lmshift(x, tau, deg_poly = 0, ...)

fit_lmshift_ar1(x, tau, ...)
}
\arguments{
\item{x}{A time series}

\item{tau}{a set of indices representing a changepoint set}

\item{deg_poly}{integer indicating the degree of the polynomial spline to be
fit. Passed to \code{\link[stats:poly]{stats::poly()}}.}

\item{...}{arguments passed to \code{\link[stats:lm]{stats::lm()}}}
}
\description{
Regression-based model fitting
}
\details{
\link{fit_lmshift_ar1}: will apply auto-regressive lag 1 errors
}
\examples{
tau <- c(365, 826)
mod <- fit_lmshift(DataCPSim, tau)
logLik(mod)
deg_free(mod)

cpts <- c(1700, 1739, 1988)
ids <- time2tau(cpts, as_year(time(CET)))
mod <- fit_lmshift(CET, tau = ids)
glance(mod)
glance(fit_lmshift(CET, tau = ids, deg_poly = 1))
glance(fit_lmshift_ar1(CET, tau = ids))
glance(fit_lmshift_ar1(CET, tau = ids, deg_poly = 1))
glance(fit_lmshift_ar1(CET, tau = ids, deg_poly = 2))
}
