% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cptshift.R
\name{fit_meanshift}
\alias{fit_meanshift}
\alias{logLik.cptshift}
\alias{glance.cptshift}
\title{Regression-based model fitting}
\usage{
fit_meanshift(x, tau, trends = FALSE, ar1 = FALSE, ...)

\method{logLik}{cptshift}(object, ...)

\method{glance}{cptshift}(x, ...)
}
\description{
Regression-based model fitting
}
\examples{
tau <- c(365, 826)
mod <- fit_meanshift(DataCPSim, tau)
logLik(mod)
deg_free(mod)

cpts <- c(1700, 1739, 1988)
ids <- time2tau(cpts, lubridate::year(time(CET)))
mod <- fit_meanshift(CET, tau = ids)
glance(mod)
glance(fit_meanshift(CET, tau = ids, trends = TRUE))
glance(fit_meanshift(CET, tau = ids, ar1 = TRUE))
glance(fit_meanshift(CET, tau = ids, trends = TRUE, ar1 = TRUE))
}
