% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cptshift.R
\name{fit_meanshift}
\alias{fit_meanshift}
\alias{logLik.cptshift}
\alias{MBIC.cptshift}
\alias{MDL.cptshift}
\alias{glance.cptshift}
\title{Regression-based model fitting}
\usage{
fit_meanshift(x, tau, trends = FALSE, ar1 = FALSE, ...)

\method{logLik}{cptshift}(object, ...)

\method{MBIC}{cptshift}(object, ...)

\method{MDL}{cptshift}(object, ...)

\method{glance}{cptshift}(x, ...)
}
\arguments{
\item{x}{A time series}

\item{tau}{a set of indices representing a changepoint set}

\item{trends}{logical indicating whether you want trends within regions}

\item{ar1}{logicial indicating whether you want auto-regressive lag 1 errors}

\item{...}{arguments passed to \code{\link[stats:lm]{stats::lm()}}}

\item{object}{any object from which a log-likelihood value, or a
   contribution to a log-likelihood value, can be extracted.}
}
\description{
Regression-based model fitting
}
\examples{
tau <- c(365, 826)
mod <- fit_meanshift(DataCPSim, tau)
logLik(mod)
deg_free(mod)

cpts <- c(1700, 1739, 1988)
ids <- time2tau(cpts, lubridate::year(time(CET)))
mod <- fit_meanshift(CET, tau = ids)
glance(mod)
glance(fit_meanshift(CET, tau = ids, trends = TRUE))
glance(fit_meanshift(CET, tau = ids, ar1 = TRUE))
glance(fit_meanshift(CET, tau = ids, trends = TRUE, ar1 = TRUE))
}
