% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_meanshift.R
\name{fit_meanshift}
\alias{fit_meanshift}
\alias{fit_meanshift_ar1}
\title{Fast implementation of meanshift model}
\usage{
fit_meanshift(x, tau, ar1 = FALSE, ...)

fit_meanshift_ar1(x, tau, ...)
}
\arguments{
\item{x}{A time series}

\item{tau}{a set of indices representing a changepoint set}

\item{ar1}{logicial indicating whether you want auto-regressive lag 1 errors}

\item{...}{arguments passed to \code{\link[stats:lm]{stats::lm()}}}
}
\description{
Fast implementation of meanshift model
}
\examples{
tau <- c(365, 826)
mod <- fit_meanshift_ar1(DataCPSim, tau)
logLik(mod)
deg_free(mod)

cpts <- c(1700, 1739, 1988)
ids <- time2tau(cpts, as_year(time(CET)))
mod <- fit_meanshift(CET, tau = ids)
glance(mod)
mod <- fit_meanshift_ar1(CET, tau = ids)
glance(mod)
}
\author{
Xueheng Shi
}
