% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptshift.R
\name{fit_meanshift}
\alias{fit_meanshift}
\alias{fit_meanshift_ar1}
\alias{fit_trendshift}
\title{Regression-based model fitting}
\usage{
fit_meanshift(x, tau, ...)

fit_meanshift_ar1(x, tau, ...)

fit_trendshift(x, tau, trends = TRUE, ...)
}
\description{
Regression-based model fitting
}
\examples{
tau <- c(365, 826)
mod <- fit_meanshift(DataCPSim, tau)
logLik(mod)
deg_free(mod)

cpts <- c(1700, 1739, 1988)
ids <- time2tau(cpts, lubridate::year(time(CET)))
mod <- fit_meanshift(CET, tau = ids)
glance(mod)

mod2 <- fit_meanshift_ar1(CET, tau = ids)
glance(mod2)

mod <- fit_trendshift(CET, tau = ids, trends = TRUE)
logLik(mod)
deg_free(mod)
}
