% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_meanvar.R
\name{fit_meanvar}
\alias{fit_meanvar}
\title{Fit a model for mean and variance}
\usage{
fit_meanvar(x, tau, ...)
}
\description{
Fit a model for mean and variance
}
\examples{
fit_meanvar(CET, tau = c(42, 330))
}
