% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_meanvar.R
\name{fit_meanvar}
\alias{fit_meanvar}
\title{Fit a model for mean and variance}
\usage{
fit_meanvar(x, tau, ...)
}
\arguments{
\item{x}{A time series}

\item{tau}{a set of indices representing a changepoint set}

\item{...}{arguments passed to \code{\link[stats:lm]{stats::lm()}}}
}
\description{
Fit a model for mean and variance
}
\examples{
fit_meanvar(CET, tau = c(42, 330))
}
\seealso{
Other model-fitting: 
\code{\link{fit_lmshift}()},
\code{\link{fit_meanshift}()},
\code{\link{fit_nhpp_region}()}
}
\concept{model-fitting}
