% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhpp.R
\name{fit_nhpp_region}
\alias{fit_nhpp_region}
\alias{fit_nhpp}
\alias{logLik.nhpp}
\alias{BMDL.nhpp}
\alias{MBIC.nhpp}
\alias{changepoints.nhpp}
\alias{exceedances.nhpp}
\title{Fit a non-homogeneous Poisson process model to the exceedances of a time series.}
\usage{
fit_nhpp_region(exc, tau_left, tau_right, params = parameters_weibull(), ...)

fit_nhpp(x, tau)

\method{logLik}{nhpp}(object, ...)

\method{BMDL}{nhpp}(x, ...)

\method{MBIC}{nhpp}(object, ...)

\method{changepoints}{nhpp}(x, ...)

\method{exceedances}{nhpp}(x, ...)
}
\arguments{
\item{x}{A time series}

\item{tau}{A vector of changepoints.}
}
\value{
A \code{tbl_df} with each row representing one region.
}
\description{
Any times series can be modeled as a non-homogeneous Poisson process of the
locations of the exceedances of the mean of the series. This function
uses the \code{\link[=BMDL]{BMDL()}} criteria to determine the best fit parameters for each
region defined by the changepoint set returned by \code{\link[=changepoints]{changepoints()}}.
}
\examples{
fit_nhpp_region(exceedances(lista_AG), 0, 575)
fit_nhpp_region(exceedances(lista_AG), 0, 575, initial_val_optim = c(1, 10))
fit_nhpp(DataCPSim, tau = 826)
fit_nhpp(as.ts(lista_AG), tau = changepoints(lista_AG))
}
\references{
Zhang and Seigmmund (2007) for MBIC: \doi{10.1111/j.1541-0420.2006.00662.x}
}
