% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_nhpp.R
\name{fit_nhpp_region}
\alias{fit_nhpp_region}
\alias{fit_nhpp}
\alias{logLik.nhpp}
\alias{glance.nhpp}
\alias{fitted.nhpp}
\alias{mcdf}
\alias{plot.nhpp}
\alias{plot_intensity}
\title{Fit a non-homogeneous Poisson process model to the exceedances of a time series.}
\usage{
fit_nhpp_region(exc, tau_left, tau_right, params = parameters_weibull(), ...)

fit_nhpp(x, tau, ...)

\method{logLik}{nhpp}(object, ...)

\method{glance}{nhpp}(x, ...)

\method{fitted}{nhpp}(object, ...)

mcdf(x, dist = "weibull")

\method{plot}{nhpp}(x, ...)

plot_intensity(x, ...)
}
\arguments{
\item{exc}{Output from \code{\link[=exceedances]{exceedances()}}}

\item{tau_left}{left-most changepoint}

\item{tau_right}{right-most changepoint}

\item{params}{Output from \code{\link[=parameters_weibull]{parameters_weibull()}}}

\item{...}{arguments passed to \code{\link[stats:optim]{stats::optim()}}}

\item{x}{An \code{nhpp} object}

\item{tau}{A vector of changepoints.}

\item{object}{An \code{nhpp} object}

\item{dist}{Name of the distribution}
}
\value{
A \code{tbl_df} with each row representing one region.

a numeric vector of length equal to the \link{exceedances} of \code{x}
}
\description{
Any times series can be modeled as a non-homogeneous Poisson process of the
locations of the exceedances of the mean of the series. This function
uses the \code{\link[=BMDL]{BMDL()}} criteria to determine the best fit parameters for each
region defined by the changepoint set returned by \code{\link[=changepoints]{changepoints()}}.
}
\examples{
fit_nhpp_region(exceedances(DataCPSim), 0, 575)
fit_nhpp(DataCPSim, tau = 826)
fit_nhpp(DataCPSim, tau = 826, threshold = 20)
fit_nhpp(DataCPSim, tau = 826, threshold = 200)
fit_nhpp(DataCPSim, tau = changepoints(segment(DataCPSim, method = "pelt")))
nhpp <- fit_nhpp(DataCPSim, tau = 826)
mcdf(nhpp)

plot(fit_nhpp(DataCPSim, tau = 826))
plot_intensity(fit_nhpp(DataCPSim, tau = 826))
mod <- segment(bogota_pm, method = "pelt")
plot_intensity(mod$nhpp)
}
