% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhpp.R
\name{fit_nhpp_region}
\alias{fit_nhpp_region}
\title{Estimadores MAP}
\usage{
fit_nhpp_region(
  t,
  tau_left,
  tau_right,
  initial_val_optim = tidychangepoint::param$initial_val_optim,
  mat_low_upp = tidychangepoint::param$mat_low_upp,
  rf_type = tidychangepoint::param$rf_type,
  vec_dist_a_priori = tidychangepoint::param$vec_dist_a_priori,
  mat_phi = tidychangepoint::param$mat_phi,
  ...
)
}
\arguments{
\item{initial_val_optim}{valores iniciales que utiliza la función optim para
encontrar el mínimo}

\item{mat_low_upp}{matriz con lugares donde buscar; cada renglon es para un
parámetro del NHPP}

\item{rf_type}{nombre de tasa de NHPP}

\item{vec_dist_a_priori}{nombres de distribuciones a priori}

\item{mat_phi}{matriz cuyos renglones tiene los parámetros de las
distribuciones a priori}

\item{vec_d_i}{vector de días de un régimen}

\item{tau1}{valor del primer punto de cambio}

\item{tau2}{valor del segundo punto de cambio}
}
\value{
regresa un el resultado de optim
}
\description{
Estimadores MAP
}
\examples{
fit_nhpp_region(exceedances(lista_AG$data), 0, 575)
fit_nhpp_region(exceedances(lista_AG$data), 0, 575, initial_val_optim = c(1, 10))


}
