% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_nhpp.R
\name{fit_nhpp_region}
\alias{fit_nhpp_region}
\title{Fit an NHPP model to one specific region}
\usage{
fit_nhpp_region(exc, tau_left, tau_right, params = parameters_weibull(), ...)
}
\arguments{
\item{exc}{Output from \code{\link[=exceedances]{exceedances()}}}

\item{tau_left}{left-most changepoint}

\item{tau_right}{right-most changepoint}

\item{params}{Output from \code{\link[=parameters_weibull]{parameters_weibull()}}}

\item{...}{arguments passed to \code{\link[stats:optim]{stats::optim()}}}
}
\value{
Modified output from \code{\link[stats:optim]{stats::optim()}}.
}
\description{
Fit an NHPP model to one specific region
}
\details{
This is an internal function not to be called by users.
Use \code{\link[=fit_nhpp]{fit_nhpp()}}.
}
\seealso{
\code{\link[=fit_nhpp]{fit_nhpp()}}
}
\keyword{internal}
