% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA.R, R/changepoint.R, R/generics.R,
%   R/seg_default.R, R/tidycpt.R, R/wbs.R
\name{fitness.ga}
\alias{fitness.ga}
\alias{fitness.cpt}
\alias{fitness}
\alias{fitness.seg_default}
\alias{fitness.tidycpt}
\alias{fitness.wbs}
\title{Retrieve the optimal fitness (or objective function) value used by an algorithm}
\usage{
\method{fitness}{ga}(object, ...)

\method{fitness}{cpt}(object, ...)

fitness(object, ...)

\method{fitness}{seg_default}(object, ...)

\method{fitness}{tidycpt}(object, ...)

\method{fitness}{wbs}(object, ...)
}
\arguments{
\item{object}{An object, typically returned by \code{\link[=segment]{segment()}}}

\item{...}{currently ignored}
}
\description{
Retrieve the optimal fitness (or objective function) value used by an algorithm
}
\examples{
x <- segment(DataCPSim, method = "ga")
fitness(x)

x <- segment(DataCPSim, method = "pelt")
fitness(x)

x <- segment(DataCPSim, method = "wbs")
fitness(x)

}
