% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-seg_default.R, R/class-tidycpt.R,
%   R/generics.R, R/pkg-GA.R, R/pkg-changepoint.R, R/pkg-wbs.R
\name{fitness.seg_default}
\alias{fitness.seg_default}
\alias{fitness.tidycpt}
\alias{fitness}
\alias{fitness.ga}
\alias{fitness.cpt}
\alias{fitness.wbs}
\title{Retrieve the optimal fitness (or objective function) value used by an algorithm}
\usage{
\method{fitness}{seg_default}(object, ...)

\method{fitness}{tidycpt}(object, ...)

fitness(object, ...)

\method{fitness}{ga}(object, ...)

\method{fitness}{cpt}(object, ...)

\method{fitness}{wbs}(object, ...)
}
\arguments{
\item{object}{A segmenter object, typically returned by \code{\link[=segment]{segment()}}}

\item{...}{currently ignored}
}
\description{
Retrieve the optimal fitness (or objective function) value used by an algorithm
}
\examples{
x <- segment(DataCPSim, method = "ga", maxiter = 10)
fitness(x)

x <- segment(DataCPSim, method = "pelt")
fitness(x)

x <- segment(DataCPSim, method = "wbs")
fitness(x)

}
\seealso{
Other tidychangepoint-generics: 
\code{\link{changepoints.mod_cpt}()},
\code{\link{diagnose.seg_default}()},
\code{\link{exceedances}()}
}
\concept{tidychangepoint-generics}
