% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-GA.R
\name{ga-generics}
\alias{ga-generics}
\alias{as.seg_cpt.ga}
\alias{nobs.ga}
\title{Compatibility layer for GA}
\usage{
\method{as.seg_cpt}{ga}(object, ...)

\method{nobs}{ga}(object, ...)
}
\arguments{
\item{object}{A \code{ga} object.}

\item{...}{arguments passed to methods}

\item{x}{A \code{GA} object returned by \code{\link[GA:ga]{GA::ga()}}}
}
\description{
Methods for generic functions applied to \code{ga} objects
}
\examples{
cpts <- segment(DataCPSim, method = "ga", maxiter = 10)
x <- as.segmenter(cpts)
class(x)
as.seg_cpt(x)
as.ts(x) 
changepoints(x)
fitness(x)
model_name(x)
model_args(x)
nobs(x)
seg_params(x)
cpts <- segment(DataCPSim, method = "ga", maxiter = 5)
nobs(cpts$segmenter)
}
\seealso{
\link[GA:GA-package]{GA::GA-package}
}
