% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoint.R
\name{glance.cpt}
\alias{glance.cpt}
\alias{as.ts.cpt}
\alias{nobs.cpt}
\alias{logLik.cpt}
\alias{changepoints.cpt}
\title{Compatibility layer for changepoint}
\usage{
\method{glance}{cpt}(x, ...)

\method{as.ts}{cpt}(x, ...)

\method{nobs}{cpt}(object, ...)

\method{logLik}{cpt}(object, ...)

\method{changepoints}{cpt}(x, ...)
}
\description{
Compatibility layer for changepoint
}
\examples{
cpts <- segment(DataCPSim, method = "cpt-pelt")
class(cpts)
y <- augment(cpts)
class(y)
y
tidy(cpts)
glance(cpts)
cpts <- segment(DataCPSim, method = "cpt-pelt")
as.ts(cpts)

cpts <- segment(DataCPSim, method = "cpt-pelt")
nobs(cpts)

cpts <- segment(DataCPSim, method = "cpt-pelt", penalty = "BIC")
logLik(cpts)
cpts <- segment(DataCPSim, method = "cpt-pelt", penalty = "AIC")
logLik(cpts)

cpts <- segment(DataCPSim, method = "cpt-pelt")
changepoints(cpts)

}
