% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpt_gbmdl.R
\name{glance.cpt_gbmdl}
\alias{glance.cpt_gbmdl}
\title{Broom compatibility layer for changepoint}
\usage{
\method{glance}{cpt_gbmdl}(x, ...)
}
\arguments{
\item{x}{A \code{cpt_gbmdl} object}

\item{...}{arguments passed to methods}
}
\description{
Broom compatibility layer for changepoint
}
\examples{
cpts <- lista_AG
y <- augment(lista_AG)
class(y)
y
tidy(cpts)
glance(cpts)
}
