% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GA.R
\name{glance.ga}
\alias{glance.ga}
\title{Compatibility layer for GA}
\usage{
\method{glance}{ga}(x, ...)
}
\arguments{
\item{x}{A \code{GA} object returned by \code{\link[GA:ga]{GA::ga()}}}

\item{...}{arguments passed to methods}
}
\description{
Compatibility layer for GA
}
\examples{
cpts <- segment(DataCPSim, method = "ga", maxiter = 20)
class(cpts)
y <- augment(cpts)
class(y)
y
tidy(cpts)
glance(cpts)
}
