% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbs.R
\name{glance.wbs}
\alias{glance.wbs}
\title{Compatibility layer for wbs}
\usage{
\method{glance}{wbs}(x, ...)
}
\arguments{
\item{x}{A \code{wbs} object returned by \code{\link[wbs:wbs]{wbs::wbs()}}}

\item{...}{arguments passed to methods}
}
\description{
Compatibility layer for wbs
}
\examples{
cpts <- segment(DataCPSim, method = "wbs")
class(cpts)
y <- augment(cpts)
class(y)
y
tidy(cpts)
glance(cpts)
}
