% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-FnAGBayesianMDL.R
\name{grafica_ajuste_NHPP}
\alias{grafica_ajuste_NHPP}
\title{Graficando un ajuste de NHPP}
\usage{
grafica_ajuste_NHPP(
  d_i,
  tau1,
  tau2,
  initial_val_optim,
  mat_low_upp,
  rf_type,
  vec_dist_a_priori,
  mat_phi
)
}
\arguments{
\item{d_i}{vector de días de revases}

\item{tau1}{}

\item{tau2}{}

\item{mat_low_upp}{}

\item{mat_phi}{}
}
\description{
Esta función solo grafica un ajuste de NHPP
}
