% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{grafica_datos_e_intervalos}
\alias{grafica_datos_e_intervalos}
\alias{plot_BMDL}
\title{Graficas de los intervalos y los datos}
\usage{
grafica_datos_e_intervalos(
  lista_AG,
  param,
  n_puntos_cambio,
  d,
  alpha,
  sigma,
  tau
)

plot_BMDL(cpt_list, destdir = tempdir())
}
\arguments{
\item{lista_AG}{description}

\item{param}{description}
}
\value{
una gráfica con :
1 Datos reales
2 Evolución del algoritmo genético
3 Puntos de cambio que más se repitieron
4 Number of change points in the best chromosomes
}
\description{
Graficas de los intervalos y los datos

Graficas del AG con BMDL
}
\examples{
x <- sample(1:1096)
alpha <- runif(6) 
sigma <- runif(6)
tau <- sample(x, 5)
grafica_datos_e_intervalos(
  lista_AG = list(x = x), param,
  n_puntos_cambio = 5, d = x, 
  alpha = alpha, sigma = sigma, tau = tau
)

\dontrun{
lista_AG <- AG_BMDL_r_paso(DataCPSimRebases, param)
plot_BMDL(lista_AG)
}

}
