% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpt_gbmdl_graphics.R
\name{plot_BMDL}
\alias{plot_BMDL}
\alias{plot_evolution}
\alias{plot_cpt_repeated}
\alias{plot_best_chromosome}
\title{Graficas del AG con BMDL}
\usage{
plot_BMDL(cpt_list, destdir = tempdir(), data_name_slug = "data", pdf = FALSE)

plot_evolution(cpt_list, i = length(cpt_list$vec_min_BMDL))

plot_cpt_repeated(cpt_list, i = nrow(cpt_list$historia_mejores))

plot_best_chromosome(cpt_list)
}
\value{
una gráfica con :
1 Datos reales
2 Evolución del algoritmo genético
3 Puntos de cambio que más se repitieron
4 Number of change points in the best chromosomes
}
\description{
Graficas del AG con BMDL
}
\examples{
plot_BMDL(lista_AG$segmenter)

plot_evolution(lista_AG$segmenter)
plot_evolution(lista_AG$segmenter, 5)

plot_cpt_repeated(lista_AG$segmenter)
plot_cpt_repeated(lista_AG$segmenter, 5)
plot_best_chromosome(lista_AG$segmenter)
}
