% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{italy_grads}
\alias{italy_grads}
\title{Italian University graduates by disciplinary groups from 1926-2013}
\format{
An object of class \code{tbl_ts} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 88 rows and 17 columns.
}
\source{
\url{https://seriestoriche.istat.it/}

Source: Istat- Ministero dell'istruzione pubblica, years 1926-1942

Istat- Rilevazione sulle Università, years 1943-1997

Miur- Rilevazione sulle Università, years 1998-2013
}
\usage{
italy_grads
}
\description{
Italian University graduates by disciplinary groups during the years 1926-2013.
}
\keyword{datasets}
