% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_weibull.R
\name{iweibull}
\alias{iweibull}
\alias{mweibull}
\alias{parameters_weibull}
\alias{log_likelihood_region_weibull}
\alias{log_prior_region_weibull}
\alias{D_log_prior_region_weibull}
\alias{D_log_likelihood_region_weibull}
\title{Weibull distribution functions}
\usage{
iweibull(x, shape, scale = 1)

mweibull(x, shape, scale = 1)

parameters_weibull(...)

log_likelihood_region_weibull(t, tau_left, tau_right, theta)

log_prior_region_weibull(theta, params = parameters_weibull())

D_log_prior_region_weibull(theta, params = parameters_weibull())

D_log_likelihood_region_weibull(t, tau_left, tau_right, theta)
}
\arguments{
\item{x}{A numeric vector}

\item{shape}{Shape parameter for Weibull distribution. See \code{\link[stats:Weibull]{stats::dweibull()}}.}

\item{scale}{Scale parameter for Weibull distribution. See \code{\link[stats:Weibull]{stats::dweibull()}}.}

\item{...}{currently ignored}

\item{t}{vector of exceedances}

\item{tau_left}{valor del primer punto de cambio}

\item{tau_right}{valor del segundo punto de cambio}

\item{theta}{vector de parámetros de verosimilitud del NHPP}

\item{params}{Possibly modified output from \code{\link[=parameters_weibull]{parameters_weibull()}}}
}
\value{
regresa la evaluación

regresa la evaluación
}
\description{
Weibull distribution functions
}
\examples{
iweibull(1, shape = 1, scale = 1)
plot(x = 1:10, y = iweibull(1:10, shape = 2, scale = 2))
mweibull(1, shape = 1, scale = 1)
plot(x = 1:10, y = mweibull(1:10, shape = 1, scale = 1))
plot(x = 1:10, y = mweibull(1:10, shape = 1, scale = 2))
plot(x = 1:10, y = mweibull(1:10, shape = 0.5, scale = 2))
plot(x = 1:10, y = mweibull(1:10, shape = 0.5, scale = 100))
plot(x = 1:10, y = mweibull(1:10, shape = 2, scale = 2))
plot(x = 1:10, y = mweibull(1:10, shape = 2, scale = 100))

log_likelihood_region_weibull(DataCPSim, 0, 575, theta = c(0.5, 2))

log_prior_region_weibull(theta = c(0.5, 2))
D_log_prior_region_weibull(theta = c(0.5, 2))

D_log_likelihood_region_weibull(DataCPSim, 0, 575, theta = c(0.5, 2))
}
\seealso{
\code{\link[stats:Weibull]{stats::dweibull()}}
}
