% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_weibull.R
\name{log_likelihood_region_weibull}
\alias{log_likelihood_region_weibull}
\alias{log_prior_region_weibull}
\alias{D_log_prior_region_weibull}
\alias{D_log_likelihood_region_weibull}
\title{Log-Likelihood functions for regions (Weibull)}
\usage{
log_likelihood_region_weibull(t, tau_left, tau_right, theta)

log_prior_region_weibull(theta, params = parameters_weibull())

D_log_prior_region_weibull(theta, params = parameters_weibull())

D_log_likelihood_region_weibull(t, tau_left, tau_right, theta)
}
\arguments{
\item{t}{vector of \code{\link[=exceedances]{exceedances()}}}

\item{tau_left}{Left endpoint of the region}

\item{tau_right}{Right endpoint of the region}

\item{theta}{numeric vector of parameters for the NHPP model}

\item{params}{Possibly modified output from \code{\link[=parameters_weibull]{parameters_weibull()}}}
}
\value{
A numeric vector
}
\description{
Log-Likelihood functions for regions (Weibull)
}
\keyword{internal}
