% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{logLik.ts}
\alias{logLik.ts}
\alias{exceedances.ts}
\alias{exceedances.double}
\alias{MDL.logLik}
\alias{MBIC.logLik}
\title{Generic methods for classes from \link{stats}}
\usage{
\method{logLik}{ts}(object, ...)

\method{exceedances}{ts}(x, ...)

\method{exceedances}{double}(x, ...)

\method{MDL}{logLik}(x, ...)

\method{MBIC}{logLik}(x, ...)
}
\description{
Generic methods for classes from \link{stats}
}
\details{
loc.ind is the binary input of length N
Note \code{X[1]} cannot be a changepoint,
so \code{loc.ind[1]} is set to be zero
loc.ind = 0/1. 1 corresponds a CPT location, 0 is not.
}
\examples{
x <- england_temperature$annual_mean_temp
logLik(as.ts(x), loc.ind = round(runif(length(x))))
x <- england_temperature$annual_mean_temp
MDL(as.ts(x), tau = 123)
}
\references{
Zhang and Seigmmund (2007) for MBIC: \doi{10.1111/j.1541-0420.2006.00662.x}
}
\author{
Xueheng Shi
}
