% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{exceedances.ts}
\alias{exceedances.ts}
\alias{exceedances.double}
\alias{MDL.logLik}
\alias{MBIC.logLik}
\title{Generic methods for classes from \link{stats}}
\usage{
\method{exceedances}{ts}(x, ...)

\method{exceedances}{double}(x, ...)

\method{MDL}{logLik}(object, ...)

\method{MBIC}{logLik}(object, ...)
}
\arguments{
\item{x}{A numeric vector}
}
\description{
Generic methods for classes from \link{stats}
}
\examples{
MDL(logLik(as.ts(CET), loc.ind = tau2binary(123, n = length(CET))))
}
\references{
Zhang and Seigmmund (2007) for MBIC: \doi{10.1111/j.1541-0420.2006.00662.x}
}
