% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-MaMBOltzmann.R, R/3-GABoltzmannRanks.R
\name{MaMBOltzmann}
\alias{MaMBOltzmann}
\alias{GABoltzmannMutation2}
\title{Boltzman}
\usage{
MaMBOltzmann(mat_cp, Fitness, Mutation, Temperature)

GABoltzmannMutation2(param, Mutation = 0.03, Temperature = 27)
}
\arguments{
\item{mat_cp}{description}

\item{Fitness}{description}

\item{Mutation}{description}

\item{Temperature}{description}

\item{param}{description}
}
\description{
Boltzman
}
