% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boltzmann.R
\name{GABoltzmannMutation2}
\alias{GABoltzmannMutation2}
\alias{MaMBOltzmann}
\title{Boltzman}
\usage{
GABoltzmannMutation2(x, r = 50, k = 50, Mutation = 0.03, Temperature = 27)

MaMBOltzmann(mat_cp, Fitness, Mutation, Temperature)
}
\arguments{
\item{x}{An \code{integer} vector. See \code{\link[=exceedances]{exceedances()}}.}

\item{Mutation}{description}

\item{Temperature}{description}

\item{mat_cp}{description}

\item{Fitness}{description}
}
\value{
A \code{list} of length 2:
\itemize{
\item A \code{numeric} vector of \code{FitnessGeneraciones}
\item A \code{matrix} with \code{k} rows and \code{N} columns
}
}
\description{
What does this function do?
}
\examples{
\dontrun{
  GABoltzmannMutation2(exceedances(DataCPSim))
  GABoltzmannMutation2(rlnorm_ts_1)
  
}
}
