% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoint.R, R/mod_lmshift.R,
%   R/mod_meanshift.R, R/mod_nhpp.R, R/stats.R, R/tidycpt.R
\name{MBIC.cpt}
\alias{MBIC.cpt}
\alias{MBIC.lmshift}
\alias{MBIC.meanshift}
\alias{MBIC.nhpp}
\alias{MBIC.logLik}
\alias{MBIC}
\alias{MBIC.tidycpt}
\title{Modified Bayesian Information Criterion}
\usage{
\method{MBIC}{cpt}(object, ...)

\method{MBIC}{lmshift}(object, ...)

\method{MBIC}{meanshift}(object, ...)

\method{MBIC}{nhpp}(object, ...)

\method{MBIC}{logLik}(object, ...)

MBIC(object, ...)

\method{MBIC}{tidycpt}(object, ...)
}
\arguments{
\item{object}{any object from which a log-likelihood value, or a
   contribution to a log-likelihood value, can be extracted.}

\item{...}{some methods for this generic function require additional
   arguments.}
}
\description{
Modified Bayesian Information Criterion
}
\references{
Zhang and Seigmmund (2007) for MBIC: \doi{10.1111/j.1541-0420.2006.00662.x}

Zhang and Seigmmund (2007) for MBIC: \doi{10.1111/j.1541-0420.2006.00662.x}

Zhang and Seigmmund (2007) for MBIC: \doi{10.1111/j.1541-0420.2006.00662.x}
}
\seealso{
\code{\link[stats:AIC]{stats::BIC()}}
}
