% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoint.R, R/mod_lmshift.R,
%   R/mod_meanshift.R, R/mod_nhpp.R, R/stats.R, R/tidycpt.R
\name{MDL.cpt}
\alias{MDL.cpt}
\alias{MDL.lmshift}
\alias{MDL.meanshift}
\alias{MDL.nhpp}
\alias{MDL.logLik}
\alias{MDL}
\title{Maximum Descriptive Length}
\usage{
\method{MDL}{cpt}(object, ...)

\method{MDL}{lmshift}(object, ...)

\method{MDL}{meanshift}(object, ...)

\method{MDL}{nhpp}(object, ...)

\method{MDL}{logLik}(object, ...)

MDL(object, ...)
}
\arguments{
\item{object}{any object from which a log-likelihood value, or a
   contribution to a log-likelihood value, can be extracted.}

\item{...}{some methods for this generic function require additional
   arguments.}
}
\description{
Maximum Descriptive Length
}
\examples{
MDL(fit_meanshift_ar1(CET, tau = c(42, 330)))
}
