% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/mod_nhpp.R, R/stats.R
\name{MDL}
\alias{MDL}
\alias{MDL.default}
\alias{MDL.nhpp}
\alias{MDL.logLik}
\title{Maximum Descriptive Length}
\usage{
MDL(object, ...)

\method{MDL}{default}(object, ...)

\method{MDL}{nhpp}(object, ...)

\method{MDL}{logLik}(object, ...)
}
\arguments{
\item{object}{any object from which a log-likelihood value, or a
   contribution to a log-likelihood value, can be extracted.}

\item{...}{some methods for this generic function require additional
   arguments.}
}
\value{
regresa la evaluación de la penalización
\deqn{
   2 \cdot P(\tau | N) = a \cdot \ln{N}
     + b \cdot \sum_{i=1}^{m+1} \ln{(\tau_i-\tau_{i-1})} 
     + \ln{m} + \sum_{i=2}^m \ln{(\tau_i)}
 }
where \eqn{a} is the number of parameters estimated using all of the data points,
and \eqn{b} is the number of parameters estimated in each regime.
These quantites should be \code{\link[base:attributes]{base::attributes()}} of the object returned by
\code{\link[=logLik]{logLik()}}.
}
\description{
Maximum Descriptive Length
}
\examples{
MDL(fit_meanshift_ar1(CET, tau = c(42, 330)))
MDL(fit_lmshift(CET, tau = c(42, 81, 330), trends = TRUE))
}
