% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_cumulative.R
\name{media_acumulada}
\alias{media_acumulada}
\title{Mean cumulative exceedances function}
\usage{
media_acumulada(x, tau, theta, dist = "weibull")
}
\value{
a numeric vector of length equal to the \link{exceedances} of \code{x}
}
\description{
Mean cumulative exceedances function
}
\examples{
tau <- cpt_best(lista_AG)
theta <- cpt_best_params(lista_AG)
media_acumulada(x = lista_AG$data, tau, theta)

}
