% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_cumulative.R
\name{media_acumulada}
\alias{media_acumulada}
\title{Mean cumulative exceedances function}
\usage{
media_acumulada(t, tau, theta, n, dist = "weibull")
}
\value{
a numeric vector of length equal to the \link{exceedances} of \code{x}
}
\description{
Mean cumulative exceedances function
}
\examples{
tau <- changepoints(lista_AG)
theta <- cpt_best_params(lista_AG$segmenter)
media_acumulada(exceedances(as.ts(lista_AG)), tau, theta, length(as.ts(lista_AG)))

}
