% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mlb_hrs}
\alias{mlb_hrs}
\title{Differences between leagues in Major League Baseball}
\format{
An object of class \code{xts} (inherits from \code{zoo}) with 98 rows and 1 columns.
}
\usage{
mlb_hrs
}
\description{
The difference in home runs hit per plate appearance between the
American League and the National League from 1925 to 2022.
}
\keyword{datasets}
