% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/pkg-GA.R, R/pkg-changepoint.R
\name{model_args}
\alias{model_args}
\alias{model_args.ga}
\alias{model_args.cpt}
\title{Retrieve the model arguments that a segmenter used}
\usage{
model_args(object, ...)

\method{model_args}{ga}(object, ...)

\method{model_args}{cpt}(object, ...)
}
\arguments{
\item{object}{An object, typically returned by \code{\link[=segment]{segment()}}}

\item{...}{currently ignored}
}
\description{
Retrieve the model arguments that a segmenter used
}
