% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/pkg-GA.R, R/pkg-changepoint.R
\name{model_args}
\alias{model_args}
\alias{model_args.default}
\alias{model_args.ga}
\alias{model_args.cpt}
\title{Retrieve the arguments that a model-fitting function used}
\usage{
model_args(object, ...)

\method{model_args}{default}(object, ...)

\method{model_args}{ga}(object, ...)

\method{model_args}{cpt}(object, ...)
}
\arguments{
\item{object}{An object, typically returned by \code{\link[=segment]{segment()}}}

\item{...}{currently ignored}
}
\value{
\itemize{
\item \code{\link[=model_args]{model_args()}}: A named \code{list} of arguments, or \code{NULL}
}
}
\description{
Retrieve the arguments that a model-fitting function used
}
\details{
Every segmenter uses a model-fitting function, and these functions sometimes
take arguments. \code{\link[=model_args]{model_args()}} recovers the arguments that were passed to
the model fitting function when it was called. These are especially
important when using a genetic algorithm.
}
\examples{
x <- segment(CET, method = "ga-taimal", maxiter = 3)
model_args(x$segmenter)
}
\seealso{
Other modeling: 
\code{\link{model_name.mod_default}()}
}
\concept{modeling}
