% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-seg_cpt.R, R/generics.R, R/pkg-GA.R,
%   R/pkg-changepoint.R, R/pkg-wbs.R
\name{model_args.seg_cpt}
\alias{model_args.seg_cpt}
\alias{model_args}
\alias{model_args.default}
\alias{model_args.ga}
\alias{model_args.cpt}
\alias{model_args.wbs}
\title{Retrieve the arguments that a model-fitting function used}
\usage{
\method{model_args}{seg_cpt}(object, ...)

model_args(object, ...)

\method{model_args}{default}(object, ...)

\method{model_args}{ga}(object, ...)

\method{model_args}{cpt}(object, ...)

\method{model_args}{wbs}(object, ...)
}
\arguments{
\item{object}{A \code{segmenter} object.}

\item{...}{currently ignored}
}
\value{
A named \code{list} of arguments, or \code{NULL}
}
\description{
Retrieve the arguments that a model-fitting function used
}
\details{
Every model is fit by a model-fitting function, and these functions sometimes
take arguments.
\code{\link[=model_args]{model_args()}} recovers the arguments that were passed to
the model fitting function when it was called.
These are especially
important when using a genetic algorithm.
}
\examples{
# Segment a time series using Coen's algorithm
x <- segment(CET, method = "ga-coen", maxiter = 3)

# Recover the arguments passed to the model-fitting function
x |>
  as.segmenter() |>
  model_args()
  
}
\seealso{
Other modeling: 
\code{\link{model_name.mod_cpt}()},
\code{\link{new_fun_cpt}()}
}
\concept{modeling}
