% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{model_name}
\alias{model_name}
\title{Retrieve the model that a segmenter used}
\usage{
model_name(x, ...)
}
\arguments{
\item{...}{currently ignored}
}
\description{
Retrieve the model that a segmenter used
}
