% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mod_cpt.R, R/class-seg_basket.R,
%   R/class-seg_cpt.R, R/class-tidycpt.R, R/generics.R, R/pkg-GA.R,
%   R/pkg-changepoint.R, R/pkg-wbs.R
\name{model_name.mod_cpt}
\alias{model_name.mod_cpt}
\alias{model_name.seg_basket}
\alias{model_name.seg_cpt}
\alias{model_name.tidycpt}
\alias{model_name}
\alias{model_name.default}
\alias{model_name.character}
\alias{model_name.ga}
\alias{model_name.cpt}
\alias{model_name.wbs}
\title{Retrieve the name of the model that a segmenter or model used}
\usage{
\method{model_name}{mod_cpt}(object, ...)

\method{model_name}{seg_basket}(object, ...)

\method{model_name}{seg_cpt}(object, ...)

\method{model_name}{tidycpt}(object, ...)

model_name(object, ...)

\method{model_name}{default}(object, ...)

\method{model_name}{character}(object, ...)

\method{model_name}{ga}(object, ...)

\method{model_name}{cpt}(object, ...)

\method{model_name}{wbs}(object, ...)
}
\arguments{
\item{object}{A \linkS4class{tidycpt} object, typically returned by \code{\link[=segment]{segment()}}}

\item{...}{currently ignored}
}
\value{
\itemize{
\item \code{\link[=model_name]{model_name()}}: A \code{character} vector of length 1.
}
}
\description{
Retrieve the name of the model that a segmenter or model used
}
\details{
Every segmenter works by fitting a model to the data. \code{\link[=model_name]{model_name()}} returns
the name of a model that can be passed to \code{\link[=whomademe]{whomademe()}} to retrieve the
model fitting function. These functions must begin with the prefix \code{fit_}.
Note that the model fitting functions exist in \code{tidychangepoint} are are
not necessarily the actual functions used by the segmenter.

Models also implement \code{model_name()}.
}
\examples{
x <- segment(CET, method = "pelt")
model_name(x$segmenter)
whomademe(model_name(x))
model_name(x$segmenter)
model_name(x$model)
}
\seealso{
Other modeling: 
\code{\link{model_args.seg_cpt}()},
\code{\link{new_fun_cpt}()}
}
\concept{modeling}
