% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mod_default.R, R/class-seg_default.R,
%   R/class-tidycpt.R, R/generics.R, R/pkg-GA.R, R/pkg-changepoint.R,
%   R/pkg-wbs.R, R/utils.R
\name{model_name.mod_default}
\alias{model_name.mod_default}
\alias{model_name.seg_default}
\alias{model_name.tidycpt}
\alias{model_name}
\alias{model_name.default}
\alias{model_name.character}
\alias{model_name.ga}
\alias{model_name.cpt}
\alias{model_name.wbs}
\alias{model_fit}
\title{Retrieve the name of the model that a segmenter or model used}
\usage{
\method{model_name}{mod_default}(object, ...)

\method{model_name}{seg_default}(object, ...)

\method{model_name}{tidycpt}(object, ...)

model_name(object, ...)

\method{model_name}{default}(object, ...)

\method{model_name}{character}(object, ...)

\method{model_name}{ga}(object, ...)

\method{model_name}{cpt}(object, ...)

\method{model_name}{wbs}(object, ...)

model_fit(x, ...)
}
\arguments{
\item{object}{A segmenter object, typically returned by \code{\link[=segment]{segment()}}}

\item{...}{currently ignored}

\item{x}{A \code{character} giving the name of a model. Typically the result of
\code{\link[=model_name]{model_name()}}.}
}
\value{
\itemize{
\item \code{\link[=model_name]{model_name()}}: A \code{character} vector of length 1.
}

\itemize{
\item \code{\link[=model_fit]{model_fit()}}: A function
}
}
\description{
Retrieve the name of the model that a segmenter or model used
}
\details{
Every segmenter works by fitting a model to the data. \code{\link[=model_name]{model_name()}} returns
the name of a model that can be passed to \code{\link[=model_fit]{model_fit()}} to retrieve the
model fitting function. These functions must begin with the prefix \code{fit_}.
Note that the model fitting functions exist in \code{tidychangepoint} are are
not necessarily the actual functions used by the segmenter.

Models also implement \code{model_name()}.
}
\examples{
x <- segment(CET, method = "pelt")
model_name(x$segmenter)
model_fit(model_name(x))
model_name(x$segmenter)
model_name(x$model)
}
\seealso{
Other modeling: 
\code{\link{model_args.seg_default}()}

Other modeling: 
\code{\link{model_args.seg_default}()}
}
\concept{modeling}
