% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mod_default.R, R/class-seg_default.R,
%   R/class-tidycpt.R, R/generics.R, R/pkg-GA.R, R/pkg-changepoint.R,
%   R/pkg-wbs.R
\name{model_name.mod_default}
\alias{model_name.mod_default}
\alias{model_name.seg_default}
\alias{model_name.tidycpt}
\alias{model_name}
\alias{model_name.default}
\alias{model_name.character}
\alias{model_args.default}
\alias{model_name.ga}
\alias{model_name.cpt}
\alias{model_name.wbs}
\title{Retrieve the model that a segmenter used}
\usage{
\method{model_name}{mod_default}(object, ...)

\method{model_name}{seg_default}(object, ...)

\method{model_name}{tidycpt}(object, ...)

model_name(object, ...)

\method{model_name}{default}(object, ...)

\method{model_name}{character}(object, ...)

\method{model_args}{default}(object, ...)

\method{model_name}{ga}(object, ...)

\method{model_name}{cpt}(object, ...)

\method{model_name}{wbs}(object, ...)
}
\arguments{
\item{object}{An object, typically returned by \code{\link[=segment]{segment()}}}

\item{...}{currently ignored}
}
\description{
Retrieve the model that a segmenter used
}
