% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-mod_cpt.R
\name{model_variance}
\alias{model_variance}
\title{Compute model variance}
\usage{
model_variance(object, ...)
}
\arguments{
\item{object}{A model object implementing \code{\link[=residuals]{residuals()}} and \code{\link[=nobs]{nobs()}}}

\item{...}{currently ignored}
}
\value{
A \code{double} vector of length 1
}
\description{
Compute model variance
}
\details{
Using the generic functions \code{\link[=residuals]{residuals()}} and \code{\link[=nobs]{nobs()}}, this function
computes the variance of the residuals.

Note that unlike \code{\link[stats:cor]{stats::var()}}, it does not use \eqn{n-1} as the denominator.
}
