% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_bmdl.R
\name{muta_1_cp_BMDL}
\alias{muta_1_cp_BMDL}
\alias{muta_k_cp_BMDL}
\title{Mutaciones un cp en el caso BMDL}
\usage{
muta_1_cp_BMDL(cp, x, param)

muta_k_cp_BMDL(mat_cp, x, param)
}
\arguments{
\item{cp}{puntos de cambio}

\item{x}{vector de revases}

\item{param}{parametros}
}
\value{
regresa un vector mutado

regreas una mat_cp mutada
}
\description{
Mutaciones un cp en el caso BMDL
}
