% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpt_gbmdl.R
\name{new_cpt_gbmdl}
\alias{new_cpt_gbmdl}
\alias{validate_cpt_gbmdl}
\alias{cpt_gbmdl}
\alias{plot.cpt_gbmdl}
\alias{print.cpt_gbmdl}
\alias{summary.cpt_gbmdl}
\alias{nobs.cpt_gbmdl}
\alias{logLik.cpt_gbmdl}
\alias{cpt_best_bmdl}
\alias{cpt_best_bmdl_string}
\alias{cpt_best_params}
\alias{chromosome_best}
\alias{cpt_best}
\alias{chromo2tau}
\alias{mat_cp_2_tbl}
\alias{mat_cp_2_list}
\alias{max_num_cp}
\alias{generation_size}
\alias{num_generations}
\alias{write_cpt_gbmdl}
\alias{file_name}
\title{Class for candidate changepoints using Genetic BMDL heuristic}
\usage{
new_cpt_gbmdl(
  x = numeric(),
  nhpp_dist = c("W", "EW", "GGO", "MO", "GO")[1],
  vec_dist_a_priori = c("Gamma", "Gamma"),
  mat_phi = matrix(c(1, 3, 2, 1.2), ncol = 2),
  num_generations = 50,
  generation_size = 50,
  max_num_cp = 20
)

validate_cpt_gbmdl(x)

cpt_gbmdl(x, ...)

\method{plot}{cpt_gbmdl}(x, ...)

\method{print}{cpt_gbmdl}(x, ...)

\method{summary}{cpt_gbmdl}(object, ...)

\method{nobs}{cpt_gbmdl}(object, ...)

\method{logLik}{cpt_gbmdl}(object, ...)

cpt_best_bmdl(x)

cpt_best_bmdl_string(x)

cpt_best_params(x)

chromosome_best(x)

cpt_best(x)

chromo2tau(chromo)

mat_cp_2_tbl(mat_cp)

mat_cp_2_list(mat_cp)

max_num_cp(x)

generation_size(x)

num_generations(x)

write_cpt_gbmdl(x, destdir = tempdir())

file_name(x, data_name_slug = "data")
}
\arguments{
\item{x}{a numeric vector}

\item{nhpp_dist}{toma valores en c("W","EW","GGO","MO","GO") y es el nombre de
la función de tasa del NHPP}

\item{vec_dist_a_priori}{vector de los nobmres de las distribuciones a priori
que se utilizan; eg c("Gamma","Gamma") y c("Gamma","Gamma","Gamma")}

\item{mat_phi}{matriz cuyos renglones tiene los parámetros de las
distribuciones a priori; cada renglón tiene todos los parametros de una
distribución}

\item{num_generations}{número de generaciones}

\item{generation_size}{tamaño de las generaciones}

\item{data_name_slug}{character string that will identify the data set used
in the file name}
}
\description{
Class for candidate changepoints using Genetic BMDL heuristic
}
\examples{
cpts <- cpt_gbmdl(DataCPSim)
str(cpts)
cpt_best_params(lista_AG$segmenter)
chromo <- c(4, 1, 557, 877 , 905, 986, 1096, 0, 0, 0)
chromo2tau(chromo)
write_cpt_gbmdl(lista_AG$segmenter)
file_name(cpt_gbmdl(DataCPSim))
}
