% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpt_list.R
\name{new_cpt_list}
\alias{new_cpt_list}
\alias{validate_cpt_list}
\alias{cpt_list}
\alias{plot.cpt_list}
\alias{print.cpt_list}
\alias{summary.cpt_list}
\alias{cpt_best_bmdl}
\alias{cpt_best_bmdl_string}
\alias{cpt_best_params}
\alias{chromosome_best}
\alias{cpt_best}
\alias{write_cpt_list}
\title{Class for candidate changepoints}
\usage{
new_cpt_list(x = numeric(), param = list())

validate_cpt_list(x)

cpt_list(x, ...)

\method{plot}{cpt_list}(x, ...)

\method{print}{cpt_list}(x, ...)

\method{summary}{cpt_list}(object, ...)

cpt_best_bmdl(x)

cpt_best_bmdl_string(x)

cpt_best_params(x)

chromosome_best(x)

cpt_best(x)

write_cpt_list(x, destdir = tempdir())
}
\arguments{
\item{x}{a numeric vector}
}
\description{
Class for candidate changepoints
}
\examples{
cpts <- cpt_list(DataCPSim, param)
str(cpts)
plot(cpts)
cpt_best_params(lista_AG)

write_cpt_list(lista_AG)
}
