% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-seg_cpt.R
\name{new_seg_cpt}
\alias{new_seg_cpt}
\alias{validate_seg_cpt}
\alias{seg_cpt}
\title{Base class for segmenters}
\usage{
new_seg_cpt(
  x = numeric(),
  pkg = character(),
  algorithm = NA,
  changepoints = integer(),
  fitness = double(),
  seg_params = list(),
  model_name = "meanshift_norm",
  penalty = "BIC",
  ...
)

validate_seg_cpt(x)

seg_cpt(x, ...)
}
\arguments{
\item{x}{a numeric vector coercible into a \code{ts} object}

\item{pkg}{name of the package providing the segmenter}

\item{algorithm}{Algorithm used to find the changepoints}

\item{changepoints}{a possibly empty \code{list()} of candidate changepoints}

\item{fitness}{A named \code{double} vector whose name reflects the penalty applied}

\item{seg_params}{a possibly empty \code{list()} of segmenter parameters}

\item{model_name}{character indicating the model used to find the changepoints.}

\item{penalty}{character indicating the name of the penalty function used to
find the changepoints.}

\item{...}{currently ignored}
}
\description{
Base class for segmenters
}
\examples{
x <- seg_cpt(DataCPSim, pkg = "tidychangepoint", changepoints = c(330))
str(x)
as.ts(x)
changepoints(x)

y <- segment(CET, method = "pelt")
z <- as.seg_cpt(y$segmenter)
class(z)
fitness(x)
glance(z)
model_name(z)
model_args(z)
nobs(z)
seg_params(z)
}
