% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_default.R
\name{new_seg_default}
\alias{new_seg_default}
\alias{validate_seg_default}
\alias{seg_default}
\alias{as.ts.seg_default}
\alias{changepoints.seg_default}
\alias{evaluate}
\alias{glance.seg_default}
\title{Default class for candidate changepoint sets}
\usage{
new_seg_default(x = numeric(), cpt_list = list(), params = list(), ...)

validate_seg_default(x)

seg_default(x, ...)

\method{as.ts}{seg_default}(x, ...)

\method{changepoints}{seg_default}(x, ...)

evaluate(x, ...)

\method{glance}{seg_default}(x, ...)
}
\arguments{
\item{x}{a numeric vector coercible into a \code{ts} object}

\item{cpt_list}{a possibly empty \code{list()} of candidate changepoints}

\item{params}{a possibly empty \code{list()} of parameters}
}
\description{
Default class for candidate changepoint sets
}
\examples{
seg <- seg_default(DataCPSim)
str(seg)
as.ts(seg)
changepoints(seg)
}
