% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-seg_default.R, R/generics.R
\name{new_seg_default}
\alias{new_seg_default}
\alias{validate_seg_default}
\alias{seg_default}
\alias{num_candidates}
\alias{evaluate_cpts.seg_default}
\alias{evaluate_cpts.list}
\alias{evaluate_cpts.tbl_df}
\alias{evaluate_cpts}
\title{Default class for candidate changepoint sets}
\usage{
new_seg_default(
  x = numeric(),
  algorithm = NA,
  cpt_list = list(),
  params = list(),
  model_name = "meanshift",
  penalty = "BIC",
  ...
)

validate_seg_default(x)

seg_default(x, ...)

num_candidates(x)

\method{evaluate_cpts}{seg_default}(x, model_fn, ...)

\method{evaluate_cpts}{list}(x, model_fn, ...)

\method{evaluate_cpts}{tbl_df}(x, model_fn, ...)

evaluate_cpts(x, ...)
}
\arguments{
\item{x}{a numeric vector coercible into a \code{ts} object}

\item{algorithm}{Algorithm used to find the changepoints}

\item{cpt_list}{a possibly empty \code{list()} of candidate changepoints}

\item{params}{a possibly empty \code{list()} of parameters}

\item{model_name}{character indicating the model used to find the changepoints.}

\item{penalty}{character indicating the name of the penalty function used to
find the changepoints.}

\item{...}{currently ignored}

\item{model_fn}{Name of the function to fit the model.
See, for examples, \code{\link[=fit_meanshift]{fit_meanshift()}}.}
}
\description{
Default class for candidate changepoint sets
}
\examples{
seg <- seg_default(DataCPSim, cpt_list = list(c(365), c(330, 839)))
str(seg)
as.ts(seg)
changepoints(seg)
fitness(seg)
glance(seg)
}
