% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pad_tau}
\alias{pad_tau}
\alias{unpad_tau}
\alias{cut_inclusive}
\alias{test_set}
\alias{split_by_tau}
\alias{deg_free}
\title{Utility functions}
\usage{
pad_tau(tau, n)

unpad_tau(padded_tau)

cut_inclusive(x, tau)

test_set(n = 1, sd = 1, seed = NULL)

split_by_tau(x, tau)

deg_free(x)
}
\arguments{
\item{tau}{a numeric vector of changepoints}

\item{n}{Number of changepoints}

\item{padded_tau}{Output from \code{\link[=pad_tau]{pad_tau()}}}

\item{x}{A numeric vector}

\item{sd}{Standard deviation passed to \code{\link[stats:Normal]{stats::rnorm()}}}

\item{seed}{Value passed to \code{\link[base:Random]{base::set.seed()}}}
}
\description{
Utility functions
}
\examples{
split_by_tau(as.ts(lista_AG), changepoints(lista_AG))
deg_free(segment(DataCPSim))
}
