% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pad_tau}
\alias{pad_tau}
\alias{unpad_tau}
\alias{is_valid_tau}
\alias{validate_tau}
\alias{binary2tau}
\alias{tau2binary}
\alias{tau2time}
\alias{time2tau}
\alias{cut_inclusive}
\alias{split_by_tau}
\alias{regions_by_tau}
\alias{deg_free}
\alias{as_year}
\alias{whoami}
\title{Utility functions}
\usage{
pad_tau(tau, n)

unpad_tau(padded_tau)

is_valid_tau(tau, n)

validate_tau(tau, n)

binary2tau(x)

tau2binary(tau, n)

tau2time(tau, index)

time2tau(cpts, index)

cut_inclusive(x, tau)

split_by_tau(x, tau)

regions_by_tau(n, tau)

deg_free(x)

as_year(x)

whoami(x = fit_meanshift, ...)
}
\arguments{
\item{tau}{a numeric vector of changepoints}

\item{n}{the length of the original time series}

\item{padded_tau}{Output from \code{\link[=pad_tau]{pad_tau()}}}

\item{x}{A numeric vector}

\item{index}{Index of times, typically returned by \code{\link[stats:time]{stats::time()}}}

\item{cpts}{Time series observation labels to be converted to indices}

\item{...}{currently ignored}
}
\description{
Utility functions
}
\examples{
is_valid_tau(0, length(DataCPSim))
is_valid_tau(1, length(DataCPSim))
is_valid_tau(826, length(DataCPSim))
is_valid_tau(1096, length(DataCPSim))
is_valid_tau(1097, length(DataCPSim))

validate_tau(0, length(DataCPSim))
validate_tau(1, length(DataCPSim))
validate_tau(826, length(DataCPSim))
validate_tau(c(826, 826), length(DataCPSim))
validate_tau(1096, length(DataCPSim))
validate_tau(1097, length(DataCPSim))
validate_tau(c(-4, 0, 1, 4, 5, 5, 824, 1096, 1097, 182384), length(DataCPSim))

binary2tau(c(0, 0, 1, 0, 1))
binary2tau(round(runif(10)))

tau2binary(c(7, 17), n = 24)
tau2binary(binary2tau(c(0, 0, 1, 1, 0, 1)), n = 6)
tau2time(c(42, 81, 330), index = as_year(time(CET)))
time2tau(c(1700, 1739, 1988), index = as_year(time(CET)))
split_by_tau(DataCPSim, c(365, 826))
regions_by_tau(1096, c(365, 826))
deg_free(segment(DataCPSim)$model)
as_year("1988-01-01")
whoami(fit_meanshift)
whoami(fit_meanshift_ar1)
whoami(fit_lmshift)
whoami(fit_nhpp)
}
\seealso{
\code{\link[stats:time]{stats::time()}}
}
