% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-FnAGBayesianMDL.R
\name{penalization_MDL}
\alias{penalization_MDL}
\title{Penalizaci贸n MDL}
\usage{
penalization_MDL(cp, rf_type)
}
\arguments{
\item{cp}{vector de extendido de puntos de cambio}
}
\value{
regresa la evaluaci贸n de la penalizaci贸n
\deqn{
   P_{\theta,\tau} = \sum_{i=1}^{m+1}\dfrac{\ln(\tau_i-\tau_{i-1})}{2}+\ln(m)+\sum_{i=2}^m\ln(\tau_i)
 }
}
\description{
Penalizaci贸n MDL
}
