% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalty.R
\name{penalization_MDL}
\alias{penalization_MDL}
\title{Penalizaci贸n MDL}
\usage{
penalization_MDL(cp, nhpp_dist = "W", N)
}
\arguments{
\item{cp}{vector de extendido de puntos de cambio}
}
\value{
regresa la evaluaci贸n de la penalizaci贸n
\deqn{
   P_{\theta,\tau} = \sum_{i=1}^{m+1}\dfrac{\ln(\tau_i-\tau_{i-1})}{2}+\ln(m)+\sum_{i=2}^m\ln(\tau_i)
 }
}
\description{
Penalizaci贸n MDL
}
\examples{
mat_cp <- sim_k_cp_BMDL(DataCPSim)
penalization_MDL(mat_cp, N = max(exceedances(DataCPSim)))


}
