% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalty.R
\name{penalty_mdl}
\alias{penalty_mdl}
\title{Penalizaci贸n MDL}
\usage{
penalty_mdl(padded_tau)
}
\arguments{
\item{cp}{vector de extendido de puntos de cambio}
}
\value{
regresa la evaluaci贸n de la penalizaci贸n
\deqn{
   P_{\theta,\tau} = \sum_{i=1}^{m+1}\dfrac{\ln(\tau_i-\tau_{i-1})}{2}+\ln(m)+\sum_{i=2}^m\ln(\tau_i)
 }
}
\description{
Penalizaci贸n MDL
}
\examples{
x <- segment(DataCPSim, method = "cpt-pelt")
tau <- changepoints(x)
penalty_mdl(pad_tau(tau, n = length(as.ts(x))))

}
