% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_cpt_repetidos}
\alias{plot_cpt_repetidos}
\title{Grafica los n puntos de cambio más repetidos sobre el plot.stepfun de los datos}
\usage{
plot_cpt_repetidos(cpt_list, destdir = tempdir(), pdf = FALSE)
}
\arguments{
\item{cpt_list}{description}
}
\value{
guarda un pdf con la gráfica
}
\description{
Grafica los n puntos de cambio más repetidos sobre el plot.stepfun de los datos
}
\examples{
plot_cpt_repetidos(lista_AG)

}
