% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_generics.R, R/class-mod_cpt.R,
%   R/class-tidycpt.R
\name{regions}
\alias{regions}
\alias{regions.mod_cpt}
\alias{regions.tidycpt}
\title{Extract the regions from a \linkS4class{tidycpt} object}
\usage{
regions(x, ...)

\method{regions}{mod_cpt}(x, ...)

\method{regions}{tidycpt}(x, ...)
}
\arguments{
\item{x}{An object that has regions}

\item{...}{Currently ignored}
}
\value{
A \code{\link[base:factor]{base::factor()}} of intervals indicating the region
}
\description{
Extract the regions from a \linkS4class{tidycpt} object
}
\examples{

cpt <- fit_meanshift_norm(CET, tau = 330)
regions(cpt)

}
