% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_bmdl.R
\name{revisor_param}
\alias{revisor_param}
\title{Validador de la lista param para ejecutar el AG-BMDL}
\usage{
revisor_param(
  param,
  r = 50,
  k = 50,
  max_num_cp = 40,
  prob_inicial = 0.01,
  probs_muta = c(0.4, 0.2, 0.4),
  mutaciones = c(-1, 0, 1),
  dist_extremos = 10,
  prob_para_sin_cp = 0.5,
  probs_rank0_MDL1 = 0,
  cuantos_mejores_cp_graf = 100,
  minimo_numero_de_cp = 5,
  probs_nuevos_muta0N = c(0.5, 0.2, 0.2, 0.1),
  rf_type = c("W", "EW", "GGO", "MO", "GO")[1],
  initial_val_optim = c(0.1, 0.5),
  mat_low_upp = matrix(c(c(1e-04, 1e-08), c(10, 1e+05)), nrow = 2),
  vec_dist_a_priori = c("Gamma", "Gamma"),
  mat_phi = matrix(c(1, 3, 2, 1.2), ncol = 2)
)
}
\arguments{
\item{param}{es la lista original de parámetros la cual contiene todos los
siguientes}

\item{r}{número de generaciones}

\item{k}{tamaño de las generaciones}

\item{max_num_cp}{el máximo número de rebases. Este parámetro se ocupa en
particular para que todos los cromosomas quepan en una matriz.}

\item{prob_inicial}{probabilidad de que en la primera generación un punto
cualquiera sea punto de cambio. Se recomienda =.5 ya que con esto se
distribuyen relativamente uniformes los puntos de cambio}

\item{probs_muta}{probabilidades de mutación. Las longitudes de este vector y
mutaciones tienen que ser iguales; eg si mutaciones=c(-1,0,1) y probs_muta
= c(.2, .6, .2) entonces se tiene una probabilidad .2 de que el punto de
cambio se desplace a la izquierda, probabilidad .6 de quedar igual, y
probabilidad . 2 de ser movido a la derecha}

\item{mutaciones}{vector con mutaciones posibles; eg si mutaciones=c(-1,0,1)
entonces un punto de cambio puede ser movido una unidad a la izquierda,
puede quedarse igual, o moverse una unidad a la derecha}

\item{dist_extremos}{distancia entre el primer los puntos de cambio v_0 y v_1
al igual que entre v_m y v_{m+1}; distancia minima que debe de haber de un
punto de cambio y los valores 1 y T, donde T es la longitud total de la
serie}

\item{prob_para_sin_cp}{En caso de querer mutar un cp sin puntos de cambio se
lanza un volado con probabilidad prob_para_sin_cp, si es cae 1 se regresa
el cp original, si cae 0 se simula un punto de cambio y se regresa este cp
con este punto de cambio}

\item{probs_rank0_MDL1}{para medir obtener la probabilidad de los padres se
pueden tomar o las probabilidades con respecto a los rangos (como en el
artículo) o se pueden tomar las probabilidades con respecto a el MDL. La
diferencia radica en que si se toma con respecto al MDL se tendrá que un
cromosoma con un gran MDL este tendrá una gran ventaja de ocurrir, en
cambio cuando solo se tiene rank esta ventaja gran ventaja se reduce}

\item{cuantos_mejores_cp_graf}{al final se generan unas graficas de los
mejores puntos de cambio, este parámetro dicta cuantos cromosomas se
graficarán}

\item{minimo_numero_de_cp}{es la cota inferior del número de puntos de cambio
que puede tener un cromosoma}

\item{probs_nuevos_muta0N}{probabilidades de mutar 0,1,2,...,l hasta cierto
numero l; eg si vale c(.5,.2,.2,.1) se tiene una probabilidad 0.5 de mutar
0 (de no mutar), probabilidad 0.2 de mutar 1,, probabilidad 0.2 de mutar 2,
y, probabilidad 0.1 de mutar 3.}

\item{rf_type}{toma valores en c("W","EW","GGO","MO","GO") y es el nombre de
la función de tasa del NHPP}

\item{initial_val_optim}{valores iniciales de busqueda del MAP para los
parámetros del NHPP}

\item{mat_low_upp}{rango de busqueda de los parámetros para el MAP}

\item{vec_dist_a_priori}{vector de los nobmres de las distribuciones a priori
que se utilizan; eg c("Gamma","Gamma") y c("Gamma","Gamma","Gamma")}

\item{mat_phi}{matriz cuyos renglones tiene los parámetros de las
distribuciones a priori; cada renglón tiene todos los parametros de una
distribución}
}
\value{
regresa un mensaje en caso de que no sea de la longitud correcta el
vector
}
\description{
Validador de la lista param para ejecutar el AG-BMDL
}
\examples{
revisor_param(param)
}
