% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_bmdl.R
\name{revisor_param}
\alias{revisor_param}
\title{Validador de la lista param para ejecutar el AG-BMDL}
\usage{
revisor_param(
  param,
  nombre_datos = c("ciclones", "O3", "O3_U61", "PM10", "PM10_U100")[5],
  frecuencia_datos = 7,
  n_datos = c("TODOS", 403, 1003)[1],
  diarios0_rebases1 = 0,
  valor_de_rebase = 100,
  r = 50,
  k = 50,
  penalty = c("MDL", "BMDL")[2],
  max_num_cp = 40,
  prob_inicial = 0.01,
  prob_volado = 0.6,
  probs_muta = c(0.4, 0.2, 0.4),
  mutaciones = c(-1, 0, 1),
  dist_extremos = 10,
  prob_para_sin_cp = 0.5,
  cp_real = "sin cp_real",
  quita_ini0_fin1 = 0,
  probs_rank0_MDL1 = 0,
  nombre_carpeta_pdf = "Figures",
  nombre_carpeta_RData = "Data",
  cuantos_mejores_cp_graf = 100,
  my_data = list(nulo = NULL, "my_data=my_data")[[1]],
  minimo_numero_de_cp = 5,
  probs_nuevos_muta0N = c(0.5, 0.2, 0.2, 0.1),
  rf_type = c("W", "EW", "GGO", "MO", "GO")[1],
  initial_val_optim = c(0.1, 0.5),
  mat_low_upp = matrix(c(c(1e-04, 1e-08), c(10, 1e+05)), nrow = 2),
  vec_dist_a_priori = c("Gamma", "Gamma"),
  mat_phi = matrix(c(1, 3, 2, 1.2), ncol = 2),
  ajuste_bloque = T
)
}
\arguments{
\item{param}{es la lista original de parámetros la cual contiene todos los
siguientes}

\item{nombre_datos}{esta variable indica la base de datos que se va a ocupar.
Esta base se toma del archivo "Datos.R"; los posibles valores que toma son
c("ciclones","O3","O3_U61","PM10","PM10_U100")}

\item{frecuencia_datos}{indica se hay que hacer un cambio en la frecuencia de
los datos. En particular, se utiliza junto con la variable
diarios0_rebases1 = 1, (si esta última vale 0, frecuencia_datos no tiene
efecto)}

\item{n_datos}{Es una variable para hacer pruebas que cirve para tocar un
subconjunto de los datos, o en caso de ser "TODOS" tomar todos los datos}

\item{diarios0_rebases1}{indicadora si se toman los revasos o los datos
crudos. En particular se utiliza para el caso de dist="poisson", ya que al
tomar los datos distribuidos poisson se estará contando el número de
revases}

\item{valor_de_rebase}{es la indicadora del nivel para revase, por lo general
vale 100}

\item{r}{número de generaciones}

\item{k}{tamaño de las generaciones}

\item{penalty}{tipo de penalidad que se ocupa, por ahora solo se tiene
programado MDL}

\item{max_num_cp}{el máximo número de rebases. Este parámetro se ocupa en
particular para que todos los cromosomas quepan en una matriz.}

\item{prob_inicial}{probabilidad de que en la primera generación un punto
cualquiera sea punto de cambio. Se recomienda =.5 ya que con esto se
distribuyen relativamente uniformes los puntos de cambio}

\item{prob_volado}{probabilidad de quitar un tiempo de cambio existente
utilizado por mata_k_tau_volado para quitar elementos de más. Se recomienda
dejar el valor de 0.5 ya que así al juntar los pc del padre y madre se
eliminará la mitad de estos}

\item{probs_muta}{probabilidades de mutación. Las longitudes de este vector y
mutaciones tienen que ser iguales; eg si mutaciones=c(-1,0,1) y probs_muta
= c(.2, .6, .2) entonces se tiene una probabilidad .2 de que el punto de
cambio se desplace a la izquierda, probabilidad .6 de quedar igual, y
probabilidad . 2 de ser movido a la derecha}

\item{mutaciones}{vector con mutaciones posibles; eg si mutaciones=c(-1,0,1)
entonces un punto de cambio puede ser movido una unidad a la izquierda,
puede quedarse igual, o moverse una unidad a la derecha}

\item{dist_extremos}{distancia entre el primer los puntos de cambio v_0 y v_1
al igual que entre v_m y v_{m+1}; distancia minima que debe de haber de un
punto de cambio y los valores 1 y T, donde T es la longitud total de la
serie}

\item{prob_para_sin_cp}{En caso de querer mutar un cp sin puntos de cambio se
lanza un volado con probabilidad prob_para_sin_cp, si es cae 1 se regresa
el cp original, si cae 0 se simula un punto de cambio y se regresa este cp
con este punto de cambio}

\item{cp_real}{en caso de estar haciendo pruebas y conocer los puntos de
cambio reales da un vector de tamaño max_num_cp con entradas
m,tau_0,...,tau_{m+1},0,...,0, en particual esto cirve para que al final se
grafiquen los puntos de cambio reales contra los estimados; en caso de no
conocer los pc reales, esta variable tiene el valor "sin cp_real"}

\item{quita_ini0_fin1}{es una indicadora para cuando se hacen rebases y la
serie de tiempo x (la serie original) no es exanctamente de la longitud de
frecuencia_datos. Entonces se quitan los restantes del final o del
principio}

\item{probs_rank0_MDL1}{para medir obtener la probabilidad de los padres se
pueden tomar o las probabilidades con respecto a los rangos (como en el
artículo) o se pueden tomar las probabilidades con respecto a el MDL. La
diferencia radica en que si se toma con respecto al MDL se tendrá que un
cromosoma con un gran MDL este tendrá una gran ventaja de ocurrir, en
cambio cuando solo se tiene rank esta ventaja gran ventaja se reduce}

\item{nombre_carpeta_pdf}{normalmente esta carpeta se llama "Figures", y
contiene las figuras que genera el programa}

\item{nombre_carpeta_RData}{normalmente esta carpeta se llama "Data", y
contiene archivos .RData que tienen los resultados del programa}

\item{cuantos_mejores_cp_graf}{al final se generan unas graficas de los
mejores puntos de cambio, este parámetro dicta cuantos cromosomas se
graficarán}

\item{my_data}{en caso de que se quiera hacer pruebas con otra serie en esta
variable se carga la serie con la que se desea hacer las pruebas. En este
caso, se comenta la entrada llamada my_data para que se pueda correr la
función}

\item{minimo_numero_de_cp}{es la cota inferior del número de puntos de cambio
que puede tener un cromosoma}

\item{probs_nuevos_muta0N}{probabilidades de mutar 0,1,2,...,l hasta cierto
numero l; eg si vale c(.5,.2,.2,.1) se tiene una probabilidad 0.5 de mutar
0 (de no mutar), probabilidad 0.2 de mutar 1,, probabilidad 0.2 de mutar 2,
y, probabilidad 0.1 de mutar 3.}

\item{rf_type}{toma valores en c("W","EW","GGO","MO","GO") y es el nombre de
la función de tasa del NHPP}

\item{initial_val_optim}{valores iniciales de busqueda del MAP para los
parámetros del NHPP}

\item{mat_low_upp}{rango de busqueda de los parámetros para el MAP}

\item{vec_dist_a_priori}{vector de los nobmres de las distribuciones a priori
que se utilizan; eg c("Gamma","Gamma") y c("Gamma","Gamma","Gamma")}

\item{mat_phi}{matriz cuyos renglones tiene los parámetros de las
distribuciones a priori; cada renglón tiene todos los parametros de una
distribución}

\item{ajuste_bloque}{en caso de ser TRUE se utiliza el ajuste por bloque de
NHPP para cada bloque del ajuste; ie, siempre se ajusta como primer bloque
y se toma en cuenta en que pedazo se empieza}
}
\value{
regresa un mensaje en caso de que no sea de la longitud correcta el
vector
}
\description{
Validador de la lista param para ejecutar el AG-BMDL
}
\examples{
revisor_param(param)
}
