% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-seg_basket.R
\name{seg-basket-generics}
\alias{seg-basket-generics}
\alias{as.segmenter.seg_basket}
\alias{as.ts.seg_basket}
\alias{best_cpt}
\alias{plot.seg_basket}
\title{Methods for seg_basket objects}
\usage{
\method{as.segmenter}{seg_basket}(object, ...)

\method{as.ts}{seg_basket}(x, ...)

best_cpt(x, ...)

\method{plot}{seg_basket}(x, ...)
}
\description{
Methods for seg_basket objects
}
