% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seg_default.R
\name{seg-default-generics}
\alias{seg-default-generics}
\alias{as.ts.seg_default}
\alias{params.seg_default}
\alias{glance.seg_default}
\alias{plot.seg_default}
\title{Methods for seg_default objects}
\usage{
\method{as.ts}{seg_default}(x, ...)

\method{params}{seg_default}(x, ...)

\method{glance}{seg_default}(x, ...)

\method{plot}{seg_default}(x, ...)
}
\arguments{
\item{x}{An \code{seg_default} object}

\item{...}{arguments passed to methods}
}
\description{
Methods for seg_default objects
}
