% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-seg_cpt.R, R/generics.R, R/pkg-GA.R,
%   R/pkg-changepoint.R, R/pkg-wbs.R
\name{seg_params.seg_cpt}
\alias{seg_params.seg_cpt}
\alias{seg_params}
\alias{seg_params.ga}
\alias{seg_params.cpt}
\alias{seg_params.wbs}
\title{Retrieve parameters from a segmenter}
\usage{
\method{seg_params}{seg_cpt}(x, ...)

seg_params(x, ...)

\method{seg_params}{ga}(x, ...)

\method{seg_params}{cpt}(x, ...)

\method{seg_params}{wbs}(x, ...)
}
\value{
A named \code{list} of parameters with their values.
}
\description{
Retrieve parameters from a segmenter
}
\details{
Most segmenting algorithms have parameters.
This function retrieves an informative set of those parameter values.
}
\examples{
# Segment a time series using PELT
x <- segment(CET, method = "pelt")
x |>
  as.segmenter() |>
  seg_params()

}
\seealso{
Other tidychangepoint-generics: 
\code{\link{changepoints.mod_cpt}()},
\code{\link{diagnose.mod_cpt}()},
\code{\link{exceedances}()},
\code{\link{fitness.seg_basket}()}
}
\concept{tidychangepoint-generics}
