% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment}
\alias{segment}
\alias{segment.tbl_ts}
\alias{segment.xts}
\alias{segment.numeric}
\alias{segment.ts}
\title{Segment a time series using a variety of algorithms}
\usage{
segment(x, method = "null", ...)

\method{segment}{tbl_ts}(x, method = "null", ...)

\method{segment}{xts}(x, method = "null", ...)

\method{segment}{numeric}(x, method = "null", ...)

\method{segment}{ts}(x, method = "null", ...)
}
\arguments{
\item{x}{a numeric vector coercible into a \link[stats:ts]{stats::ts} object}

\item{method}{a character string indicating the algorithm to use. See Details.}

\item{...}{arguments passed to methods}
}
\value{
An object of class \linkS4class{tidycpt}.
}
\description{
A wrapper function that encapsulates various algorithms for detecting changepoint
sets in univariate time series.
}
\details{
Currently, \code{\link[=segment]{segment()}} can use the following algorithms, depending
on the value of the \code{method} argument:
\itemize{
\item \code{pelt}: Uses the PELT algorithm as implemented in
\code{\link[changepoint:cpt.meanvar]{changepoint::cpt.meanvar()}}. The \code{segmenter} is of class \code{cpt}.
\item \code{binseg}: Uses the Binary Segmentation algorithm as implemented by
\code{\link[changepoint:cpt.meanvar]{changepoint::cpt.meanvar()}}. The \code{segmenter} is of class \code{cpt}.
\item \code{segneigh}: Uses the Segmented Neighborhood algorithm as implemented by
\code{\link[changepoint:cpt.meanvar]{changepoint::cpt.meanvar()}}. The \code{segmenter} is of class \code{cpt}.
\item \code{single-best}: Uses the AMOC criteria as implemented by
\code{\link[changepoint:cpt.meanvar]{changepoint::cpt.meanvar()}}. The \code{segmenter} is of class \code{cpt}.
\item \code{wbs}: Uses the Wild Binary Segmentation algorithm as implemented by
\code{\link[wbs:wbs]{wbs::wbs()}}. The \code{segmenter} is of class \code{wbs}.
\item \code{ga}: Uses the Genetic algorithm implemented by \code{\link[=segment_ga]{segment_ga()}}, which wraps
\code{\link[GA:ga]{GA::ga()}}. The \code{segmenter} is of class \code{tidyga}.
\item \code{coen}: Uses Coen's heuristic as implemented by
\code{\link[=segment_coen]{segment_coen()}}. The \code{segmenter} is of class \code{\link[=seg_basket]{seg_basket()}}.
\item \code{manual}: Uses the vector of changepoints in the \code{tau} argument.
The \code{segmenter} is of class \code{\link[=seg_cpt]{seg_cpt()}}`.
\item \code{null}: The default. Uses no changepoints.
The \code{segmenter} is of class \code{\link[=seg_cpt]{seg_cpt()}}`.
}
}
\examples{
mod_null <- segment(DataCPSim)
augment(mod_null)
tidy(mod_null)
glance(mod_null)
segment(DataCPSim, method = "pelt")
segment(DataCPSim, method = "pelt", penalty = "AIC")
segment(DataCPSim, method = "binseg", penalty = "AIC")
segment(DataCPSim, method = "segneigh", penalty = "BIC")
segment(DataCPSim, method = "random")
segment(DataCPSim, method = "manual", tau = c(826))
two_cpts <- segment(DataCPSim, method = "manual", tau = c(365, 826))
plot(two_cpts)
diagnose(two_cpts)
segment(bogota_pm, method = "pelt")
\dontrun{
x <- segment(DataCPSim, method = "gbmdl", num_generations = 10)
}

}
\seealso{
\code{\link[changepoint:cpt.meanvar]{changepoint::cpt.meanvar()}}, \code{\link[wbs:wbs]{wbs::wbs()}}, \code{\link[GA:ga]{GA::ga()}},
\code{\link[=segment_ga]{segment_ga()}}
}
