% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment}
\alias{segment}
\alias{segment.tbl_ts}
\alias{segment.xts}
\alias{segment.numeric}
\alias{segment.ts}
\title{Segment a time series using a variety of algorithms}
\usage{
segment(x, method = "null", ...)

\method{segment}{tbl_ts}(x, method = "null", ...)

\method{segment}{xts}(x, method = "null", ...)

\method{segment}{numeric}(x, method = "null", ...)

\method{segment}{ts}(x, method = "null", ...)
}
\arguments{
\item{x}{a \link[tsibble:tsibble]{tsibble::tsibble} or numeric vector coercible into a \link[stats:ts]{stats::ts} object}

\item{method}{a character string indicating the algorithm to use. See Details.}

\item{...}{arguments passed to methods}
}
\value{
An object of class \code{tidycpt}. Every \code{tidycpt} object contains:
\itemize{
\item \code{segmenter}: The object returned by the underlying function.
\item \code{nhpp}: A \code{tbl_df} showing the best fit parameters for each region, as
defined by the changepoint set returned by \code{\link[=changepoints]{changepoints()}} when called
on the \code{segmenter}. These parameters
are fit using the non-homogeneous Poisson process model in \code{\link[=fit_nhpp]{fit_nhpp()}}.
This object also has class \code{nhpp}.
\item \code{elapsed_time}: The clock time that passed while the algorithm was running.
}
}
\description{
A wrapper function that encapsulates various algorithms for detecting changepoint
sets in univariate time series.
}
\details{
Currently, \code{\link[=segment]{segment()}} can use the following algorithms, depending
on the value of the \code{method} argument:
\itemize{
\item \code{pelt}: Uses the PELT algorithm as implemented in
\code{\link[changepoint:cpt.meanvar]{changepoint::cpt.meanvar()}}. The \code{segmenter} is of class \code{cpt}.
\item \code{binseg}: Uses the Binary Segmentation algorithm as implemented by
\code{\link[changepoint:cpt.meanvar]{changepoint::cpt.meanvar()}}. The \code{segmenter} is of class \code{cpt}.
\item \code{segneigh}: Uses the Segmented Neighborhood algorithm as implemented by
\code{\link[changepoint:cpt.meanvar]{changepoint::cpt.meanvar()}}. The \code{segmenter} is of class \code{cpt}.
\item \code{single-best}: Uses the AMOC criteria as implemented by
\code{\link[changepoint:cpt.meanvar]{changepoint::cpt.meanvar()}}. The \code{segmenter} is of class \code{cpt}.
\item \code{wbs}: Uses the Wild Binary Segmentation algorithm as implemented by
\code{\link[wbs:wbs]{wbs::wbs()}}. The \code{segmenter} is of class \code{wbs}.
\item \code{ga-shi}: Uses the Genetic algorithm used by Shi and implemented by
\code{\link[GA:ga]{GA::ga()}}. The \code{segmenter} is of class \code{tidyga}.
\item \code{gbmdl}: Uses the Genetic BMDL heuristic as implemented by
\code{\link[=segment_gbmdl]{segment_gbmdl()}}. The \code{segmenter} is of class \code{\link[=seg_default]{seg_default()}}.
\item \code{manual}: Uses the vector of changepoints in the \code{cpts} argument.
The \code{segmenter} is of class \code{\link[=seg_default]{seg_default()}}`.
\item \code{null}: The default. Uses no changepoints.
The \code{segmenter} is of class \code{\link[=seg_default]{seg_default()}}`.
}
}
\examples{
mod_null <- segment(DataCPSim)
augment(mod_null)
tidy(mod_null)
glance(mod_null)
segment(DataCPSim, method = "pelt")
segment(DataCPSim, method = "pelt", penalty = "AIC")
segment(DataCPSim, method = "binseg", penalty = "AIC")
segment(DataCPSim, method = "segneigh", penalty = "BIC")
segment(DataCPSim, method = "random", num_generations = 10)
segment(DataCPSim, method = "manual", cpts = c(826))
two_cpts <- segment(DataCPSim, method = "manual", cpts = c(365, 826))
plot(two_cpts)
diagnose(two_cpts)
segment(bogota_pm, method = "pelt")
\dontrun{
x <- segment(DataCPSim, method = "gbmdl", num_generations = 10)
}

}
\seealso{
\code{\link[changepoint:cpt.meanvar]{changepoint::cpt.meanvar()}}, \code{\link[wbs:wbs]{wbs::wbs()}}, \code{\link[GA:ga]{GA::ga()}}
}
