% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycpt.R
\name{segment}
\alias{segment}
\alias{segment.tbl_ts}
\alias{segment.numeric}
\alias{segment.ts}
\title{Segment a time series using a variety of algorithms}
\usage{
segment(x, method = "null", ...)

\method{segment}{tbl_ts}(x, method = "null", ...)

\method{segment}{numeric}(x, method = "null", ...)

\method{segment}{ts}(x, method = "null", ...)
}
\arguments{
\item{x}{a \link[tsibble:tsibble]{tsibble::tsibble} or numeric vector coercible into a \link[stats:ts]{stats::ts} object}

\item{method}{a character string indicating the algorithm to use. See Details.}

\item{...}{arguments passed to methods}
}
\value{
an object of class \code{tidycpt}. Every \code{tidycpt} object contains:
\itemize{
\item \code{segmenter}: The object returned by the underlying function.
\item \code{nhpp}: A \code{tbl_df} showing the best fit parameters for each region, as
defined by the chnagepoint set returned by \code{changepoints()}. These parameters
are fit using the non-homogeneous Poisson process model in \code{\link[=fit_nhpp]{fit_nhpp()}}.
}
}
\description{
A wrapper function that encapsulates various algorithms for detecting changepoint
sets in univariate time series.
}
\details{
Currently, \code{\link[=segment]{segment()}} can use the following algorithms, depending
on the value of the \code{method} argument:
\itemize{
\item \code{cpt-pelt}: Uses the PELT algorithm as implemented in
\code{\link[changepoint:cpt.meanvar]{changepoint::cpt.meanvar()}}. The \code{segmenter} is of class \code{cpt}.
\item \code{cpt-binseg}: Uses the Binary Segmentation algorithm as implemented by
\code{\link[changepoint:cpt.meanvar]{changepoint::cpt.meanvar()}}. The \code{segmenter} is of class \code{cpt}.
\item \code{cpt-segneigh}: Uses the Segmented Neighborhood algorithm as implemented by
\code{\link[changepoint:cpt.meanvar]{changepoint::cpt.meanvar()}}. The \code{segmenter} is of class \code{cpt}.
\item \code{single-best}: Uses the AMOC criteria as implemented by
\code{\link[changepoint:cpt.meanvar]{changepoint::cpt.meanvar()}}. The \code{segmenter} is of class \code{cpt}.
\item \code{cpt-gbmdl}: Uses the Genetic BMDL heuristic as implemented by
\code{\link[=segment_gbmdl]{segment_gbmdl()}}. The \code{segmenter} is of class \link{cpt_gbmdl}.
\item \code{cpt-manual}: Uses the vector of changepoints in the \code{cpts} arugment and
\code{\link[stats:lm]{stats::lm()}}. The \code{segmenter} is of class \code{lm}.
\item \code{null}: The default. Uses \code{\link[stats:lm]{stats::lm()}} with no changepoints.
The \code{segmenter} is of class \code{lm}.
\link[stats:lm]{stats::lm}, or \link{cpt_gbmdl}
}
}
\examples{
mod_null <- segment(DataCPSim)
augment(mod_null)
tidy(mod_null)
glance(mod_null)
segment(DataCPSim, method = "cpt-pelt")
segment(DataCPSim, method = "cpt-pelt", penalty = "AIC")
segment(DataCPSim, method = "cpt-binseg", penalty = "AIC")
segment(DataCPSim, method = "cpt-segneigh", penalty = "BIC")
segment(DataCPSim, method = "cpt-manual", cpts = c(826))
two_cpts <- segment(DataCPSim, method = "cpt-manual", cpts = c(365, 826))
plot(two_cpts)
diagnose(two_cpts)
\dontrun{
x <- segment(DataCPSim, method = "cpt-gbmdl")
}

}
\seealso{
\code{\link[changepoint:cpt.meanvar]{changepoint::cpt.meanvar()}}
}
