% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment}
\alias{segment}
\alias{segment.numeric}
\alias{segment.ts}
\title{Segment a time series using Genetic BMDL heuristic}
\usage{
segment(x, method = "null", ...)

\method{segment}{numeric}(x, method = "null", ...)

\method{segment}{ts}(x, method = "null", ...)
}
\description{
Segment a time series using Genetic BMDL heuristic
}
\examples{
mod_null <- segment(DataCPSim)
augment(mod_null)
}
