% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycpt.R
\name{segment}
\alias{segment}
\alias{segment.numeric}
\alias{segment.ts}
\alias{as.ts.tidycpt}
\alias{length.tidycpt}
\alias{changepoints}
\alias{changepoints.tidycpt}
\alias{augment.tidycpt}
\alias{tidy.tidycpt}
\alias{glance.tidycpt}
\alias{regions}
\alias{regions.tidycpt}
\alias{plot.tidycpt}
\alias{diagnose}
\alias{diagnose.tidycpt}
\title{Segment a time series using a variety of algorithms}
\usage{
segment(x, method = "null", ...)

\method{segment}{numeric}(x, method = "null", ...)

\method{segment}{ts}(x, method = "null", ...)

\method{as.ts}{tidycpt}(x, ...)

\method{length}{tidycpt}(x, ...)

changepoints(x, ...)

\method{changepoints}{tidycpt}(x, ...)

\method{augment}{tidycpt}(x, ...)

\method{tidy}{tidycpt}(x, ...)

\method{glance}{tidycpt}(x, ...)

regions(x, ...)

\method{regions}{tidycpt}(x, ...)

\method{plot}{tidycpt}(x, ...)

diagnose(x, ...)

\method{diagnose}{tidycpt}(x, ...)
}
\arguments{
\item{x}{a numeric vector coercible into a \link[stats:ts]{stats::ts} object}

\item{method}{a character string indicating the algorithm to use}

\item{...}{arguments passed to methods}
}
\value{
an object of class \code{changepoint::cpt}, \link[stats:lm]{stats::lm}, or \link{cpt_gbmdl}
}
\description{
Segment a time series using a variety of algorithms
}
\examples{
mod_null <- segment(DataCPSim)
augment(mod_null)
tidy(mod_null)
glance(mod_null)
segment(DataCPSim, method = "cpt-pelt")
segment(DataCPSim, method = "cpt-pelt", penalty = "AIC")
segment(DataCPSim, method = "cpt-binseg", penalty = "AIC")
segment(DataCPSim, method = "cpt-segneigh", penalty = "BIC")
segment(DataCPSim, method = "cpt-manual", cpts = c(826))
two_cpts <- segment(DataCPSim, method = "cpt-manual", cpts = c(365, 826))
plot(two_cpts)
\dontrun{
segment(DataCPSim, method = "cpt-gbmdl")
}

diagnose(segment(DataCPSim))
diagnose(segment(DataCPSim, method = "single-best"))
diagnose(segment(DataCPSim, method = "cpt-pelt"))
diagnose(segment(test_set()))
diagnose(segment(test_set(n = 2, sd = 4), method = "cpt-pelt"))

}
\seealso{
\code{\link[changepoint:cpt.meanvar]{changepoint::cpt.meanvar()}}
}
