% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_boltzmann.R
\name{segment_boltzmann}
\alias{segment_boltzmann}
\alias{evolve_boltzmann}
\title{Segment a times series using Boltzmann criteria for a genetic algorithm}
\usage{
segment_boltzmann(
  x,
  num_generations = 10,
  generation_size = 50,
  mutation_rate = 0.03,
  temperature = 27,
  show_progress_bar = TRUE
)

evolve_boltzmann(mat_cp, Fitness, mutation_rate, temperature)
}
\arguments{
\item{x}{An \code{integer} vector. See \code{\link[=exceedances]{exceedances()}}.}

\item{num_generations}{number of generations}

\item{generation_size}{size of generation}

\item{mutation_rate}{description}

\item{temperature}{description}

\item{show_progress_bar}{show a progress bar?}

\item{mat_cp}{description}

\item{Fitness}{description}
}
\value{
A \code{list} of length 2:
\itemize{
\item A \code{numeric} vector of \code{FitnessGeneraciones}
\item A \code{matrix} with \code{k} rows and \code{N} columns
}
}
\description{
What does this function do?
}
\examples{
\dontrun{
x <- segment(DataCPSim, method = "boltzmann", num_generations = 5)
y <- segment_boltzmann(rlnorm_ts_1)
  
}
}
