% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_ga.R
\name{segment_ga}
\alias{segment_ga}
\title{Segment a time series using a genetic algorithm}
\usage{
segment_ga(x, initial_prob = 0.01, ...)
}
\description{
Segment a time series using a genetic algorithm
}
\examples{
\dontrun{
res <- segment_ga(DataCPSim, initial_prob = 0.01, maxiter = 50)
summary(res)
str(res)
plot(res)
}
}
