% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_ga.R
\name{segment_ga}
\alias{segment_ga}
\title{Segment a time series using a genetic algorithm}
\usage{
segment_ga(x, ...)
}
\description{
Segment a time series using a genetic algorithm
}
\examples{
res <- segment_ga(DataCPSim)
summary(res)
str(res)
plot(res)
}
