% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_ga.R
\name{segment_ga}
\alias{segment_ga}
\alias{segment_ga_shi}
\alias{segment_ga_taimal}
\title{Segment a time series using a genetic algorithm}
\usage{
segment_ga(
  x,
  model_fn = fit_meanshift_norm,
  penalty_fn = BIC,
  model_fn_args = list(),
  ...
)

segment_ga_shi(x, ...)

segment_ga_taimal(x, ...)
}
\arguments{
\item{x}{A time series}

\item{model_fn}{A \code{character} or \code{name} coercible into \code{fun_cpt} function.
See, for example, \code{\link[=fit_meanshift_norm]{fit_meanshift_norm()}}.}

\item{penalty_fn}{A function that evaluates the changepoint set returned by
\code{model_fn}. We provide \code{\link[=AIC]{AIC()}}, \code{\link[=BIC]{BIC()}}, \code{\link[=MBIC]{MBIC()}}, \code{\link[=MDL]{MDL()}}, and \code{\link[=BMDL]{BMDL()}}.}

\item{model_fn_args}{A \code{\link[=list]{list()}} of parameters passed to \code{model_fn}}

\item{...}{arguments passed to \code{\link[GA:ga]{GA::ga()}}}
}
\description{
Segment a time series using a genetic algorithm
}
\details{
Shi's algorithm is the algorithm used in \doi{10.1175/JCLI-D-21-0489.1}.
Note that in order to achieve the reported results you have to run the algorithm
for a really long time. Pass the values \code{maxiter} = 50000 and \code{run} = 10000
to \code{\link[GA:ga]{GA::ga()}} using the dots.

Taimal's algorithm is the one used in \doi{10.1007/978-3-031-47372-2_20}.

Note that the speed of the algorithm is highly sensitive to the size of the
changepoint sets under consideration, with large changepoint sets begin slow.
Consider setting the \code{population} argument to \code{\link[GA:ga]{GA::ga()}} to improve
performance. Taimal's algorithm uses the \code{\link[=build_gabin_population]{build_gabin_population()}} function
for this purpose by default.
}
\examples{
\dontrun{
res <- segment_ga(DataCPSim, maxiter = 10)
summary(res)
str(res)
plot(res)
# Shi's algorithm
x <- segment(CET, method = "ga-shi", maxiter = 50)
str(x)
# GeneticBMDL
y <- segment(CET, method = "ga-gbmdl", maxiter = 20)
changepoints(y)

z <- segment(CET, method = "ga-gbmdl", maxiter = 10, model_fn_args = list(threshold = 2))
changepoints(z)
}
\dontrun{
# This will take a really long time!
x <- segment(CET, method = "ga-shi", maxiter = 500, run = 100)
changepoints(x)

y <- segment(CET, method = "ga", model_fn = fit_lmshift, penalty_fn = BIC, 
  popSize = 200, maxiter = 5000, run = 1000, model_fn_args = list(trends = TRUE), 
  population = build_gabin_population(CET))

}

\dontrun{
x <- segment(method = "ga-taimal", maxiter = 50)
}

}
\references{
Shi, et al. (2022, \doi{10.1175/JCLI-D-21-0489.1})

Taimal, et al. (2023, \doi{10.1007/978-3-031-47372-2_20})
}
\seealso{
\code{\link[=build_gabin_population]{build_gabin_population()}}
}
