% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_ga.R
\name{segment_ga}
\alias{segment_ga}
\title{Segment a time series using a genetic algorithm}
\usage{
segment_ga(
  x,
  model_fn = fit_meanshift_ar1,
  penalty_fn = BIC,
  model_params = list(initial_prob = 0.01),
  ...
)
}
\arguments{
\item{x}{A time series}

\item{model_fn}{A function that takes two arguments: \code{x} (a time series) and
\code{tau} (a set of changepoint indices). See \code{\link[=fit_meanshift_ar1]{fit_meanshift_ar1()}},
\code{\link[=fit_lmshift]{fit_lmshift()}}, and \code{\link[=fit_nhpp]{fit_nhpp()}}}

\item{penalty_fn}{A function that evaluates the changepoint set returned by
\code{model_fn}. We provide \code{\link[=AIC]{AIC()}}, \code{\link[=BIC]{BIC()}}, \code{\link[=MBIC]{MBIC()}}, \code{\link[=MDL]{MDL()}}, and \code{\link[=BMDL]{BMDL()}}.}

\item{model_params}{A \code{\link[=list]{list()}} of parameters passed to \code{model_fn}}

\item{...}{arguments passed to \code{\link[GA:ga]{GA::ga()}}}
}
\description{
Segment a time series using a genetic algorithm
}
\examples{
\dontrun{
res <- segment_ga(DataCPSim, model_params = list(initial_prob = 0.06), maxiter = 50)
summary(res)
str(res)
plot(res)
# Shi's algorithm
x <- segment(CET, method = "ga", penalty_fn = BIC, 
model_params = list(initial_prob = 0.06), maxiter = 50, popSize = 200)
str(x)
# GeneticBMDL
y <- segment(CET, method = "ga", model_fn = fit_nhpp, penalty_fn = BMDL, 
  model_params = list(initial_prob = 0.06), maxiter = 20, popSize = 50)
changepoints(y)

z <- segment(CET, method = "ga", model_fn = fit_nhpp, penalty_fn = BMDL, 
  maxiter = 10, popSize = 20, 
  model_params = list(initial_prob = 0.06, threshold = 2))
changepoints(z)
}
}
