% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_gbmdl.R
\name{segment_gbmdl}
\alias{segment_gbmdl}
\title{Algoritmo genético de Bayesian MDL a un paso}
\usage{
segment_gbmdl(
  x,
  param,
  destdir = tempdir(),
  show_progress_bar = TRUE,
  write_rda = FALSE
)
}
\arguments{
\item{x}{a vector that can be coerced into a \code{ts}}

\item{param}{description. See \link{param}.}

\item{show_progress_bar}{show the progress bar?}
}
\value{
A \code{cpt_gbmdl} object
}
\description{
Algoritmo genético de Bayesian MDL a un paso
}
\examples{
\dontrun{
x <- segment_gbmdl(DataCPSim, param)
y <- segment_gbmdl(rlnorm_ts_1, param)
}

}
