% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_gbmdl.R
\name{segment_gbmdl}
\alias{segment_gbmdl}
\title{Algoritmo genético de Bayesian MDL a un paso}
\usage{
segment_gbmdl(x, param, destdir = tempdir())
}
\arguments{
\item{x}{a vector that can be coerced into a \code{ts}}

\item{param}{description. See \link{param}.}
}
\value{
A \code{cpt_gbmdl} object
}
\description{
Algoritmo genético de Bayesian MDL a un paso
}
\examples{
\dontrun{
lista_AG <- cpt_bmdl_genetic(DataCPSim, param)
lista_AG <- cpt_bmdl_genetic(rlnorm_ts_1, param)
}

}
