% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_gbmdl.R
\name{segment_gbmdl}
\alias{segment_gbmdl}
\title{Algoritmo genético de Bayesian MDL a un paso}
\usage{
segment_gbmdl(x, num_generations = 50, show_progress_bar = TRUE)
}
\arguments{
\item{x}{a vector that can be coerced into a \code{ts}}

\item{num_generations}{Number of generations to evolve}

\item{show_progress_bar}{show the progress bar?}
}
\value{
A \code{cpt_gbmdl} object
}
\description{
Algoritmo genético de Bayesian MDL a un paso
}
\examples{
\dontrun{
x <- segment_gbmdl(DataCPSim, num_generations = 10)
y <- segment_gbmdl(rlnorm_ts_1)
z <- segment_gbmdl(bogota_pm$particulate_matter)
}

}
