% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_gbmdl_algorithm.R
\name{segment_gbmdl}
\alias{segment_gbmdl}
\title{Algoritmo genético de Bayesian MDL a un paso}
\usage{
segment_gbmdl(
  x,
  num_generations = 50,
  destdir = tempdir(),
  show_progress_bar = TRUE,
  write_rda = FALSE
)
}
\arguments{
\item{x}{a vector that can be coerced into a \code{ts}}

\item{destdir}{Directory to stare files}

\item{show_progress_bar}{show the progress bar?}

\item{write_rda}{Write the data to \code{destdir}?}
}
\value{
A \code{cpt_gbmdl} object
}
\description{
Algoritmo genético de Bayesian MDL a un paso
}
\examples{
\dontrun{
x <- segment_gbmdl(DataCPSim, num_generations = 10)
y <- segment_gbmdl(rlnorm_ts_1)
z <- segment_gbmdl(bogota_pm$particulate_matter)
}

}
