% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_ga.R, R/segment_manual.R
\name{segment_ga_random}
\alias{segment_ga_random}
\alias{segment_manual}
\title{Manually segment a times series}
\usage{
segment_ga_random(x, ...)

segment_manual(x, tau, ...)
}
\arguments{
\item{x}{A time series}

\item{...}{arguments passed to \code{\link[stats:lm]{stats::lm()}}}

\item{tau}{a set of indices representing a changepoint set}
}
\description{
Manually segment a times series
}
\details{
Randomly select candidate changepoint sets
}
\examples{
x <- segment(CET, method = "random")
segment_manual(CET, tau = c(84, 330))
segment_manual(CET, tau = NULL)

}
\seealso{
\code{\link[=build_gabin_population]{build_gabin_population()}}
}
