% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_pelt.R
\name{segment_pelt}
\alias{segment_pelt}
\title{Segment a time series using the PELT algorithm}
\usage{
segment_pelt(x, model_fn = fit_meanvar, ...)
}
\arguments{
\item{x}{A time series}

\item{model_fn}{A \code{character} or \code{name} coercible into a \code{\link[=fun_cpt]{fun_cpt()}} function.
See, for example, \code{\link[=fit_meanshift_norm]{fit_meanshift_norm()}}. The default is \code{\link[=fit_meanvar]{fit_meanvar()}}.}

\item{...}{arguments passed to \code{\link[changepoint:cpt.meanvar]{changepoint::cpt.meanvar()}}}
}
\description{
Segmenting functions for the PELT algorithm
}
\examples{
res <- segment_pelt(DataCPSim)
summary(res)
str(res)
plot(res)
segment_pelt(as.ts(CET))
segment_pelt(as.ts(CET), penalty = "BIC")
segment_pelt(as.ts(CET), model_fn = fit_meanshift_norm, penalty = "BIC")
}
