% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_generics.R
\name{SIC}
\alias{SIC}
\title{Schwarz information criterion}
\usage{
SIC(object, ...)
}
\description{
Schwarz information criterion
}
\examples{
# The SIC is just the BIC
SIC(fit_meanvar(CET, tau = NULL))

}
\seealso{
\code{\link[stats:AIC]{stats::BIC()}}
}
