% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_generics.R
\name{SIC}
\alias{SIC}
\title{Schwarz information criterion}
\usage{
SIC(object, ...)
}
\arguments{
\item{object}{a fitted model object for which there exists a
    \code{logLik} method to extract the corresponding log-likelihood, or
    an object inheriting from class \code{logLik}.}

\item{...}{optionally more fitted model objects.}
}
\value{
The value of \code{\link[stats:AIC]{stats::BIC()}}
}
\description{
Schwarz information criterion
}
\examples{
# The SIC is just the BIC
SIC(fit_meanvar(CET, tau = NULL))
BIC(fit_meanvar(CET, tau = NULL))

}
\seealso{
\code{\link[stats:AIC]{stats::BIC()}}, \code{\link[stats:AIC]{stats::AIC()}}

Other penalty-functions: 
\code{\link{BMDL}()},
\code{\link{HQC}()},
\code{\link{MBIC}()},
\code{\link{MDL}()}
}
\concept{penalty-functions}
