% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_bmdl.R
\name{sim_1_cp_BMDL}
\alias{sim_1_cp_BMDL}
\alias{sim_k_cp_BMDL}
\title{Genera un cromosoma de puntos de cambio para el Bayesian MDL}
\usage{
sim_1_cp_BMDL(x, param, max_num_cp = 20, prob_inicial = 0.06)

sim_k_cp_BMDL(x, generation_size = 50, max_num_cp = 20)
}
\arguments{
\item{x}{vector de excedentes}

\item{param}{lista de parámetros globales. See \link{param}.}

\item{prob_inicial}{probabilidad de que en la primera generación un punto
cualquiera sea punto de cambio. Se recomienda =.5 ya que con esto se
distribuyen relativamente uniformes los puntos de cambio}
}
\value{
regresa una matriz de \code{k} por \code{max_num_cp+3}, la cual en cada renglón tiene
una simulación de un vector de tiempos de cambio
}
\description{
Genera un cromosoma de puntos de cambio para el Bayesian MDL

Simula k vectores change point para Bayesian MDL
}
\details{
regresa un vector de tamaño \code{max_num_cp+3} donde la primera entrada es
m, la segunda \eqn{v_0=1, ...., v_{m+1}=N,0,...,0}

por ejemplo: \code{c(4,1,3,8,11,15,20,0,0,0,0)} para \code{m=4}, \eqn{max\_num\_cp=8}, \eqn{N=20}.
Se tienen \code{m} puntos de cambio, los cuales
\eqn{\tau_0=1} y \eqn{\tau_{m+1}= N+1},
pero en nuestro caso
tenemos que los vectores \code{cp} tienen \eqn{c(m,\tau_0=1,\tau_1,...,\tau_{m-1},\tau_m= N,0,0,0)}
por lo cual se nosotros:
\itemize{
\item empieza con el número de puntos de cambio;
\item la segunda entrada es un uno;
\item la tercera entrada es el primer punto de cambio;
\item las siguientes son otros puntos de cambio;
\item la siguiente entrada después de punto de cambio tiene el valor  \code{N}; y
\item los siguientes son númores cero hasta llenarlo para que sea de tamaño \code{max_num_cp}
}
}
\examples{
sim_1_cp_BMDL(exceedances(DataCPSim))
sim_1_cp_BMDL(exceedances(rlnorm_ts_1))
sim_1_cp_BMDL(exceedances(rlnorm_ts_2))
sim_1_cp_BMDL(exceedances(rlnorm_ts_3))

sim_k_cp_BMDL(DataCPSim)

}
