% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sims}
\alias{sims}
\title{Simulated time series data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1096 rows and 3 columns.
}
\usage{
sims
}
\description{
Each column contains a \code{\link[stats:ts]{stats::ts()}} vector.
\itemize{
\item For \code{ts1}, there is one changepoint located at 826.
\item For \code{ts2}, there are two changepoints, located at 366 and 731.
\item For \code{ts3}, there are three changepoints, located at 548, 823, and 973.
}
}
\examples{
plot(sims$ts1)

}
\seealso{
\code{\link[stats:ts]{stats::ts()}}
}
\keyword{datasets}
