% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lmshift.R
\name{tbl_coef}
\alias{tbl_coef}
\title{Format the coefficients from a linear model as a tibble}
\usage{
tbl_coef(mod, ...)
}
\arguments{
\item{mod}{An \code{lm} model object}

\item{...}{currently ignored}
}
\description{
Format the coefficients from a linear model as a tibble
}
\examples{
ds <- data.frame(y = as.ts(CET), t = 1:length(CET))
tbl_coef(lm(y ~ 1, data = ds))
tbl_coef(lm(y ~ (t >= 42) + (t >= 81), data = ds))
tbl_coef(lm(y ~ t * (t >= 42) + t * (t >= 81), data = ds))
}
