% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_set}
\alias{test_set}
\title{Simulate time series with known changepoint sets}
\usage{
test_set(n = 1, sd = 1, seed = NULL)
}
\arguments{
\item{n}{Number of true changepoints in set}

\item{sd}{Standard deviation passed to \code{\link[stats:Normal]{stats::rnorm()}}}

\item{seed}{Value passed to \code{\link[base:Random]{base::set.seed()}}}
}
\value{
A \code{\link[stats:ts]{stats::ts()}} object
}
\description{
Simulate time series with known changepoint sets
}
\examples{
x <- test_set()
plot(x)
changepoints(x)
}
\seealso{
\link{DataCPSim}
}
