% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidychangepoint-package.R
\docType{package}
\name{tidychangepoint-package}
\alias{tidychangepoint}
\alias{tidychangepoint-package}
\title{tidychangepoint: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://beanumber.github.io/tidychangepoint/}
}

}
\author{
\strong{Maintainer}: Benjamin S. Baumer \email{ben.baumer@gmail.com} (\href{https://orcid.org/0000-0002-3279-0516}{ORCID})

Authors:
\itemize{
  \item Biviana Suarez Sierra \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})
  \item Carlos A. Taimal \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})
}

Other contributors:
\itemize{
  \item Xueheng Shi [contributor]
}

}
\keyword{internal}
