% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidychangepoint-package.R
\docType{package}
\name{tidychangepoint-package}
\alias{tidychangepoint}
\alias{tidychangepoint-package}
\title{tidychangepoint: A Tidy Framework for Changepoint Detection Analysis}
\description{
Changepoint detection algorithms for R are widespread but have different interfaces and reporting conventions. This makes the comparative analysis of results difficult. We solve this problem by providing a tidy, unified interface for several different changepoint detection algorithms. We also provide consistent numerical and graphical reporting leveraging the 'broom' and 'ggplot2' packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://beanumber.github.io/tidychangepoint/}
}

}
\author{
\strong{Maintainer}: Benjamin S. Baumer \email{ben.baumer@gmail.com} (\href{https://orcid.org/0000-0002-3279-0516}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Biviana Marcela Suárez Sierra \email{bmsuarezs@eafit.edu.co} (\href{https://orcid.org/0000-0003-2151-3537}{ORCID})
  \item Arrigo Coen (\href{https://orcid.org/0000-0001-7798-7104}{ORCID})
  \item Carlos A. Taimal (\href{https://orcid.org/0000-0002-8716-1282}{ORCID})
}

Other contributors:
\itemize{
  \item Xueheng Shi [contributor]
}

}
\keyword{internal}
