% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycpt.R
\name{tidycpt-generics}
\alias{tidycpt-generics}
\alias{as.ts.tidycpt}
\alias{length.tidycpt}
\alias{logLik.tidycpt}
\alias{nobs.tidycpt}
\alias{augment.tidycpt}
\alias{tidy.tidycpt}
\alias{glance.tidycpt}
\alias{plot.tidycpt}
\title{Generic functions for tidycpt objects}
\usage{
\method{as.ts}{tidycpt}(x, ...)

\method{length}{tidycpt}(x, ...)

\method{logLik}{tidycpt}(object, ...)

\method{nobs}{tidycpt}(object, ...)

\method{augment}{tidycpt}(x, ...)

\method{tidy}{tidycpt}(x, ...)

\method{glance}{tidycpt}(x, ...)

\method{plot}{tidycpt}(x, ...)
}
\arguments{
\item{x}{An \code{tidycpt} object, typically the output from \code{\link[=segment]{segment()}}`}

\item{...}{arguments passed to methods}

\item{object}{A \code{tidycpt} object}
}
\description{
Generic functions for tidycpt objects
}
\seealso{
\code{\link[stats:logLik]{stats::logLik()}}

\code{\link[stats:nobs]{stats::nobs()}}

\code{\link[broom:reexports]{broom::augment()}}

\code{\link[broom:reexports]{broom::tidy()}}

\code{\link[broom:reexports]{broom::glance()}}
}
