% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbs.R
\name{wbs-generics}
\alias{wbs-generics}
\alias{glance.wbs}
\alias{params.wbs}
\alias{as.ts.wbs}
\alias{nobs.wbs}
\title{Compatibility layer for wbs}
\usage{
\method{glance}{wbs}(x, ...)

\method{params}{wbs}(x, ...)

\method{as.ts}{wbs}(x, ...)

\method{nobs}{wbs}(object, ...)
}
\arguments{
\item{x}{A \code{wbs} object returned by \code{\link[wbs:wbs]{wbs::wbs()}}}

\item{...}{arguments passed to methods}

\item{object}{A \code{cpt} object.}
}
\description{
Compatibility layer for wbs
}
\examples{
cpts <- segment(DataCPSim, method = "wbs")
class(cpts)
y <- augment(cpts)
class(y)
y
tidy(cpts)
glance(cpts)
cpts <- segment(DataCPSim, method = "wbs")
as.ts(cpts)

cpts <- segment(DataCPSim, method = "wbs")
nobs(cpts)

}
