% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-wbs.R
\name{wbs-generics}
\alias{wbs-generics}
\alias{as.segmenter.wbs}
\alias{as.ts.wbs}
\alias{nobs.wbs}
\alias{seg_params.wbs}
\title{Compatibility layer for wbs}
\usage{
\method{as.segmenter}{wbs}(object, ...)

\method{as.ts}{wbs}(x, ...)

\method{nobs}{wbs}(object, ...)

\method{seg_params}{wbs}(x, ...)
}
\arguments{
\item{object}{A \code{cpt} object.}

\item{...}{arguments passed to methods}

\item{x}{A \code{wbs} object returned by \code{\link[wbs:wbs]{wbs::wbs()}}}
}
\description{
Compatibility layer for wbs
}
\examples{
cpts <- segment(DataCPSim, method = "wbs")
x <- cpts$segmenter
class(x)
as.segmenter(x)
as.ts(x) 
changepoints(x)
fitness(x)
model_name(x)
model_args(x)
nobs(x)
seg_params(x)

cpts <- segment(DataCPSim, method = "wbs")
as.ts(cpts)

cpts <- segment(DataCPSim, method = "wbs")
nobs(cpts$segmenter)
}
