% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-wbs.R
\name{wbs-generics}
\alias{wbs-generics}
\alias{as.seg_cpt.wbs}
\title{Compatibility layer for wbs}
\usage{
\method{as.seg_cpt}{wbs}(object, ...)
}
\arguments{
\item{...}{arguments passed to methods}

\item{x}{A \code{wbs} object returned by \code{\link[wbs:wbs]{wbs::wbs()}}}
}
\description{
Methods for generic functions applied to \code{wbs} objects
}
\examples{
cpts <- segment(DataCPSim, method = "wbs")
x <- as.segmenter(cpts)
class(x)
as.seg_cpt(x)
as.ts(x) 
changepoints(x)
fitness(x)
model_name(x)
model_args(x)
nobs(x)
seg_params(x)

}
\seealso{
\link[wbs:wbs-package]{wbs::wbs-package}
}
